% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset_featureTables.R
\name{activeFT<-}
\alias{activeFT<-}
\title{activeFT}
\usage{
activeFT(x) <- value
}
\arguments{
\item{x}{an MseekTree (reactivevalues) object}

\item{value}{a character() value corresponding to a table ID}
}
\description{
changes the currently active Feature Table if supplied value is a valid table ID
}
