% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_plotting_EICs.R
\name{addLines}
\alias{addLines}
\title{addLines}
\usage{
addLines(EIClist, TIC = F, colr = topo.colors(1, alpha = 1),
  relto = NULL, liwi = 2)
}
\arguments{
\item{EIClist}{matrix or list of EIC objects (potentially subsetted)}

\item{TIC}{logical() if TRUE, TIC will be plotted instead of EIC}

\item{colr}{character() of color values for lines}

\item{relto}{if not NULL, intensities will be given as relative to this numeric(1)}

\item{liwi}{line width of plotted lines}
}
\value{
plots EIC trace lines into the current plot window
}
\description{
Helper function for EICplot to plot the lines of EICs into a PlottingWindow. 
Mass shifts will be handled by plotting different line types
}
