% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_FeatureTable_analysis.R
\name{analyzeTable}
\alias{analyzeTable}
\title{analyzeTable}
\usage{
analyzeTable(
  df,
  intensities,
  groups,
  analyze = c("Basic analysis", "clara_cluster", "t-test", "Peak shapes",
    "Fast peak shapes", "PCA features", "PCA samples", "mzMatch"),
  normalize = T,
  useNormalized = T,
  logNormalized = F,
  MSData = NULL,
  ppm = 5,
  controlGroup = NULL,
  numClusters = 2,
  mzMatchParam = list("smid-db_pos.csv", ppm = 5, mzdiff = 0.001),
  workers = 1
)
}
\arguments{
\item{df}{a data.frame with numeric (intensity) values, rt and mz values}

\item{intensities}{the intensity column names, before normalization 
(without __norm suffix), will be automatically renamed if useNormalized.}

\item{groups}{named list of non-normalized intensity columns listed by group 
(as supplied by $anagroupnames of MseekFT objects), will be automatically 
renamed if useNormalized.}

\item{analyze}{character vector to select the analyses to be run: 
"Basic analysis", "clara_cluster", "t-test", "Peak shapes"}

\item{normalize}{normalze intensity columns}

\item{useNormalized}{use normalized values for analyses; will trigger 
normalize if there is no normalized data available for all selected 
intensity columns}

\item{logNormalized}{if TRUE, applies a log10 to intensity values after normalization}

\item{MSData}{a (named) list of xcmsRaw objects, e.g. generated by 
Mseek::loadRawM, OR an MSnbase OnDiskMSnExp object required for peak 
shape analysis}

\item{ppm}{ppm range for peak shape analysis}

\item{controlGroup}{control group for foldChange (part of Basic analysis) 
analysis (optional)}

\item{numClusters}{number of clusters for clara_clusters analysis}

\item{mzMatchParam}{list of parameters passed to mass}

\item{workers}{number of workers to use for multithreaded analyses}
}
\value{
\code{df} with additional columns (or overridden columns) as generated by the selected analyses
}
\description{
Run a series of analyses on a feature table data.frame
}
