% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_class_featureTable.R
\name{asInclusionlist}
\alias{asInclusionlist}
\title{asInclusionlist}
\usage{
asInclusionlist(parenttable, rtwin = 2.5, polarity = "Positive",
  instrument = c("QExactive", "QExactive-HF"), listType = "inclusion",
  restrictRT = F)
}
\arguments{
\item{parenttable}{feature table as data.frame, with mz and rt columns.}

\item{rtwin}{retention time window (+/- df$rt) in seconds}

\item{polarity}{"Positive" or "Negative"}

\item{instrument}{which instrument is this table for?}

\item{listType}{"inclusion" or "exclusion"}

\item{restrictRT}{use retention time column from parenttable and rtwin to restrict retention time in inclusion/exclusion list}
}
\description{
Reformats a feature table into a format that can be imported into MS instruments as inclusion or exclusion lists.
}
