% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_plotting_genericData.R
\name{assignColor}
\alias{assignColor}
\title{assignColor}
\usage{
assignColor(datarange, colscale, scalerange = range(datarange),
  center = NULL, symmetric = F)
}
\arguments{
\item{datarange}{the data range for the legend}

\item{colscale}{character vector of colors}

\item{scalerange}{scale that will be used to spread the colors across}

\item{center}{force the middle of the color vector to correspond to this value}

\item{symmetric}{if true, will spread colors on both sides of center along 
the same scale}
}
\value{
a character vector of same length as \code{datarange} with color values
}
\description{
Assign a color from a range of colors to all values in a numeric vector (datarange)
}
