% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_EICs.R
\name{bestgauss}
\alias{bestgauss}
\title{bestgauss}
\usage{
bestgauss(...)
}
\arguments{
\item{...}{arguments passed on to multiEIC().}
}
\value{
best gauss curve fit score for any of the files in rawdata for
 each feature returned by multiEIC.
}
\description{
wrapper function for using \code{multiEIC} with \code{getgauss = TRUE}
}
