% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_calcMZ.R
\name{calcMZs}
\alias{calcMZs}
\title{calcMZs}
\usage{
calcMZs(df, mf_column = "formula", ...)
}
\arguments{
\item{df}{data.frame that contains a column with molecular formulas (MFs)}

\item{mf_column}{name of column in df that contains MFs}

\item{...}{additional arguments passed to \code{\link[MassTools]{calcIons}()}}
}
\value{
input data.frame with additional columns \code{ion} and \code{charge},
 and updated column \code{mz}
}
\description{
Calculate charge-to-mass ratios for molecula formulas in a data.frame, 
retaining the dataframes columns, but overriding columns mz, charge and ion
}
