% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_xcmsRunner.R
\name{cameraWrapper}
\alias{cameraWrapper}
\title{cameraWrapper}
\usage{
cameraWrapper(xset, polarity = NULL, sigma = 6, perfwhm = 0.6,
  cor_eic_th = 0.75, pval = 0.05, maxcharge = 3, maxiso = 4,
  ppm = 5, mzabs = 0.01, minfrac = 0.5, filter = TRUE,
  workers = 1)
}
\arguments{
\item{polarity}{passed on to \code{\link[CAMERA]{xsAnnotate}()} and 
\code{\link[CAMERA]{findAdducts}()}}

\item{sigma}{passed on to \code{\link[CAMERA]{groupFWHM}()}}

\item{perfwhm}{passed on to \code{\link[CAMERA]{groupFWHM}()}}

\item{cor_eic_th}{passed on to \code{\link[CAMERA]{groupCorr}()}}

\item{pval}{passed on to \code{\link[CAMERA]{groupCorr}()}}

\item{maxcharge}{passed on to \code{\link[CAMERA]{findIsotopes}()}}

\item{maxiso}{passed on to \code{\link[CAMERA]{findIsotopes}()}}

\item{ppm}{passed on to \code{\link[CAMERA]{findIsotopes}()} and 
\code{\link[CAMERA]{findAdducts}()}}

\item{mzabs}{passed on to \code{\link[CAMERA]{findIsotopes}()} and 
\code{\link[CAMERA]{findAdducts}()}}

\item{minfrac}{passed on to \code{\link[CAMERA]{findIsotopes}()}}

\item{filter}{passed on to \code{\link[CAMERA]{findIsotopes}()}}

\item{workers}{number of workers, passed on to \code{\link[CAMERA]{xsAnnotate}()}
(as \code{nSlaves})}
}
\value{
an \code{xsAnnotate} object
}
\description{
converts a \code{XCMSnExp} object to \code{xcmsSet}, then consecutively calls
\code{xsAnnotate}, \code{groupFWHM}, \code{groupCorr}, \code{findIsotopes} and 
\code{findAdducts} from the \code{CAMERA} package, and finally adds an entry to
its \code{.processHistory} slot.
}
