% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_ChangeFTGroupingModule.R
\name{ChangeFTGroupingModule}
\alias{ChangeFTGroupingModule}
\title{ChangeFTGroupingModule}
\usage{
ChangeFTGroupingModule(input, output, session, reactives = reactive({    
  list() }), values = reactiveValues(featureTables = featureTables,
  MSData = MSData, projectData = projectData), static = list())
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{reactives}{Import data from the shiny session}

\item{values}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
server module to select and group intensity columns in Feature tables.
}
