% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_ChangeFTGroupingModule.R
\name{ChangeFTGroupingModuleUI}
\alias{ChangeFTGroupingModuleUI}
\title{ChangeFTGroupingModuleUI}
\usage{
ChangeFTGroupingModuleUI(id)
}
\arguments{
\item{id}{id of the shiny module}
}
\description{
ChangeFTGroupingModuleUI
}
