% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_formatting_utilities.R
\name{checkFolders}
\alias{checkFolders}
\title{checkFolders}
\usage{
checkFolders(query = paste0(LETTERS, ":/"))
}
\arguments{
\item{query}{character vector with folders to search for,
by default looks for drives in Windows file system}
}
\value{
a named character vector of the folders in \code{query} which exist.
}
\description{
Looks for folders as specified in \code{query}
}
