% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_plotting_genericData.R
\name{colorRampLegend}
\alias{colorRampLegend}
\title{colorRampLegend}
\usage{
colorRampLegend(datarange, colscale, title = "")
}
\arguments{
\item{datarange}{the data range for the legend}

\item{colscale}{character vector of colors}

\item{title}{legend title}
}
\description{
Make a figure legend for a continuous color range
}
