% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_columnSelModule.R
\name{ColumnSelModule}
\alias{ColumnSelModule}
\title{columnSelModule}
\usage{
ColumnSelModule(input, output, session,
  values = reactiveValues(featureTables = featureTables, MainTable =
  MainTable))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{import reactiveValues from the shiny session}

\item{load}{data to load from previous session (not implemented)}
}
\description{
server module for selection of columns to show in main Table
}
