% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_columnSelModule.R
\name{columnSelModuleUI}
\alias{columnSelModuleUI}
\title{columnSelModuleUI}
\usage{
columnSelModuleUI(id)
}
\arguments{
\item{id}{id to be used in ns()}
}
\description{
UI module for columnsel Module
}
