% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_class_MSdata.R
\name{constructRawLayout}
\alias{constructRawLayout}
\title{constructRawLayout}
\usage{
constructRawLayout(rawgrouptable, stem = NULL)
}
\arguments{
\item{rawgrouptable}{a data.frame with columns File and Group , holding file paths and group names, respectively.}

\item{stem}{if the file paths in rawgrouptable are not full (e.g. subdirectories of the working directory), this should be the path of the working directory.}
}
\description{
Constructor function for the rawLayout S3 class, holding information on MSdata grouping and layout options in METABOseek.
}
