% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_components.R
\name{dashboardHeaderM}
\alias{dashboardHeaderM}
\title{dashboardHeaderM}
\usage{
dashboardHeaderM(..., title = NULL, titleWidth = NULL,
  disable = FALSE, .list = NULL, left = NULL)
}
\description{
METABOseek-specific, modified version of \code{\link[shinydashboard]{dashboardHeader}()} 
that allows placing icons left and right in the navbar
}
