% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMosaic.R
\docType{data}
\name{default__root}
\alias{default__root}
\title{default__root}
\format{An object of class \code{character} of length 1.}
\usage{
default__root
}
\description{
default__root
}
\keyword{datasets}
