% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_stats.R
\name{densplot}
\alias{densplot}
\title{densplot}
\usage{
densplot(densin = log10(as.numeric(unlist(filtrate3[, scol]))),
  perc = c(0.01, 0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 0.95,
  0.99), ...)
}
\arguments{
\item{densin}{numeric vector or matrix}

\item{perc}{numeric(): draw lines at these quantiles}
}
\description{
generate a density plot for numeric data.
}
\details{
To accomodate logarithmized data,
Inf values will be set to 1.1*largest value that is not Inf
-Inf values will be set to 0.9*smallest value that is not -Inf
}
