% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics_modules.R
\name{densplotModuleUI}
\alias{densplotModuleUI}
\title{#' NormalizeModuleUI
#' 
#' Normalize a matrix (UI single button)
#' 
#' @param id 
#' 
#' @export
NormalizeModuleUI <- function(id){
    ns <- NS(id)
    
   actionButton(ns('normbutton'),"Normalize data") 
    
}}
\usage{
densplotModuleUI(id)
}
\arguments{
\item{id}{}
}
