% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_densPlotWidget.R
\name{densplotWidget}
\alias{densplotWidget}
\alias{densplotModuleUI}
\title{densplotWidget}
\usage{
densplotWidget(input, output, session, reactives = reactive({     list()
  }))

densplotModuleUI(id)
}
\arguments{
\item{reactives}{list of arguments passed to \code{\link{densplot}}}
}
\description{
Plot density of a vector or matrix
}
\section{Functions}{
\itemize{
\item \code{densplotWidget}: server logic for densplotWidget

\item \code{densplotModuleUI}: UI elements for densplotWidget
}}

