% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_FeatureTable_analysis.R
\name{.calculateM}
\alias{.calculateM}
\title{.calculateM}
\usage{
.calculateM(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{a matrix which is expected to contain imputed, log2-transformed intensity values only}

\item{na.rm}{remove NA values for sd calculation?}

\item{...}{Arguments passed to \code{bplapply()}}
}
\description{
Calculates M value as detailed by Vandesompele et al. (2002)
}
\references{

}
\author{
Aiden Kolodziej, Maximilian Helf
}
