% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_class_featureTable.R
\name{.getFTFormat}
\alias{.getFTFormat}
\title{.getFTFormat}
\usage{
.getFTFormat(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
Returns the format of the df object. Can be either of these:
\itemize{
 \item "xcms PeakTable with Metaboseek Intensities"
 \item "xcms Peak Table"
 \item "mzMINE Aligned Peak List"
 \item "unknown"
}
}
\description{
Tries to identify the format of a feature table input data frame
}
