% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{.withHistory}
\alias{.withHistory}
\title{.withHistory}
\usage{
.withHistory(
  fun = character(),
  args = list(),
  longArgs = list(),
  addHistory = TRUE,
  continueWithErrors = FALSE,
  returnIfError = NULL
)
}
\arguments{
\item{fun}{character(1), name of the function to apply}

\item{args}{named list of arguments to fun, will be kept in the ProcessHistory as they are}

\item{longArgs}{arguments to fun that are large. They will be supplied to fun as they are, but 
will be summarized using \code{summary()} in the ProcessHistory.}

\item{addHistory}{if TRUE, adds the processing history to the output object if an 
\code{addProcessHistory} method exists for it. If FALSE, this function
 returns a list(), see \code{Value}!}

\item{continueWithErrors}{if TRUE, if an error occurs during the call to fun,
it is recorded in the ProcessHistory, and the value specified in returnIfError
is returned.}

\item{returnIfError}{object to return if error occurs. Will be treated like the 
object returned by fun (ProcessHistory will be added to the object if possible to keep track of failed analyses).}
}
\value{
if \code{addHistory = TRUE}: the object returned by fun, with analyses performed as
defined by args and longArgs and ProcessHistory added to the object.
if \code{addHistory = FALSE}: a list with elements \code{history} and \code{result}
}
\description{
apply a function to an object and add this event to the object's
processHistory, or return a list that contains the event history
}
