% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_plotting_modules.R
\name{EICmodule}
\alias{EICmodule}
\title{EICmodule}
\usage{
EICmodule(input, output, session, tag, set, keys)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{tag}{id to be used in ns()}

\item{set}{Import data from the shiny session}
}
\description{
server module for interactive EIC view
}
