% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_EICOptionsModule.R
\name{EICOptionsModule}
\alias{EICOptionsModule}
\title{EICOptionsModule}
\usage{
EICOptionsModule(input, output, session, values = reactiveValues(GlobalOpts =
  GlobalOpts))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{reactives}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
Module to change EIC options
}
