% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_EICOptionsModule.R
\name{EICOptionsModuleUI}
\alias{EICOptionsModuleUI}
\title{EICOptionsModuleUI}
\usage{
EICOptionsModuleUI(id)
}
\arguments{
\item{id}{id of the module}
}
\description{
Module to change EIC options
}
