% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_msdata.R
\name{EICplot}
\alias{EICplot}
\title{EICplot}
\usage{
EICplot(EIClistItem = res[[1]], cx = 1, ylim = c(0,
  max(unlist(EIClistItem[, "tic"]))), xlim = c(min(unlist(EIClistItem[,
  "rt"])), max(unlist(EIClistItem[, "rt"])))/60,
  legendtext = paste(sub("^([^.]*).*", "\\\\1",
  basename(row.names(EIClistItem)))), colr = topo.colors(nrow(EIClistItem),
  alpha = 1), heading = "test", relto = NULL, TIC = T, single = F,
  midline = 100)
}
\arguments{
\item{EIClistItem}{item from a list of EICs from Mosaic:multiEIC}

\item{ylim}{numeric(2) min and max visible rt value (in seconds)}

\item{xlim}{numeric(2) min and max visible intensity value (in seconds)}

\item{legendtext}{character() with item for each shown EIC for the plot legend}

\item{colr}{color range (actual vector of color values)}

\item{heading}{plot title}

\item{relto}{normalize intensities to this number}

\item{TIC}{if TRUE plot TIC}

\item{single}{TRUE if this is not part of a composite plot}
}
\description{
generate one EIC plot for multiple files
}
