% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_plotting_EICs.R
\name{EICtitles}
\alias{EICtitles}
\title{EICtitles}
\usage{
EICtitles(mzs, rts, ppm)
}
\arguments{
\item{mzs}{vector of mz values (not ranges)}

\item{rts}{vector of retention time values (not ranges)}

\item{ppm}{mz window +/- mzmid in ppm that will be plotted}
}
\description{
helper function to generate titles for EICgeneral
}
