% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_msdata.R
\name{exIntensities}
\alias{exIntensities}
\title{multiEIC}
\usage{
exIntensities(rawfile = rawdata[[1]], mz = tb$mz, ppm = 5,
  rtw = data.frame(rta[[1]]$rtmin - 5, rta[[1]]$rtmax + 5))
}
\description{
Get intensities from a table of mz/rt features from an EICraw- created object. Requires rtmin and rtmax
}
\details{
New version with apply instead of loops 5x faster for EICs
}
