% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_EICs.R
\name{fastPeakShapes}
\alias{fastPeakShapes}
\title{fastPeakShapes}
\usage{
fastPeakShapes(rawdata, mz, ppm, rtw, workers = 1)
}
\arguments{
\item{mz}{numeric (): m/z values (same length as nrow(rtw)).}

\item{ppm}{m/z width (+/- ppm from values defined in mz)}

\item{rtw}{a data.frame with minimum (column 1) and maximum (column 2) retention time values (in seconds) in each row.}

\item{rawfile}{an xcmsRaw object}
}
\description{
Calculate peak quality for features in a feature table across a set of MS data files
}
