% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Labelfinder.R
\name{featlistCompare}
\alias{featlistCompare}
\title{featlistCompare}
\usage{
featlistCompare(
  reflist = reflist,
  complist = complist,
  mzdiff = 2 * 1.00335,
  pktol = 2,
  rtdiff = 10,
  ppm = 5
)
}
\arguments{
\item{reflist}{data.frame with molecular features and columns \code{mz, rt, rtmin, rtmax}}

\item{complist}{data.frame with molecular features and columns \code{mz, rt, rtmin, rtmax}}

\item{mzdiff}{expected mz difference between features of interest in \code{reflist} and \code{complist}}

\item{pktol}{maximum fold difference in retention time peak width between matched peaks}

\item{rtdiff}{maximum retention time difference (same unit as \code{rt} columns in \code{reflist} and \code{complist}, typically seconds)}

\item{ppm}{maximum relative \code{mz} difference between matched peaks in parts per million (ppm)}
}
\value{
a list of data.frames
}
\description{
compare feature tables and find molecular features that are shifted from the 
mz values in \code{reflist} by \code{mzdiff} and are within \code{rtdiff}.
}
\details{
Will first find matches in \code{complist} and then go back and find the shifted
mz in \code{reflist} as well as the unshifted mass in \code{complist}.
}
\examples{
\dontrun{
 MseekExamplePreload(data = TRUE, tables = TRUE)
demo_reflist = data.frame(mz = 101:200,
                     rt = 201:300,
                     rtmin = 198:297,
                     rtmax = 201:300)

demo_complist = demo_reflist

demo_complist$mz <- demo_complist$mz + rnorm(100, 2*1.00335, sd = 2*1.00335/600)

demo_complist[101:150,] <- demo_complist[1:50,]

demo_complist$mz[101:150] <- demo_complist$mz[101:150] + 1e-10


compareResults <- featlistCompare(reflist = demo_reflist, complist = demo_complist)

  MseekExamplePreload(data = FALSE, tables = TRUE)
compareResultsBig <- featlistCompare(reflist = tab2$df, complist = tab2$df)

}

}
