% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics_modules.R
\name{featurePlotModule}
\alias{featurePlotModule}
\title{featurePlotModule}
\usage{
featurePlotModule(input, output, session, FT, rname, heading = "Default")
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{FT}{Mosaic Feature Table object}

\item{rname}{row names of rows in FT selected for plotting}

\item{heading}{}
}
\description{
Quickplots (ggplot-based plots of feature properties)
}
