% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics_modules.R
\name{featurePlotModuleUI}
\alias{featurePlotModuleUI}
\title{featurePlotModuleUI}
\usage{
featurePlotModuleUI(id)
}
\arguments{
\item{id}{}
}
\description{
Quickplots (ggplot-based plots of feature properties) UI
}
