% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_featureReport.R
\name{featureReport}
\alias{featureReport}
\title{featureReport}
\usage{
featureReport(pdf_settings = list(file = "testReport.pdf", width = NULL,
  height = NULL), layout_settings = 3, EICplots = NULL, MS1 = NULL,
  MS2 = NULL, tree = NULL)
}
\arguments{
\item{pdf_settings}{if NULL, plot to current device, otherwise takes list with items file, height and width. If height and/ or width are NULL, plot to A4 format pdf with some safeguards increasing width and hieght as necessary}

\item{layout_settings}{list of arguments passed to layout(), OR an integer defining how many columns to use for the EIC plots}

\item{EICplots}{list of arguments passed to EICgeneral()}

\item{MS1}{list of arguments passed to specplot()}

\item{MS2}{list of arguments passed to specplot()}

\item{tree}{list of arguments passed to plotTree()}
}
\description{
featureReport
}
