% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_FeatureReportModule.R
\name{FeatureReportModuleUI}
\alias{FeatureReportModuleUI}
\title{FeatureReportModuleUI}
\usage{
FeatureReportModuleUI(id)
}
\arguments{
\item{id}{id of the Module}
}
\description{
FeatureReportModuleUI
}
