% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_getters.R
\name{FeatureTable<-}
\alias{FeatureTable<-}
\title{FeatureTable<-}
\usage{
FeatureTable(x, ...) <- value
}
\arguments{
\item{x}{the \code{values} object}

\item{value}{the value that gets set}

\item{replace}{}
}
\description{
Getter and setter methods to retireve a Feature Table from \code{values}
}
