% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container_FeatureTableContainer.R
\name{FeatureTableContainerUI}
\alias{FeatureTableContainerUI}
\title{FeatureTableContainerUI}
\usage{
FeatureTableContainerUI(id)
}
\arguments{
\item{id}{}
}
\description{
Module to apply filters to a featureTable (UI)
}
