% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_formatting_utilities.R
\name{filenamemaker}
\alias{filenamemaker}
\title{Make filenames for exported .csv or .pdf files}
\usage{
filenamemaker(projectName, FT)
}
\arguments{
\item{projectName}{ProjectName used as prefix}

\item{FT}{Mseek's featureTable reactiveValues, as returned by 
\code{\link{FeatureTable}(values)} (or a list with same structure)}
}
\value{
a character string with an informative filename that includes 
filter criteria
}
\description{
TODO: reimplement this
Generate a filename from project name and filter criteria
}
