% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_FilePathModule.R
\name{FilePathModule}
\alias{FilePathModule}
\title{FilePathModule}
\usage{
FilePathModule(input, output, session, filepaths = reactive({    
  .MosaicOptions$filePaths }), label = "Select Folder",
  description = "Select Folder", displayFolder = F, selectFile = F,
  pattern = "")
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{filepaths}{Named character vector of parent folder locations}

\item{label}{button label}

\item{description}{button description on hover}

\item{displayFolder}{show filepath (boolean)}
}
\description{
module to find filepaths for select directory operations, can display just a button or a button with a displayof the filepath.
}
