% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container_FilterContainer.R
\name{FilterContainer}
\alias{FilterContainer}
\title{FilterContainer}
\usage{
FilterContainer(input, output, session,
  values = reactiveValues(featureTables = featureTables, MainTable =
  MainTable))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}
}
\description{
Module to apply filters to a featureTable (UI)
}
