% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container_FilterContainer.R
\name{FilterContainerUI}
\alias{FilterContainerUI}
\title{FilterContainerUI}
\usage{
FilterContainerUI(id)
}
\arguments{
\item{id}{}
}
\description{
Module to apply filters to a featureTable (UI)
}
