% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_FilterModule.R
\name{FilterModule}
\alias{FilterModule}
\title{FilterModule}
\usage{
FilterModule(input, output, session,
  values = reactiveValues(featureTables = featureTables, MultiFilter =
  MultiFilter), static = list(lab = "Filter"), initValues = list(active
  = F, filter = TRUE, colSelected = NULL, summary = NULL, numeric = T,
  minSel = NULL, maxSel = NULL, modeSel = NULL, txtSel = NULL, minSelInit =
  NULL, maxSelInit = NULL, modeSelInit = NULL, txtSelInit = NULL,
  loadingFilters = F))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{static}{Import data from the shiny session}

\item{reactives}{Import data from the shiny session}
}
\description{
Module to apply filters to a featureTable (UI)
}
