% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_peakPicking.R
\name{find_max}
\alias{find_max}
\title{find_max}
\usage{
find_max(EIC, ..., getRTs = F)
}
\arguments{
\item{EIC}{a single EIC object (list item of multiEIC output)}

\item{...}{further arguments passed to peakPick::peakpick}

\item{getRTs}{return retention time values instead of indices of maxima}
}
\description{
find maxima in an EIC object using peakPick::peakpick
}
