% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_peakPicking.R
\name{find_min}
\alias{find_min}
\title{find_min}
\usage{
find_min(EIC, ..., getRTs = F)
}
\arguments{
\item{EIC}{a single EIC object (list item of multiEIC output)}

\item{...}{further arguments passed to peakPick::peakpick}

\item{getRTs}{return retention time values instead of indices of minima}
}
\description{
find minima in an EIC object using peakPick::peakpick
}
