% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_peakPicking.R
\name{find_peaks}
\alias{find_peaks}
\title{find_peaks}
\usage{
find_peaks(EIC, maxdetect = list(neighlim = 4, deriv.lim = 0.5,
  peak.min.sd = 0.7, peak.npos = 10), mindetect = list(neighlim = 4,
  deriv.lim = 0.5, peak.min.sd = 0.7, peak.npos = 10))
}
\arguments{
\item{EIC}{a single EIC object (list item of multiEIC output)}

\item{maxdetect}{arguments passed to find_max}

\item{mindetect}{arguments passed to find_min}
}
\description{
find minima in an EIC object using peakPick::peakpick
}
