% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_FindMS2scansModule.R
\name{FindMS2ScansModule}
\alias{FindMS2ScansModule}
\title{FindMS2ScansModule}
\usage{
FindMS2ScansModule(input, output, session, values = reactiveValues(MSData
  = MSData, featureTables = featureTables, MainTable = MainTable),
  static = list(tooltip =
  "Find MS2 scans for all parent m/zs in feature table", label =
  "Find MS2 scans"))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
server module for saving Tables
}
