% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_FindMS2scansModule.R
\name{FindMS2ScansModuleUI}
\alias{FindMS2ScansModuleUI}
\title{FindMS2ScansModuleUI}
\usage{
FindMS2ScansModuleUI(id)
}
\arguments{
\item{id}{id of the shiny module}
}
\description{
FindMS2ScansModuleUI
}
