% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset_MSData.R
\name{findMSnScans}
\alias{findMSnScans}
\alias{findMSnScans.reactivevalues}
\alias{findMSnScans.OnDiskMSnExp}
\title{findMSnScans}
\usage{
findMSnScans(data, ...)

\method{findMSnScans}{reactivevalues}(data, ...)

\method{findMSnScans}{OnDiskMSnExp}(data, mz, rt, ppm = 5, mzwid = 0,
  rtwid = 30, MSlevel = 2)
}
\arguments{
\item{data}{an OnDiskMSnExp or MseekValues (reactivevalues) object}

\item{mz}{parentMZ value to search for. mz and rt can be vectors, but need to be of equal length}

\item{rt}{retention time to search for. mz and rt can be vectors, but need to be of equal length}

\item{ppm}{mz tolerance in ppm}

\item{mzwid}{mz tolerance in absolute mz values}

\item{rtwid}{rt tolerance in seconds}

\item{MSlevel}{which mz level to}
}
\description{
Get a list of MS scans based on parent m/z and retention time
Will return all scans with parentMZ within mz +/- ppm OR mzwid, AND retentionTime within rt +/- rtwid.
}
