% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_FindPatternsModule.R
\name{FindPatternsModule}
\alias{FindPatternsModule}
\alias{FindPatternsModuleUI}
\title{FindPatternModule}
\usage{
FindPatternsModule(input, output, session, values, useActionLink = F)

FindPatternsModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{useActionLink}{if TRUE, will use an \code{actionLink} instead 
of an \code{actionButton} to open the modal Dialog}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
nothing, but modifies values
}
\description{
Module for loading an entire Metaboseek session
}
\section{Functions}{
\itemize{
\item \code{FindPatternsModule}: server logic

\item \code{FindPatternsModuleUI}: UI elements
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
