% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_modules.R
\name{findsubgraph}
\alias{findsubgraph}
\title{findsubgraph}
\usage{
findsubgraph(id, graphlist)
}
\arguments{
\item{id}{ID of vertex}

\item{graphlist}{list of graphs, such as those in $subgraphs of objects generated by layout_components_qgraph}
}
\description{
Find the graph in a list of graphs that contains a vertex with a given id.
}
