% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{FTAnalysisParam-class}
\alias{FTAnalysisParam-class}
\title{Parameter class for Feature Table analysis}
\description{
Objects of this class hold parameters for feature table analysis,
to be used with \code{\link{analyzeFT}()}
}
\details{
FTAnalysisParam
}
\section{Slots}{

\describe{
\item{\code{intensities}}{the intensity column names, before normalization 
(without __norm suffix), will be automatically renamed if useNormalized.}

\item{\code{groups}}{named list of non-normalized intensity columns listed by group 
(as supplied by $anagroupnames of MseekFT objects), will be automatically 
renamed if useNormalized.}

\item{\code{analyze}}{character vector to select the analyses to be run: 
"Basic analysis", "clara_cluster", "t-test", "Peak shapes"}

\item{\code{normalize}}{normalze intensity columns}

\item{\code{useNormalized}}{use normalized values for analyses; will trigger 
normalize if there is no normalized data available for all selected 
intensity columns}

\item{\code{logNormalized}}{if TRUE, applies a log10 to intensity values after normalization}

\item{\code{.files}}{character() with file paths to the MS data files to be used
in an analysis}

\item{\code{ppm}}{ppm range for peak shape analysis}

\item{\code{controlGroup}}{control group for foldChange (part of Basic analysis) 
analysis (optional)}

\item{\code{numClusters}}{number of clusters for clara_clusters analysis}

\item{\code{mzMatchParam}}{list of parameters passed to mass}

\item{\code{workers}}{number of workers to use for multithreaded analyses}

\item{\code{normalizationFactors}}{normalizationFactors vector with factors to apply to each column for normalization.}

\item{\code{zeroReplacement}}{value to replace zeros with}

\item{\code{replaceNAs}}{Numeric value to replace NA values with}

\item{\code{p.adjust.method}}{method to adjust p-values (see \code{\link[stats]{p.adjust}()})}
}}

