% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{FTAnalysisParam}
\alias{FTAnalysisParam}
\title{Constructor for FTAnalysisParam class}
\usage{
FTAnalysisParam(
  intensities = character(),
  groups = list(),
  .files = character(),
  analyze = c("Basic analysis", "clara_cluster", "t-test", "Peak shapes",
    "Fast peak shapes", "PCA features", "PCA samples", "mzMatch", "Calculate M"),
  normalize = T,
  useNormalized = T,
  logNormalized = F,
  normalizationFactors = NULL,
  zeroReplacement = 1,
  replaceNAs = 0,
  ppm = 5,
  controlGroup = NULL,
  numClusters = 2,
  mzMatchParam = list(db = "smid-db_pos.csv", ppm = 5, mzdiff = 0.001),
  workers = 1
)
}
\arguments{
\item{intensities}{the intensity column names, before normalization 
(without __norm suffix), will be automatically renamed if useNormalized.}

\item{groups}{named list of non-normalized intensity columns listed by group 
(as supplied by $anagroupnames of MseekFT objects), will be automatically 
renamed if useNormalized.}

\item{.files}{character() with file paths to the MS data files to be used
in an analysis}

\item{analyze}{character vector to select the analyses to be run: 
"Basic analysis", "clara_cluster", "t-test", "Peak shapes"}

\item{normalize}{normalze intensity columns}

\item{useNormalized}{use normalized values for analyses; will trigger 
normalize if there is no normalized data available for all selected 
intensity columns}

\item{logNormalized}{if TRUE, applies a log10 to intensity values after normalization}

\item{ppm}{ppm range for peak shape analysis}

\item{controlGroup}{control group for foldChange (part of Basic analysis) 
analysis (optional)}

\item{numClusters}{number of clusters for clara_clusters analysis}

\item{mzMatchParam}{list of parameters passed to mass}

\item{workers}{number of workers to use for multithreaded analyses}
}
\value{
a \code{\link{FTAnalysisParam-class}} object
}
\description{
Construct a \code{FTAnalysisParam} object with parameters for 
use with \code{\link{analyzeFT}()}
}
\details{
FTAnalysisParam
}
