% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_FTHistoryWidget.R
\name{FTHistoryWidget}
\alias{FTHistoryWidget}
\title{FTHistoryWidget}
\usage{
FTHistoryWidget(input, output, session, FT = reactive({     NULL }))
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}
}
\description{
One-button module to display and export the process history of MseekFT objects.
}
\section{Functions}{
\itemize{
\item \code{FTHistoryWidget}: server logic
}}

