% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{FTProcessHistory-class}
\alias{FTProcessHistory-class}
\alias{show,FTProcessHistory-method}
\alias{hasError,FTProcessHistory-method}
\alias{error,FTProcessHistory-method}
\title{Parameter class for Feature Table analysis}
\usage{
\S4method{show}{FTProcessHistory}(object)

\S4method{hasError}{FTProcessHistory}(object)

\S4method{error}{FTProcessHistory}(object)
}
\arguments{
\item{fileNames}{names of files used in this analysis step (a more human-readable 
variant of the fileIndex slot)}
}
\description{
Objects of this class hold parameters for feature table analysis,
to be used with \code{\link{analyzeFT}()}
}
\details{
FTProcessHistory
}
\section{Slots}{

\describe{
\item{\code{error}}{a named list of character() vectors, reporting errors that 
occured in this analysis step}

\item{\code{changes}}{did changes occur on the associated object. If FALSE, this step
did not result in relevant changes and could be dropped for reporting}

\item{\code{inputDFhash}}{character digest of the data.frame object before this analysis step}

\item{\code{outputDFhash}}{character digest of the data.frame object after this analysis step}

\item{\code{sessionInfo}}{a \code{\link[utils]{sessionInfo}} object, should be genereated
at time of the recorded event and at least once in every session (by default,
will be populated by load and constructor methods for MseekFT class).}
}}

