% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{FTProcessHistory}
\alias{FTProcessHistory}
\title{Constructor for FTProcessHistory class}
\usage{
FTProcessHistory(
  error = list(),
  changes = TRUE,
  inputHash = NULL,
  outputHash = NULL,
  sessionInfo = NULL,
  fileNames = character(),
  processingTime = NA_real_,
  msLevel = NA_integer_,
  ...
)
}
\arguments{
\item{error}{a named list of character() vectors, reporting errors that 
occured in this analysis step}

\item{changes}{did changes occur on the associated object. If FALSE, this step
did not result in relevant changes and could be dropped for reporting}

\item{inputHash}{character \code{\link{MseekHash}} of the object before this analysis step}

\item{outputHash}{character \code{\link{MseekHash}} of the object after this analysis step}

\item{sessionInfo}{a \code{\link[utils]{sessionInfo}} object, should be generated
at time of the recorded event and at least once in every session (by default,
will be populated by load and constructor methods for MseekFT class).}

\item{fileNames}{names of files used in this analysis step (a more human-readable 
variant of the fileIndex slot)}

\item{processingTime}{time spent on processing this step, in seconds}

\item{msLevel, ...}{additional arguments passed to \code{\link[xcms]{XProcessHistory}()}}
}
\value{
a \code{\link{FTProcessHistory-class}} object
}
\description{
Construct a \code{FTProcessHistory} object with parameters for 
use with \code{\link{analyzeFT}()}
}
\details{
FTProcessHistory
}
\examples{
FTProcessHistory(error = list(),
changes = TRUE,
inputHash = NULL,
outputHash = NULL,
sessionInfo = utils::sessionInfo(),
info = "Example")

}
