% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset_featureTables.R
\name{FTselection}
\alias{FTselection}
\alias{FTselection.reactivevalues}
\title{FTselection}
\usage{
FTselection(x, ...)

\method{FTselection}{reactivevalues}(x, liveView = F)
}
\arguments{
\item{x}{a MseekValues (reactivevalues) object}

\item{...}{additional arguments passed to S3 methods}

\item{liveView}{if TRUE, will get data from the liveView of the 
MainTable directly (default: FALSE)}
}
\value{
a data.frame that is a subset of \code{FeatureTable(x)$df}
}
\description{
get the rows that are currently selected in MainTable window.
}
\details{
By default, will get its data from the underlying Feature Table,
 not using changes (such as comments) made immediately prior to selection, 
 but including all columns, including those not visible in the current table view.
}
