% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{FunParam-class}
\alias{FunParam-class}
\alias{FTAnalysisParam-class}
\title{Parameter class for Feature Table analysis}
\description{
Objects of this class hold parameters for any function call.
}
\details{
FunParam
}
\section{Slots}{

\describe{
\item{\code{fun}}{function name}

\item{\code{args}}{arguments to fun, as supplied to fun.}

\item{\code{longArgs}}{parameters fed to the function that are excessively large 
should be summarized using \code{summary()} and then listed in this slot.}
}}

