% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{FunParam}
\alias{FunParam}
\title{FunParam}
\usage{
FunParam(fun = character(), args = list(), longArgs = list())
}
\arguments{
\item{fun}{\code{character} representing the name of the function.}

\item{args}{named \code{list} with the arguments to the function \code{fun}.}

\item{longArgs}{named \code{list} of large arguments to the function 
\code{fun}, summarized using \code{summary()}.}
}
\value{
returns a \code{FunParam} object.
}
\description{
FunParam
}
