% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMosaic.R
\name{get_common_dir}
\alias{get_common_dir}
\title{Get common root folder of file paths}
\usage{
get_common_dir(paths, delim = "/")
}
\arguments{
\item{paths}{vector of paths}

\item{delim}{folder delimiter}
}
\description{
From: https://rosettacode.org/wiki/Find_common_directory_path#R
}
