% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_Scans.R
\name{getAllScans}
\alias{getAllScans}
\title{getAllScans}
\usage{
getAllScans(scanlist, MSData, removeNoise = NULL)
}
\arguments{
\item{scanlist}{as returned by makeScanlist2()}

\item{MSData}{list of xcmsRaw objects}
}
\description{
get a list of all scans in a scanlist
}
