% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset_featureTables.R
\name{getFilters}
\alias{getFilters}
\title{getFilters}
\usage{
getFilters(x, activeOnly = TRUE)
}
\arguments{
\item{x}{a MseekValues (reactivevalues) object}

\item{activeOnly}{if TRUE, gets only active Filters}
}
\value{
a list of filters
}
\description{
get the currently applied filters in a Metaboseek session
}
