% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_GetIntensitiesModule.R
\name{GetIntensitiesModuleUI}
\alias{GetIntensitiesModuleUI}
\title{GetIntensitiesModuleUI}
\usage{
GetIntensitiesModuleUI(id)
}
\arguments{
\item{id}{id of the shiny module}
}
\description{
GetIntensitiesModuleUI
}
