% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_peakPicking.R
\name{getpeaks}
\alias{getpeaks}
\title{getpeaks}
\usage{
getpeaks(EIC, findProps = list(maxdetect = list(neighlim = 4, deriv.lim =
  0.5, peak.min.sd = 0.7, peak.npos = 10), mindetect = list(neighlim = 4,
  deriv.lim = 0.5, peak.min.sd = 0.7, peak.npos = 10)),
  mergeProps = list(minint = 10000, minrelint = 0.02, topN = 3))
}
\arguments{
\item{EIC}{a single EIC object (list item of multiEIC output), or a list of EIC items (multiEIC output)}

\item{findProps}{arguments passed to find_peaks}

\item{mergeProps}{arguments passed to mergepeaks}
}
\description{
find and then merge peaks for an EIC object. wrapper function for find_peaks and mergepeaks
}
