% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_SIRIUS.R
\name{getSirius}
\alias{getSirius}
\title{getSirius}
\usage{
getSirius(outfolder, splash, ts)
}
\arguments{
\item{outfolder}{folder with SIRIUS results}

\item{splash}{splash tag to look for}

\item{ts}{timestamp to look for}
}
\description{
get SIRIUS results for a given spectrum
}
