% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_GetSiriusModule.R
\name{GetSiriusModule}
\alias{GetSiriusModule}
\title{GetSiriusModule}
\usage{
GetSiriusModule(input, output, session, values = reactiveValues(),
  reactives = reactive({     list(outfolder =
  file.path(values$GlobalOpts$siriusFolder, "METABOseek"), ms2 =
  MergedSpecs, instrument = "orbitrap", parentmz = reactives()$mz, rt =
  reactives()$rt, comments = reactives()$comments, ion = "[M+H]+", charge =
  1, fingerid = T, scanindices = reactives()$scanlist, sirpath =
  file.path(values$GlobalOpts$siriusFolder, "sirius-console-64.exe"),
  moreOpts = "") }))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{reactives}{Import data from the shiny session}
}
\description{
server module for interactive mass spectrum view
}
