% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_GetSiriusModule.R
\name{GetSiriusModuleUI}
\alias{GetSiriusModuleUI}
\title{GetSiriusModuleUI}
\usage{
GetSiriusModuleUI(id)
}
\arguments{
\item{id}{id to be used in ns()}
}
\description{
UI module for interactive SIRIUS interface
}
