% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_SIRIUS.R
\name{getSiriusTree}
\alias{getSiriusTree}
\title{getSiriusTree}
\usage{
getSiriusTree(paths, formula)
}
\arguments{
\item{paths}{list generated by \code{\link{getSirius}()}}

\item{formula}{molecular formula prediction to get details for}
}
\value{
Returns a named list with fragmentation tree objects for plotting,
 see |code{Details}
}
\description{
get additional SIRIUS results (fragmentation trees) for a spectrum
}
\details{
\itemize{
\item \code{annotations} the matching file from \code{spectra} subfolder 
as data.frame
\item \code{trees_dot} the matching \code{.dot} tree file, read in by 
\code{DiagrammeR::grViz}
\item \code{trees_json} the matching \code{.json} tree file, read in by 
\code{jsonlite::read_json}
}
}
