% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_getters.R
\name{getSpectra}
\alias{getSpectra}
\alias{getSpectra.reactivevalues}
\alias{getSpectra.OnDiskMSnExp}
\title{getSpectra}
\usage{
getSpectra(data, ...)

\method{getSpectra}{reactivevalues}(data, ...)

\method{getSpectra}{OnDiskMSnExp}(data, index, xcmsRawobject = NULL,
  SpectrumClass = F)
}
\arguments{
\item{data}{an OnDiskMSnExp or MseekValues (reactivevalues) object}

\item{index}{a numeric vector of scan indices}

\item{xcmsRawobject}{list of xcmsRaw objects}

\item{SpectrumClass}{if TRUE, will return a list of Spectum class objects, 
otherwise will return a list of matrices with mz and intensity values}

\item{scanlist}{as returned by makeScanlist2()}
}
\description{
get a list of spectra using indices from the OnDiskMSnExp object
}
