% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_GlobalOptionsModule.R
\name{GlobalOptionsModuleUI}
\alias{GlobalOptionsModuleUI}
\title{GlobalOptionsModuleUI}
\usage{
GlobalOptionsModuleUI(id)
}
\arguments{
\item{id}{id of the shiny module}
}
\description{
GlobalOptionsModuleUI
}
