% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_math_utilities.R
\name{gm_mean}
\alias{gm_mean}
\title{gm_mean}
\usage{
gm_mean(x)
}
\arguments{
\item{x}{numeric vector of positive values}
}
\description{
Adapted from: https://stackoverflow.com/questions/2602583/geometric-mean-is-there-a-built-in
Ignores 0 and NA values!
}
