% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_GroupedEICModule.R
\name{GroupedEICModule}
\alias{GroupedEICModule}
\title{GroupedEICModule}
\usage{
GroupedEICModule(input, output, session, values = reactiveValues(MSData =
  MSData, featureTables = featureTables, GlobalOpts = GlobalOpts,
  projectData = projectData), keys = reactive({     keyin$keyd }))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{reactives}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
server module for loading a Project Folder
}
