% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_class_featureTable.R
\name{history}
\alias{history}
\alias{history.MseekFT}
\alias{history.xsAnnotate}
\alias{history.XCMSnExp}
\title{history}
\usage{
history(x)

\method{history}{MseekFT}(x)

\method{history}{xsAnnotate}(x)

\method{history}{XCMSnExp}(x)
}
\description{
extract a list of \code{ProcessHistory} objects from an object,
 representing changes made to the object.
}
