% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_LabelFinderModule.R
\name{LabelFinderModule}
\alias{LabelFinderModule}
\alias{LabelFinderModuleUI}
\title{LabelFinderModule}
\usage{
LabelFinderModule(input, output, session, values)

LabelFinderModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
Returns its internalValues
}
\description{
Module for finding labeled features
}
\section{Functions}{
\itemize{
\item \code{LabelFinderModule}: Server logic

\item \code{LabelFinderModuleUI}: UI elements
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
