% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_Scans.R
\name{listMS2scans}
\alias{listMS2scans}
\title{listMS2scans}
\usage{
listMS2scans(mz, rt, ppm, rtw, MSData)
}
\arguments{
\item{mz}{expected target (parent) m/z}

\item{rt}{expected target (parent) retention time in seconds}

\item{ppm}{target (parent) m/z tolerance in ppm}

\item{rtw}{target (parent) rt tolerance in seconds}

\item{MSData}{list of xcmsRaw objects}
}
\description{
Searches for MS2 scans with a given parent m/z and directly ceonverts the resulting table into a single string
}
\details{
Also, see \code{\link{Parentsearch}}
}
