% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_LoadDataModule.R
\name{LoadDataModule}
\alias{LoadDataModule}
\title{LoadDataModule}
\usage{
LoadDataModule(input, output, session,
  values = reactiveValues(projectData = projectData, featureTables =
  featureTables, MSData = MSData, GlobalOpts = GlobalOpts))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{reactives}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
server module for loading a Project Folder
}
