% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_formatting_utilities.R
\name{loadMseekSession}
\alias{loadMseekSession}
\title{loadMseekSession}
\usage{
loadMseekSession(values, savedValues)
}
\arguments{
\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{savedValues}{an MseekSession object or a file path to an MseekSession
saved as RDS file.}
}
\value{
nothing, but modifies the global values object
}
\description{
load a Metaboseek session
}
