% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure_functions.R
\name{loadOptions}
\alias{loadOptions}
\title{loadOptions}
\usage{
loadOptions(defaults = F)
}
\arguments{
\item{defaults}{if TRUE, default MosaicOptions are laoded}
}
\description{
Load .MosaicOptions from config file
}
