% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_LoadSessionModule.R
\name{LoadSessionModule}
\alias{LoadSessionModule}
\alias{LoadSessionModuleUI}
\title{LoadSessionModule}
\usage{
LoadSessionModule(input, output, session, values, useActionLink = F)

LoadSessionModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{useActionLink}{if TRUE, will use an \code{actionLink} instead 
of an \code{actionButton} to open the modal Dialog}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}

\item{static}{see \code{Details}}
}
\value{
nothing, but modifies values
}
\description{
Module for loading an entire Metaboseek session
}
\details{
Loads a Metaboseek session from a MseekSession object saved in
 an RDS file with the file extension ".Msks".
}
\section{Functions}{
\itemize{
\item \code{LoadSessionModule}: server logic

\item \code{LoadSessionModuleUI}: UI elements
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
