% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_module.R
\name{LoadTableModule}
\alias{LoadTableModule}
\title{LoadTableModule}
\usage{
LoadTableModule(input, output, session, tag, set = list(title = NULL,
  filetypes = NULL, format = list(header = T, sep = "\\t", quote = "\\"",
  stringsAsFactors = F)))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{tag}{id to be used in ns()}

\item{set}{Import data from the shiny session}
}
\description{
server module for loading Tables
}
