% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container_MainPageContainer.R
\name{MainPageContainer}
\alias{MainPageContainer}
\title{MainPageContainer}
\usage{
MainPageContainer(input, output, session,
  values = reactiveValues(projectData = values$projectData, featureTables
  = values$featureTables, MSData = values$MSData, GlobalOpts =
  values$GlobalOpts))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}
}
\description{
Module to apply filters to a featureTable (UI)
}
