% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container_MainPageContainer.R
\name{MainPageContainerUI}
\alias{MainPageContainerUI}
\title{MainPageContainerUI}
\usage{
MainPageContainerUI(id)
}
\arguments{
\item{id}{}
}
\description{
Module to apply filters to a featureTable (UI)
}
