% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container_MainPlotContainer.R
\name{MainPlotContainer}
\alias{MainPlotContainer}
\title{MainPlotContainer}
\usage{
MainPlotContainer(input, output, session,
  values = reactiveValues(projectData = projectData, featureTables =
  featureTables, MSData = MSData, GlobalOpts = GlobalOpts, MainTable =
  MainTable))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}
}
\description{
Module to apply filters to a featureTable (UI)
}
