% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container_MainPlotContainer.R
\name{MainPlotContainerUI}
\alias{MainPlotContainerUI}
\title{MainPlotContainerUI}
\usage{
MainPlotContainerUI(id)
}
\arguments{
\item{id}{}
}
\description{
Module to apply filters to a featureTable (UI)
}
