% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_class_MSdata.R
\name{makeColorscheme}
\alias{makeColorscheme}
\title{makeColorscheme}
\usage{
makeColorscheme(maingroup = MSD$layouts[[MSD$active]]$grouping,
  colorgroup = MSD$layouts[[MSD$active]]$grouping,
  colrange = "Mseek.colors", transparency = 0.8)
}
\arguments{
\item{maingroup}{main plot grouping scheme (e.g. grouping in rawlayouts)}

\item{colorgroup}{grouping to be used for coloring (e.g. grouping2 in rawlayouts)}

\item{colrange}{function to use to make color range}

\item{transparency}{setting for transparent coloring}
}
\description{
Use two grouing lists to generate a color scheme for groupPlot() and EICgeneral().
}
