% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_graph_utilities.R
\name{makeEdges}
\alias{makeEdges}
\title{makeEdges}
\usage{
makeEdges(speclist = speclist1, parentmasses, mztol = 0.005,
  minpeaks = 6, nonmatched = T)
}
\arguments{
\item{speclist}{(non-nested) list of MS spectra}

\item{parentmasses}{vector of parent m/z values (same length as speclist)}

\item{mztol}{max difference between matched peaks in m/z}

\item{minpeaks}{minimum number of peaks that have to be matched, otherwise returns 0}
}
\description{
Make an edgelist (data.table) using a vector of parentmasses and a list of MS spectra
}
