% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_peakPicking.R
\name{makeRTlist}
\alias{makeRTlist}
\title{makeRTlist}
\usage{
makeRTlist(df, rawdata, ppm = 5, retainColumns = NULL)
}
\arguments{
\item{df}{a data.frame that contains a mz column with mz values of interest}

\item{rawdata}{a list of xcmsRaw objects}

\item{ppm}{ppm tolerance for EIC construction}

\item{retainColumns}{keep and copy these columns from the original df when making the result data.frame}
}
\description{
find and then merge peaks, given a set of mz values in a dataframe and a set of xcmsRaw objects
}
