% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_msdata.R
\name{makeScanlist}
\alias{makeScanlist}
\title{makeScanlist}
\usage{
makeScanlist(splitme, MSData = NULL)
}
\arguments{
\item{splitme}{character string with format filename:scannumber###filename:scannumber###... as found in gnps networking output tables}

\item{MSData}{list of xcmsRaw objects. Only scans from files loaded in this object will be returned}
}
\description{
Make a scan list
}
