% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_Scans.R
\name{makeScanlist2}
\alias{makeScanlist2}
\title{makeScanlist2}
\usage{
makeScanlist2(splitme, MSData = NULL)
}
\arguments{
\item{splitme}{character string with format filename:scannumber scannumber|filename: scannumber... as generated by saveScanlist()}

\item{MSData}{list of xcmsRaw objects. Only scans from files loaded in this object will be returned}
}
\description{
Make a scan list
}
