% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MassShiftsModule.R
\name{MassShiftsModule}
\alias{MassShiftsModule}
\title{MassShiftsModule}
\usage{
MassShiftsModule(input, output, session, values = reactiveValues(MSData =
  MSData))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{reactives}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
Module to apply filters to a featureTable (UI)
}
