% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MassShiftsModule.R
\name{MassShiftsModuleUI}
\alias{MassShiftsModuleUI}
\title{MassShiftsModuleUI}
\usage{
MassShiftsModuleUI(id)
}
\arguments{
\item{id}{}
}
\description{
Module to apply filters to a featureTable (UI)
}
