% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_components.R
\name{mDownloadButton}
\alias{mDownloadButton}
\title{mDownloadButton}
\usage{
mDownloadButton(
  outputId,
  label = "Download",
  class = NULL,
  icon = icon("download"),
  ...
)
}
\arguments{
\item{outputId}{The name of the output slot that the \code{downloadHandler} is assigned to.}

\item{label}{The label that should appear on the button.}

\item{class}{Additional CSS classes to apply to the tag, if any.}

\item{icon}{icon to use on this button (an object generated by \code{\link[shiny]{icon}()})}

\item{...}{Other arguments to pass to the container tag function.}
}
\value{
a \code{\link[shiny]{actionButton}} that can be turned red easily
}
\description{
A \code{\link[shiny]{downloadButton}} with the option to set the icon
}
