% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Labelfinder.R
\name{mergecollectlist}
\alias{mergecollectlist}
\title{mergecollectlist}
\usage{
mergecollectlist(collectlist)
}
\arguments{
\item{collectlist}{a list of list of data.frames from featlistcompare()}
}
\description{
helper function to reorganize featlistcompare() outputs for findLabeledPeaks()
}
