% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_deprecated.R
\name{mergeMSold}
\alias{mergeMSold}
\title{mergeMSold}
\usage{
mergeMSold(speclist, ppm = 5, mergeOnly = F)
}
\arguments{
\item{speclist}{data.frame or matrix containing mz and intensity 
values of a spectrum (mz in column 1)}

\item{ppm}{accuracy}

\item{mergeOnly}{if TRUE, only the merged spectrum is returned}
}
\description{
Merge MS spectra by combining peaks that are within a ppm distance
}
\details{
NOTE: If multiple peaks inside a spectrum match another spectrum,
 only the one with higher(?) mz will be retained
}
