% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_peakPicking.R
\name{mergepeaks}
\alias{mergepeaks}
\title{mergepeaks}
\usage{
mergepeaks(pl, minint = 10000, minrelint = 0.02, topN = 3)
}
\arguments{
\item{pl}{a peaklist, such as list items returned by find_peaks() (or its rbind product)}

\item{minint}{minimum intensity for a peak to be retained (before merging)}

\item{minrelint}{minimum intensity of a peak to be retained (before merging), relative to largest peak}

\item{topN}{number of highest intensity peaks retained after merging}
}
\description{
merge peaks in a peaklist
}
