% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_msdata.R
\name{mosaic.colors}
\alias{mosaic.colors}
\title{mosaic.colors}
\usage{
mosaic.colors(n, alpha)
}
\arguments{
\item{n}{number of colors}

\item{alpha}{transparency}
}
\description{
custom color spectrum using color brewer Set1 colors plus topo.colors; good color discrimination up to n = 13
}
