% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container_MosaicContainer.R
\name{MosaicContainer}
\alias{MosaicContainer}
\title{MosaicContainer}
\usage{
MosaicContainer(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}
}
\description{
Module that contains the entire Mosaic program
}
