% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container_MosaicContainer.R
\name{MosaicContainerUI}
\alias{MosaicContainerUI}
\title{MosaicContainerUI}
\usage{
MosaicContainerUI(id)
}
\arguments{
\item{id}{}
}
\description{
Module that contains the entire Mosaic program
}
