% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_data_analysis.R
\name{MosCluster}
\alias{MosCluster}
\title{MosCluster}
\usage{
MosCluster(method = "clara", ...)
}
\arguments{
\item{method}{"clara" (from the cluster package)}

\item{...}{additional arguments to the clustering method}

\item{mx}{matrix (of intensity values)}
}
\description{
Cluster a matrix
}
