% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MS2BrowserModule.R
\name{MS2BrowserModuleUI}
\alias{MS2BrowserModuleUI}
\title{MS2BrowserModule}
\usage{
MS2BrowserModuleUI(id)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{tag}{id to be used in ns()}

\item{set}{Import data from the shiny session}
}
\description{
server module for loading Tables
}
