% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_formatting_utilities.R
\name{Mseek.colors}
\alias{Mseek.colors}
\title{Mseek.colors}
\usage{
Mseek.colors(n, alpha)
}
\arguments{
\item{n}{number of colors}

\item{alpha}{transparency}
}
\value{
a character vector representing \code{n} colors.
}
\description{
custom color spectrum using color brewer Set1 colors plus topo.colors; 
good color discrimination up to n = 13
}
