% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container_MseekContainer.R
\name{MseekContainer}
\alias{MseekContainer}
\title{MseekContainer}
\usage{
MseekContainer(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}
}
\description{
Module that contains the entire METABOseek program
}
