% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_documentation.R
\name{MseekContainers}
\alias{MseekContainers}
\alias{MseekContainersUI}
\title{MseekContainers}
\usage{
MseekContainers(input, output, session, values)

MseekContainersUI(id)
}
\arguments{
\item{input, output, session}{arguments necessary for use with \link[shiny]{callModule}()}

\item{values}{a \link[shiny]{reactiveValues} object that in effect gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \link[shiny]{NS}() (must match \code{id} in \link[shiny]{callModule} for the server Module)}
}
\description{
Regular modules in Metaboseek take a `shiny::reactiveValues` object called `values` as their first argument.
 `values` is the primary interface allowing communication between modules and is generated by `MseekMinimalServer()`.
  Modules expect this input to work properly, and will only work in shiny apps that provide the proper environment set up by `MseekMinimalServer()`.
}
\section{Functions}{
\itemize{
\item \code{MseekContainers}: generic server logic example for MseekContainers

\item \code{MseekContainersUI}: generic UI example for MseekContainers
}}

