% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container_MseekContainer.R
\name{MseekContainerUI}
\alias{MseekContainerUI}
\title{MseekContainerUI}
\usage{
MseekContainerUI(id)
}
\arguments{
\item{id}{}
}
\description{
Module that contains the entire METABOseek program
}
