% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_components.R
\name{MseekExamplePreload}
\alias{MseekExamplePreload}
\title{MseekExamplePreload}
\usage{
MseekExamplePreload(tables = T, data = T)
}
\arguments{
\item{tables}{load example tables?}

\item{data}{load example MS data?}
}
\description{
Loads example data into the parent environment for testing and development purposes
}
