% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_MseekFamily.R
\docType{methods}
\name{MseekFamily}
\alias{MseekFamily}
\alias{addProcessHistory}
\alias{processHistory,MseekFamily-method}
\alias{processHistory}
\alias{previousStep,MseekFamily-method}
\alias{previousStep}
\alias{is.MseekFT}
\alias{is.MseekGraph}
\alias{is.MseekFamily}
\alias{searchFunParam,MseekFamily-method}
\alias{groupingTable,MseekFamily-method}
\alias{groupingTable<-,MseekFamily,data.frame-method}
\alias{intensityCols,MseekFamily-method}
\alias{intensityCols<-,MseekFamily-method}
\alias{hasAdjustedRtime,MseekFamily-method}
\alias{hasAdjustedRtime}
\alias{MseekHash,MseekFamily-method}
\alias{rename,MseekFamily-method}
\alias{rename}
\alias{FTFilter,MseekFT-method}
\alias{FTFilter}
\title{MseekFamily}
\usage{
\S4method{addProcessHistory}{MseekFamily,ProcessHistory}(object, ph)

\S4method{processHistory}{MseekFamily}(object)

\S4method{previousStep}{MseekFamily}(object)

is.MseekFT(x)

is.MseekGraph(x)

is.MseekFamily(x)

\S4method{searchFunParam}{MseekFamily}(object, fun = "", index = FALSE)

\S4method{groupingTable}{MseekFamily}(object)

\S4method{groupingTable}{MseekFamily,data.frame}(object) <- value

\S4method{intensityCols}{MseekFamily}(object)

\S4method{intensityCols}{MseekFamily}(object) <- value

\S4method{hasAdjustedRtime}{MseekFamily}(object)

\S4method{MseekHash}{MseekFamily}(object)

\S4method{rename}{MseekFamily}(object, name)

\S4method{FTFilter}{MseekFT}(object, filters = list(),
  sortBy = character(), decreasing = TRUE)
}
\arguments{
\item{object}{an \code{MseekFT} object.}

\item{ph}{a \code{ProcessHistory} object}

\item{x}{object to test for class}

\item{fun}{character(1), function name to look for}

\item{index}{get index in list instead of actual value}

\item{value}{for \code{groupingTable<-}: a data.frame with columns 
\code{Column} and \code{Group}. For \code{intensityCols<-}: a character vector of column names}

\item{name}{file path to write to}

\item{filters}{a list of filters}

\item{sortBy}{sort by this column}

\item{decreasing}{logical to specify if sort should be decreasing.}
}
\value{
The \code{addProcessHistory} method returns the input object with the
provided \code{\link{ProcessHistory}} appended to the process history.

For \code{hasAdjustedRtime}: A logical, indicating if the object 
contains retention time correction information.
}
\description{
The MseekFamily of classes includes the \code{MseekFT} and 
\code{MseekGraph} S3 classes. Many methods described here will work on both classes

\code{addProcessHistory}: adds (appends) a single \code{\link{ProcessHistory}}
 object to the \code{.processHistory} slot. Copied description and Method 
 template for \code{addProcessHistory} from \code{xcms}.

\code{processHistory}: extract a list of \code{ProcessHistory} objects from an object,
 representing changes made to the object.

\code{previousStep}: extract the most recent \code{ProcessHistory} object from an object,
 representing the last recorded changes made to the object.

\code{searchFunParam}: extract the \code{ProcessHistory} objects
which were generated by a function named like \code{fun} (looking in the 
object@param@fun slots from all \code{ProcessHistory} objects).

\code{groupingTable, groupingTable<-}: Get or set the Mseek intensity column grouping

\code{intensityCols, intensityCols<-}: Get or set the Mseek intensity column names

\code{hasAdjustedRtime}: check if the object contains retention time
correction information.

\code{rename}: rename a \code{MseekFamily} object

\code{FTFilter}: apply a list of filters to a MseekFT object
}
