% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R, R/methods_MseekFT.R
\docType{methods}
\name{addProcessHistory,MseekFT,ProcessHistory-method}
\alias{addProcessHistory,MseekFT,ProcessHistory-method}
\alias{addProcessHistory}
\alias{processHistory,MseekFT-method}
\alias{processHistory}
\alias{previousStep,MseekFT-method}
\alias{previousStep}
\alias{\code{previousStep()}}
\alias{extract}
\alias{the}
\alias{most}
\alias{recent}
\alias{\code{ProcessHistory}}
\alias{object}
\alias{from}
\alias{an}
\alias{object,}
\alias{representing}
\alias{last}
\alias{recorded}
\alias{changes}
\alias{made}
\alias{to}
\alias{object.}
\alias{searchFunParam,MseekFT-method}
\alias{groupingTable,MseekFT-method}
\alias{groupingTable<-,MseekFT,data.frame-method}
\alias{intensityCols,MseekFT-method}
\alias{intensityCols<-,MseekFT-method}
\alias{rename,MseekFT-method}
\alias{rename}
\alias{hasAdjustedRtime,MseekFT-method}
\alias{hasAdjustedRtime}
\alias{FTFilter,MseekFT-method}
\alias{FTFilter}
\title{MseekFT-class}
\usage{
\S4method{addProcessHistory}{MseekFT,ProcessHistory}(object, ph)

\S4method{processHistory}{MseekFT}(object)

\S4method{previousStep}{MseekFT}(object)

\S4method{searchFunParam}{MseekFT}(object, fun = "", index = FALSE)

\S4method{groupingTable}{MseekFT}(object)

\S4method{groupingTable}{MseekFT,data.frame}(object) <- value

\S4method{intensityCols}{MseekFT}(object)

\S4method{intensityCols}{MseekFT}(object) <- value

\S4method{rename}{MseekFT}(object, name)

\S4method{hasAdjustedRtime}{MseekFT}(object)

\S4method{FTFilter}{MseekFT}(object, filters = list(),
  sortBy = charachter(), decreasing = TRUE)
}
\arguments{
\item{object}{an \code{MseekFT} object.}

\item{index}{get index in list instead of actual value}

\item{name}{file path to write to}

\item{filters}{a list of filters}

\item{sortBy}{sort by this column}

\item{decreasing}{logical to specify if sort should be decreasing.}
}
\value{
The \code{addProcessHistory} method returns the input object with the
provided \code{\link{ProcessHistory}} appended to the process history.

For \code{hasAdjustedRtime}: A logical, indicating if the object 
contains retention time correction information.
}
\description{
Copied description and Method template from \code{xcms}.
\code{addProcessHistory}: adds (appends) a single
\code{\link{ProcessHistory}} object to the \code{.processHistory} slot.

\code{processHistory}: extract a list of \code{ProcessHistory} objects from an object,
 representing changes made to the object.

save a \code{MseekFT} object to a file, registering the save event in the processHistory

\code{hasAdjustedRtime}: check if the object contains retention time
correction information.

\code{FTFilter}: apply a list of filters to a MseekFT object
}
\author{
Johannes Rainer, Maximilian Helf
}
