% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_buildMseekFT.R
\docType{methods}
\name{MseekGraphs}
\alias{MseekGraphs}
\alias{buildMseekGraph}
\title{MseekGraphs}
\usage{
\S4method{buildMseekGraph}{MseekFT}(object, cosineThreshold = 0.6)
}
\description{
\code{MseekGraph} objects are \code{\link[igraph]{igraph}} objects
which have layout information and slots that are the same as MseekFT objects,
except that the bulk of data is kept in an \code{igraph} object instead of 
a \code{data.frame}. These objects are read and displayed by the 
\code{\link{NetworkModule}}.
}
