% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_MseekHistoryWidget.R
\name{MseekHistoryWidget}
\alias{MseekHistoryWidget}
\title{MseekHistoryWidget}
\usage{
MseekHistoryWidget(input, output, session, FT = reactive({     NULL }),
  buttonLabel = "History")
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{FT}{an \code{\link{MseekFT}} object inside a \code{reactive()}}

\item{buttonLabel}{text on the button that opens the history modal dialog}
}
\description{
One-button module to display and export the process history of MseekFT objects.
}
\section{Functions}{
\itemize{
\item \code{MseekHistoryWidget}: server logic
}}

