% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_components.R
\name{MseekMinimalServer}
\alias{MseekMinimalServer}
\title{MseekMinimalServer}
\usage{
MseekMinimalServer(data = T, tables = T, diagnostics = T)
}
\arguments{
\item{data}{load example MS data?}

\item{tables}{load example tables?}

\item{diagnostics}{run diagnostics (shinyjs::runcodeServer()) code?}
}
\description{
A minimal server logic for Mseek. Function has to be called in a shiny server function context. Will load example data if it is not present in the parent environment.
Save time by providing the parent environment with objects \code{tab1}, \code{tab2} and \code{MSD} by calling MseekExamplePreload()
}
