% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MseekModalModule.R
\name{MseekModalModule}
\alias{MseekModalModule}
\title{MseekModalModule}
\usage{
MseekModalModule(input, output, session, reactives = reactive({     fp =
  NULL }), static = list(tooltip = "Tooltip", title = "title", label =
  "label", icon = icon("bar-chart", lib = "font-awesome")))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{static}{Import data from the shiny session}

\item{values}{Import data from the shiny session}
}
\description{
server module for saving Tables
}
