% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MseekModalModule.R
\name{MseekModalModuleUI}
\alias{MseekModalModuleUI}
\title{MseekModalModuleUI}
\usage{
MseekModalModuleUI(id)
}
\arguments{
\item{id}{id of the shiny module}
}
\description{
MseekModalModuleUI
}
