% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation_templates.R
\name{MseekWidgets}
\alias{MseekWidgets}
\alias{MseekWidgetsUI}
\title{MseekWidgets}
\usage{
MseekWidgets(input, output, session, reactives, ...)

MseekWidgetsUI(id)
}
\arguments{
\item{input, output, session}{arguments necessary for use with \code{\link[shiny]{callModule}()}}

\item{reactives}{a list, wrapped in a \code{\link[shiny]{reactive}()} expression with settings for this module. See \code{details}.}

\item{...}{additional arguments, as defined for the individual module. For example: constant layout options depending on the context that this module is used in}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} (must match \code{id} in \code{\link[shiny]{callModule}} for the server Module)}
}
\description{
\code{MseekWidgets} should work in a "vanilla" shiny environment, i.e. any shiny app, making it
 easy to use them in other shiny projects without significant overhead. The only expected argument 
 they need is a list supplied as reactives, and potentially additional arguments for which default
  values exist. Widgets are typically wrappers for plotting functions with some additional functionality.
}
\section{Functions}{
\itemize{
\item \code{MseekWidgets}: generic server logic example for MseekWidgets

\item \code{MseekWidgetsUI}: generic UI example for MseekWidgets
}}

