% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_FeatureTable_analysis.R
\name{mttest}
\alias{mttest}
\title{mttest}
\usage{
mttest(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\description{
helper function for multittest, calculating the p.value between two numeric vectors
}
