% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_EICs.R
\name{multiEIC}
\alias{multiEIC}
\title{multiEIC}
\usage{
multiEIC(rawdata, mz, rt, rnames = row.names(mz), byFile = F,
  XIC = F, getgauss = F, RTcorr = NULL, workers = 1,
  quickshapes = F, SN = NULL, scoreBy = "intmean")
}
\arguments{
\item{rawdata}{a (named) list of xcmsRaw objects, e.g. generated by 
METABOseek::loadRawM, OR an MSnbase OnDiskMSnExp object}

\item{mz}{a data.frame with minimum (column 1) and maximum (column 2) 
m/z values in each row.}

\item{rt}{a data.frame with minimum (column 1) and maximum (column 2) 
retention time values (in seconds) in each row.}

\item{rnames}{names of the rows in the result matrix, typically the 
row.names of the feature table, or the filenames if byFile =T.}

\item{byFile}{a data.frame with columns File and Group , holding file 
paths and group names, respectively.}

\item{XIC}{deprecated}

\item{getgauss}{additionally, fit a gauss curve to each EIC (time 
consuming), defaults to FALSE}

\item{RTcorr}{if not NULL, this RTcorr object will be used to 
adjust retention times.}

\item{workers}{number of workers for parallel processing}

\item{quickshapes}{get quickshapes score instead of geegauss results}

\item{SN}{when calculating peak quality with quickshapes, 
reject any peaks with a signal-to-noise ratio than this (intensity range 
within observed rt window, MAX/ MEAN INTENSITY!)}
}
\value{
a list of extracted ion chromatograms that can be read 
by \code{\link{EICplot}}, \code{\link{groupPlot}}, \code{\link{EICgeneral}}
}
\description{
Extract EICs from multiple files for multiple features.
}
