% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MultiEICmodule.R
\name{MultiEICmodule}
\alias{MultiEICmodule}
\title{MultiEICmodule}
\usage{
MultiEICmodule(input, output, session, values = reactiveValues(MSData =
  MSData), keys)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}
}
\description{
server module for interactive EIC view
}
