% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MultiEICmodule.R
\name{MultiEICmoduleUI}
\alias{MultiEICmoduleUI}
\title{MultiEICmoduleUI}
\usage{
MultiEICmoduleUI(id)
}
\arguments{
\item{id}{id to be used in ns()}
}
\description{
UI module for interactive EIC view
}
