% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_EICs.R
\name{multiEICplus}
\alias{multiEICplus}
\title{multiEICplus}
\usage{
multiEICplus(adducts = c(0, 1, 2, 3), mz, rt, ...)
}
\arguments{
\item{adducts}{numeric() of mass shifts}

\item{mz}{a data.frame with minimum (column 1) and maximum (column 2) 
m/z values in each row.}

\item{rt}{a data.frame with minimum (column 1) and maximum (column 2) 
retention time values (in seconds) in each row.}

\item{...}{all other arguments passed on to \code{\link{multiEIC}()}}
}
\value{
a list of extracted ion chromatograms that can be read 
by \code{\link{EICplot}}, \code{\link{groupPlot}}, \code{\link{EICgeneral}},
extended with EICs of adduct species
}
\description{
get EICs for adducts, isotopes and neutral loss masses
}
