% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MultiFilterModule.R
\name{MultiFilterModule}
\alias{MultiFilterModule}
\title{MultiFilterModule}
\usage{
MultiFilterModule(input, output, session,
  values = reactiveValues(featureTables = featureTables, MainTable =
  MainTable), static = list(lab = "Filter"))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{static}{Import data from the shiny session}

\item{reactives}{Import data from the shiny session}
}
\description{
Module to apply filters to a featureTable (UI)
}
