% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MultiFilterModule.R
\name{MultiFilterModuleUI}
\alias{MultiFilterModuleUI}
\title{MultiFilterModuleUI}
\usage{
MultiFilterModuleUI(id)
}
\arguments{
\item{id}{}
}
\description{
Module to apply filters to a featureTable (UI)
}
