% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_FeatureTable_analysis.R
\name{mzMatch}
\alias{mzMatch}
\title{mzMatch}
\usage{
mzMatch(df, db, ppm = 5, mzdiff = 0.001)
}
\arguments{
\item{df}{data.frame that contains a mz column}

\item{db}{data.frame that contains the compound database and at least columns mz and id. Can also be a character vector of .csv file names (will be combined into a single list for the search, with duplicate lines removed)}

\item{ppm}{ppm mz tolerance}

\item{mzdiff}{maximum mz difference. NOTE: either ppm OR mzdiff condition has to be met}
}
\description{
Find m/z matches for each item in a Feature Table in a database.
}
