% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MzqueryModule.R
\name{MzqueryModuleUI}
\alias{MzqueryModuleUI}
\title{massquery}
\usage{
MzqueryModuleUI(id)
}
\arguments{
\item{id}{id to be used in ns()}
}
\description{
Chemcalc mz query module
UI module to get molecular formula predictions from chemcalc
}
