% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_graph_utilities.R
\name{network1}
\alias{network1}
\title{pairCompare}
\usage{
network1(spec1, spec2, mztol = 0.005, parentshift = 0,
  method = "cosine", minpeaks = 6, nonmatched = T)
}
\arguments{
\item{spec1}{spectrum 1}

\item{spec2}{spectrum 2}

\item{mztol}{allowed m/z difference between peaks to match matching}

\item{parentshift}{m/z difference between parentmasses (importantly, has to be Parent(spec2) - Parent(spec1) ) to find alternative matches; will only be calculated if < mztol}

\item{method}{method passed on to \code{\link{pairCompare}()}; "cosine" will use lsa::cosine(), "pearson" will use stats::cor()}

\item{minpeaks}{minimum number of peaks that have to be matched, otherwise returns 0}
}
\description{
Compare two vectors of equal length pairwise and calculate a similarity score (e.g. intensity values of matched MS peaks)
NOTE: Current limitations: (1) peaks from one spectrum can match multiple peaks from the other scan (workaround: quickMergeMS with ppm = 0 and mzdiff same as mzdiff here)
}
