% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_NetworkModule.R
\name{NetworkModule}
\alias{NetworkModule}
\title{NetworkModule}
\usage{
NetworkModule(input, output, session, values = reactiveValues(Networks =
  Networks), reactives = reactive({     list(active = T, highlights =
  integer(0)) }), static = list(noSelection = T), keys)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{reactives}{Import data from the shiny session}
}
\description{
server module for interactive mass spectrum view
}
