% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_NetworkModule.R
\name{NetworkModuleUI}
\alias{NetworkModuleUI}
\title{NetworkModuleUI}
\usage{
NetworkModuleUI(id)
}
\arguments{
\item{id}{id to be used in ns()}
}
\description{
UI module for interactive spectrum view
}
