% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics_modules.R
\name{NormalizeModule}
\alias{NormalizeModule}
\title{NormalizeModule}
\usage{
NormalizeModule(input, output, session, mx)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{mx}{}
}
\description{
Normalize a matrix
}
