% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container_OptionsContainer.R
\name{OptionsContainerUI}
\alias{OptionsContainerUI}
\title{OptionsContainerUI}
\usage{
OptionsContainerUI(id)
}
\arguments{
\item{id}{}
}
\description{
Module to apply filters to a featureTable (UI)
}
