% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_graph_utilities.R
\name{pairCompare}
\alias{pairCompare}
\title{pairCompare}
\usage{
pairCompare(v1, v2, NAasZero = T, method = c("cosine", "pearson"))
}
\arguments{
\item{v1}{vector 1}

\item{v2}{vector 2}

\item{NAasZero}{replace NA values with 0}

\item{method}{"cosine" will use lsa::cosine(), "pearson" will use stats::cor()}
}
\value{
the result of the method call, a numeric similarity score
}
\description{
Compare two vectors of equal length pairwise and calculate a similarity 
score (e.g. intensity values of matched MS peaks)
}
