% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_formatting_utilities.R
\name{parsePatterns}
\alias{parsePatterns}
\title{parsePatterns}
\usage{
parsePatterns(df)
}
\arguments{
\item{df}{a data.frame with at least the columns \code{name} and
\code{pattern}, there pattern has to be numeric values separated 
by whitespace.}
}
\description{
parse patterns from a data.frame into a list that can be used with 
MassTools::findPatterns.
}
