% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_specplot.R
\name{parseTitle}
\alias{parseTitle}
\title{parseTitle}
\usage{
parseTitle(scanInfo)
}
\arguments{
\item{scanInfo}{a data.frame as returned by \link{getScanInfo}()}
}
\value{
a character vector of length 1.
}
\description{
parse a scanInfo table into a title for a mass spectrum plot
}
