% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PcaViewModule.R
\name{PcaViewModule}
\alias{PcaViewModule}
\title{PcaViewModule}
\usage{
PcaViewModule(input, output, session, values = reactiveValues(featureTables =
  featureTables))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{reactives}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
server module to view PCA results
}
