% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_PeakPickModule.R
\name{PeakPickModule}
\alias{PeakPickModule}
\title{PeakPickModule}
\usage{
PeakPickModule(input, output, session, values = reactiveValues(MSData =
  MSData, featureTables = featureTables, MainTable = MainTable))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
server module for saving Tables
}
