% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_PeptideAnnotationModule.R
\name{PeptideAnnotationModule}
\alias{PeptideAnnotationModule}
\alias{PeptideAnnotationModuleUI}
\title{PeptideAnnotationModule}
\usage{
PeptideAnnotationModule(input, output, session, values)

PeptideAnnotationModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
If the server module for this Module returns something,
 it is described in \code{Details}.
}
\description{
This module allows setting fixed and variable peptide modifications to be 
applied in peptide-related calculations (fragment annotation and peptide mass calculations)
}
\details{
Does not return anything, but
\subsection{requires}{
\code{PeptideModificationsModule},
 \code{SelectMS2Module},
 \code{MainTableModule}
} 
\subsection{initiates}{
nothing
}
\subsection{reads}{
\code{values$GlobalOpts$Peptides.variableMods},
 \code{values$GlobalOpts$Peptides.fixedMods},
 \code{FTselection()}
}
\subsection{modifies}{
\code{values$MSData$MS2.selected}
}
}
\section{Functions}{
\itemize{
\item \code{PeptideAnnotationModule}: server logic for PeptideAnnotationModule

\item \code{PeptideAnnotationModuleUI}: UI elements for PeptideAnnotationModule
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
