% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_PeptideModificationsModule.R
\name{PeptideModificationsModule}
\alias{PeptideModificationsModule}
\alias{PeptideModificationsModuleUI}
\title{PeptideModificationsModule}
\usage{
PeptideModificationsModule(input, output, session, values)

PeptideModificationsModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
If the server module for this Module returns something,
 it is described in \code{Details}.
}
\description{
This module allows setting fixed and variable peptide modifications to be 
applied in peptide-related calculations (fragment annotation and peptide mass calculations)
}
\details{
Does not return anything, but
\subsection{initiates}{
\code{values$GlobalOpts$Peptides.variableMods}, \code{values$GlobalOpts$Peptides.fixedMods}
}
\subsection{modifies}{
\code{values$GlobalOpts$Peptides.variableMods}, \code{values$GlobalOpts$Peptides.fixedMods}
}
\subsection{reads}{
\code{values$GlobalOpts$Peptides.variableMods}, \code{values$GlobalOpts$Peptides.fixedMods}
}
}
\section{Functions}{
\itemize{
\item \code{PeptideModificationsModule}: server logic for PeptideModificationsModule

\item \code{PeptideModificationsModuleUI}: UI elements for PeptideModificationsModule
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
