% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_permutate.R
\name{permutate}
\alias{permutate}
\title{permutate}
\usage{
permutate(modifiers)
}
\arguments{
\item{modifiers}{data.frame with columns name, formula, min, max.}
}
\description{
Use a data.frame with molecular formulas and combine them
}
\details{
columns min and max should be the minimum and maximum times a formula is allowed in the combinations, can be positive, negative, or 0.
formula: molecular formulas can contain positive or negative integers, but no parantheses or brackets.
The mass in the resulting data.frame is monoisotopic.
}
