% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_PlotBrowserModule.R
\name{PlotBrowserModule}
\alias{PlotBrowserModule}
\title{PlotBrowserModule}
\usage{
PlotBrowserModule(input, output, session, reactives = reactive({    
  reactiveValues(PCAtable = df, active = T) }), values = NULL,
  static = list(patterns = list(axis = "PCA__", color = "", hover = "")))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{reactives}{Import data from the shiny session}

\item{values}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
server module to customize ggplot views
}
