% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_PlotBrowserModule.R
\name{PlotBrowserModuleUI}
\alias{PlotBrowserModuleUI}
\title{PlotBrowserModuleUI}
\usage{
PlotBrowserModuleUI(id)
}
\arguments{
\item{id}{id of the shiny module}
}
\description{
PlotBrowserModuleUI
}
