% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_peakPicking.R
\name{plotpeaks}
\alias{plotpeaks}
\title{plotpeaks}
\usage{
plotpeaks(EIC, peaklist)
}
\arguments{
\item{EIC}{a single EIC object (list item of multiEIC output)}

\item{peaklist}{a peaklist, such as list items returned by \code{find_peaks()} 
(or its rbind product)}
}
\value{
plots to the current plotting device
}
\description{
plot detected peaks for each file in an EIC object
}
