% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_featureReport.R
\name{plotTree}
\alias{plotTree}
\title{plotTree}
\usage{
plotTree(tree, resolution = 2000, filename = NULL)
}
\arguments{
\item{tree}{DiagrammeR::grViz output object}

\item{resolution}{resolution  of the image along its longest edge}
}
\description{
plot a grViz- generated tree in an R plotting device
}
