% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_PlotWindow.R
\name{PlotWindow}
\alias{PlotWindow}
\title{PlotWindow}
\usage{
PlotWindow(
  cx = 1,
  ylim = c(0, 1),
  xlim = c(0, 1),
  heading = "test",
  single = F,
  par = T,
  relto = NULL,
  ylab = "Intensity",
  xlab = "RT (min)",
  ysci = T,
  liwi = 1,
  textadj = 0.5,
  ylabshift = 0
)
}
\arguments{
\item{cx}{character expansion factor (font size)}

\item{ylim}{numeric(2) of y-axis range}

\item{xlim}{numeric(2) of x-axis range}

\item{heading}{heading of the plot}

\item{single}{if TRUE, this plot is expected to be the only 
plot in a composite plot (different margin settings)}

\item{par}{if FALSE, par margin settings are not set inside 
the function and should be set outside}

\item{relto}{show y axis values relative to relto if not NULL.}

\item{ylab}{y axis label}

\item{xlab}{x axis label}

\item{ysci}{if TRUE, y axis label numbers are shown in scientific format}

\item{liwi}{line width for axes}

\item{textadj}{passed on to mtext adj for orientation 
of plot description/title text line}

\item{ylabshift}{integer to adjust the position of the y axis label}
}
\value{
plots an empty coordinate system in the current plotting device
}
\description{
Initiate a plot and draw axes. Used by plotting functions in Metaboseek.
}
