% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_class_featureTable.R
\name{previousStep}
\alias{previousStep}
\alias{previousStep.MseekFT}
\title{previousStep}
\usage{
previousStep(x)

\method{previousStep}{MseekFT}(x)
}
\description{
extract the most recent \code{ProcessHistory} object from an object,
 representing the last recorded changes made to the object.
}
