% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_deprecated.R
\name{quickMergeMS}
\alias{quickMergeMS}
\title{quickMergeMS}
\usage{
quickMergeMS(speclist, ppm = 5, mzdiff = 0.0005, removeNoise = NULL, count = F)
}
\arguments{
\item{speclist}{list of spectra (matrix with mz and intensity values)}

\item{ppm}{min difference between peaks in ppm}

\item{mzdiff}{min difference between peaks in m/z}

\item{removeNoise}{if not Null, will remove peaks with intensities below this fraction of maximum intensity}

\item{count}{if true, will count number of peaks that were merged into each 
peak in the resulting spectrum (will return a 3-column matrix)}
}
\value{
a merged mass spectrum as a matrix with mz and intensity values
}
\description{
NOTE: This function is deprecated, use MassTools::mergeMS !
}
\details{
Merge MS spectra, best used with noise-free spectra. Will merge all peaks 
that are EITHER within ppm OR within mzdiff range (mzdiff typically used 
to allow larger ppm difference in low mz range)

Replacement for mergeMS(..., mergeOnly = T)
}
