% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_QuickPlotModule.R
\name{QuickPlotsModule}
\alias{QuickPlotsModule}
\alias{QuickPlotsModuleUI}
\title{QuickPlotsModule}
\usage{
QuickPlotsModule(input, output, session,
  values = reactiveValues(featureTables = featureTables))

QuickPlotsModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
If the server module for this Module returns something,
 it is described in \code{Details}.
}
\description{
Show summary data plots of selected features using ggplot plotting devices
}
\section{Functions}{
\itemize{
\item \code{QuickPlotsModule}: server logic for QuickPlotsModule

\item \code{QuickPlotsModuleUI}: UI elements for QuickPlotsModule
@export
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
