% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_RatioPlotModule.R
\name{RatioPlotModule}
\alias{RatioPlotModule}
\alias{RatioPlotModuleUI}
\title{RatioPlotModule}
\usage{
RatioPlotModule(input, output, session, values)

RatioPlotModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
This module returns nothing
}
\description{
This is a ratio plotter
}
\section{Functions}{
\itemize{
\item \code{RatioPlotModule}: Server logic

\item \code{RatioPlotModuleUI}: UI elements
}}

\examples{
\dontrun{
MseekExamplePreload()

tab1 <- FTNormalize(tab1)

ui <- MseekMinimalUI(
  fluidPage(
    fluidRow(
      RatioPlotModuleUI("examplewidget")),
    fluidRow(
      MainTableModuleUI("mtb"))
  ),
  diagnostics = T)

server <- function(input, output) {
  MseekMinimalServer(diagnostics = T, data = F, tables = T)
  
  MainTable <- callModule(MainTableModule, "mtb",
                          values = reactiveValues(featureTables = featureTables,
                                                 GlobalOpts = GlobalOpts),
                         static = list(perpage = 100,
                                        height = 300,
                                        readOnly = T,
                                        contextMenu = T,
                                        fixedColumnsLeft = 1,
                                        invertReadOnly = NULL,
                                        controls = T,
                                        format = list(col = NULL,
                                                      format = NULL)))
  
  ExampleWidget <- callModule(RatioPlotModule, "examplewidget",
                              values = values)
  
  
}

# Create Shiny app ----
shinyApp(ui, server)
}

VennDiagramModule

server module for Venn diagrams in Metaboseek

}
\seealso{
\code{\link{values}} for a description of the \code{values} object
}
