% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_formatting_utilities.R
\name{reconstructValues}
\alias{reconstructValues}
\title{reconstructValues}
\usage{
reconstructValues(values, savedValues)
}
\arguments{
\item{values}{an MseekTree (reactivevalues) object}

\item{savedValues}{a saved values object (created by applying
\code{reactiveValuesToListRec()} to an MseekTree object)}
}
\value{
nothing, but modifies the global values object
}
\description{
reconstructs the MseekTree (values object) from a saved values object (that 
contains all data to be loaded as lists rather than reactiveValues).
}
