% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_calcMZ.R
\name{reformatSF}
\alias{reformatSF}
\title{'reformatSF}
\usage{
reformatSF(s)
}
\arguments{
\item{s}{Molecula formula string}
}
\description{
Reformats a molecula formula string so that it has a count for each element, even if the count is 1.
}
