% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_RegroupMSDataModule.R
\name{RegroupMSDataModuleUI}
\alias{RegroupMSDataModuleUI}
\title{RegroupMSDataModuleUI}
\usage{
RegroupMSDataModuleUI(id)
}
\arguments{
\item{id}{id of the shiny module}
}
\description{
RegroupMSDataModuleUI
}
