% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_calcMZ.R
\name{remakeSF}
\alias{remakeSF}
\title{'remakeSF}
\usage{
remakeSF(x)
}
\arguments{
\item{s}{Molecula formula string}
}
\description{
Converts an element count vector back into a molecula formula string
}
