% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_plotting_genericData.R
\name{reverselog_trans}
\alias{reverselog_trans}
\title{reverselog_trans}
\usage{
reverselog_trans(base = exp(1))
}
\arguments{
\item{base}{base of log}
}
\description{
reverse log scale for ggplot, as described here: https://stackoverflow.com/questions/11053899/how-to-get-a-reversed-log10-scale-in-ggplot2
}
