% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_retcorAdjust.R
\name{rtadjust}
\alias{rtadjust}
\title{rtadjust}
\usage{
rtadjust(rtexportObj, rts)
}
\arguments{
\item{rtexportObj}{\code{RTcorr} class object as generated by \code{\link{rtexport}()}}

\item{rts}{data.frame with rt values after RT correction}
}
\value{
a list of data.frames with corrected RT values
}
\description{
recalculate original rt values from rt values as reported in 
featuretables after rt correction for each file
}
\details{
each element in the returned list corresponds to \code{rts}, with RT values 
reset to pre-RT correction values
}
