% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_RtCorrViewModule.R
\name{RtCorrViewModuleUI}
\alias{RtCorrViewModuleUI}
\title{RtCorrViewModuleUI}
\usage{
RtCorrViewModuleUI(id)
}
\arguments{
\item{id}{}
}
\description{
Module to apply filters to a featureTable (UI)
}
