% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_retcorAdjust.R
\name{RTplot}
\alias{RTplot}
\title{RTplot}
\usage{
RTplot(rtt, colscheme = "Mseek.colors", liwi = 2, cx = 1)
}
\arguments{
\item{rtt}{\code{RTcorr} object as generated by \code{\link{rtexport}()}}

\item{colscheme}{character(1): name of function for color palette}

\item{liwi}{line width}

\item{cx}{font size}
}
\value{
generates a new plot in current plotting device
}
\description{
Plot retention time deviation for each file generated by an 
\code{\link{rtexport}()} generated RTcorr object
}
