% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_runMosaic.R
\name{runMosaic}
\alias{runMosaic}
\title{Run Mosaic}
\usage{
runMosaic(...)
}
\arguments{
\item{...}{Additional arguments passed to shiny::runApp, e.g. to set up the port used or to make Mosaic reachable from other computers (with host = getOption("shiny.host", "0.0.0.0"))}
}
\description{
Run Mosaic in your webbrowser.
}
