% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_SIRIUS.R
\name{runSirius}
\alias{runSirius}
\title{runSirius}
\usage{
runSirius(
  sirpath = file.path(.MseekOptions$siriusFolder, "sirius"),
  ms2,
  parentmz,
  outfolder = getwd(),
  ms1 = NULL,
  ion = "[M+?]+",
  comments = "",
  rt = "",
  charge = 1,
  fingerid = T,
  scanindices = "",
  moreOpts = "",
  config = list(IsotopeSettings.filter = TRUE, FormulaSearchDB = c("ALL_BUT_INSILICO",
    "ALL", "BIO", "METACYC", "CHEBI", "COCONUT", "ECOCYCMINE", "GNPS", "HMDB", "HSDB",
    "KEGG", "KEGGMINE", "KNAPSACK", "MACONDA", "MESH", "NORMAN", "UNDP", "PLANTCYC",
    "PUBCHEM", "PUBMED", "YMDB", "YMDBMINE", "ZINCBIO"), Timeout.secondsPerTree = 0,
    FormulaSettings.enforced = "HCNOP[5]S", Timeout.secondsPerInstance = 0,
    AdductSettings.detectable =
    "[[M+K]+,[M+H3N+H]+,[M+Na]+,[M-H4O2+H]+,[M-H2O+H]+,[M+H]+]",
    UseHeuristic.mzToUseHeuristicOnly = 650,      AlgorithmProfile = "orbitrap",
    IsotopeMs2Settings = "IGNORE", MS2MassDeviation.allowedMassDeviation = "5.0ppm",
    NumberOfCandidatesPerIon = 1, UseHeuristic.mzToUseHeuristic = 300,
    FormulaSettings.detectable = ",", NumberOfCandidates = 10, StructureSearchDB =
    c("ALL_BUT_INSILICO", "ALL", "BIO", "METACYC", "CHEBI", "COCONUT", "ECOCYCMINE",
    "GNPS", "HMDB", "HSDB", "KEGG", "KEGGMINE", "KNAPSACK", "MACONDA", "MESH", "NORMAN",
    "UNDP", "PLANTCYC", "PUBCHEM", "PUBMED", "YMDB", "YMDBMINE", "ZINCBIO"),     
    AdductSettings.fallback = "[[M+K]+,[M+Na]+,[M+H]+]", RecomputeResults = TRUE),
  force = T
)
}
\arguments{
\item{sirpath}{path to SIRIUS executable}

\item{ms2}{a list (or list of lists) of ms2 spectra 
(matrices with columns mz and intensity), see \code{Details}}

\item{parentmz}{numeric: parent ion m/z value}

\item{outfolder}{output folder for the SIRIUS results}

\item{ms1}{if not NULL, a list of MS1 spectra}

\item{ion}{ion type (e.g. \eqn{[M+H]+})}

\item{comments}{character: comments}

\item{rt}{numeric: retention time in seconds}

\item{charge}{charge (positive or negative integer)}

\item{fingerid}{if TRUE, will try to predict structures with CSI:FingerID}

\item{scanindices}{which scans were averaged into this ms2 spectrum}

\item{moreOpts}{character with additional options to be passed to SIRIUS}

\item{config}{list containing sirius configuration}

\item{force}{force calculation, even if same results should exist according to indexfile}
}
\value{
Will cause SIRIUS to run a job and save its results in
 a timestamped subfolder of  \code{outfolder}, and will register the job in
 the \code{index.csv} file in \code{outfolder}.
}
\description{
run SIRIUS externally for a list of ms2 spectra
}
\details{
\describe{
   \item{ms2}{can be a list of matrices, or a list of list of matrices. 
   In the latter case, the (optional) names of the nested list items are 
   expected to be in the format "collisionXX", where XX denotes an integer 
   collision energy.}
}
}
