% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMosaic.R
\name{runXcms}
\alias{runXcms}
\title{Run XCMS interface}
\usage{
runXcms(...)
}
\arguments{
\item{...}{Additional arguments passed to shiny::runApp.}
}
\description{
Run the XCMS runner in your webbrowser. Currently only intended for use on a local computer (file system access required).
the XCMS runner will also be integrated into MOSAiC eventually.
}
