% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_formatting_utilities.R
\name{saveMseekSession}
\alias{saveMseekSession}
\title{saveMseekSession}
\usage{
saveMseekSession(values, path = NULL, MSData = T)
}
\arguments{
\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{path}{file path to save to. if NULL, the MseekSession object is returned}

\item{MSData}{if TRUE, MSData will be included in saved file. If false, only 
MS data file paths will be saved}
}
\value{
an MseekSession object
}
\description{
save the current Metaboseek session
}
