% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SaveNetworkModule.R
\name{SaveNetworkModule}
\alias{SaveNetworkModule}
\title{SaveNetworkModule}
\usage{
SaveNetworkModule(input, output, session, reactives = reactive({    
  list(graphname = NULL, filename = "table.graphml") }),
  values = reactiveValues(Networks = Networks), static = list(tooltip =
  "Save Network as a graphml file", label = "Save Network", format =
  c("graphml")))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{reactives}{Import data from the shiny session}

\item{values}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
server module for saving Tables
}
