% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SaveNetworkModule.R
\name{SaveNetworkModuleUI}
\alias{SaveNetworkModuleUI}
\title{SaveNetworkModuleUI}
\usage{
SaveNetworkModuleUI(id)
}
\arguments{
\item{id}{id of the shiny module}
}
\description{
SaveNetworkModuleUI
}
