% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_Scans.R
\name{saveScanlist}
\alias{saveScanlist}
\title{saveScanlist}
\usage{
saveScanlist(scanlist)
}
\arguments{
\item{scanlist}{a scanlist as generated by makeScanlist or Parentsearch()}
}
\description{
makes a string representation of a MS2 parent list, including only the basenames and scan # (not acquisition #) of MS2 scans in that list.
}
