% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_xcmsRunner.R
\name{savetable}
\alias{savetable}
\title{savetable}
\usage{
savetable(xset, importResultsFrom = NULL, filename = "tableoutxx",
  bparams = SnowParam(workers = 1), intensities = list(ppm = 5, rtw =
  5, rtrange = T), rawdata = NULL, saveR = T, postProc = NULL)
}
\arguments{
\item{xset}{an xcms::XCMSnExp, xcms::xcmsSet or xcms::xsAnnotate object from 
which to extract the peaklist}

\item{importResultsFrom}{if an MseekFT object is supplied here, will import Mseek intensities
and other post-processing results if appropriate (same feature definitions,
 same files, same intensities)}

\item{filename}{filename to use when writing the csv file (including file 
extension). "_filled" will be added for filled data.}

\item{bparams}{BBPARAMs (BiocParallel::BiocParallelParam object with 
parameters for parallel computing used in fill peaks)}

\item{intensities}{list of parameters for Mseek EIC intensity extraction 
(for each feature in the peaklist). ppm: mz window, rtw: rt window 
(+/- seconds), rtrange: if TRUE, rtw extends from rtmin and rtmax, 
if FALSE rtw centers around value in column rt. Not performed if 
intensities = NULL.}

\item{rawdata}{named list of xcmsRaw objects (as returned by
Mseek::loadRawM), required for Mseek EIC intensity extraction which 
is not performed if NULL.}

\item{saveR}{save the xset object as .Rds file}

\item{postProc}{list of SOME arguments passed to \code{\link{analyzeTable}()}
for post processing, or NULL to skip processing. 
TODO: transisiton to do.call solution for postProc and document}
}
\value{
Returns a \code{writeStatus} object, writes to \code{status.csv} 
 file as well as other .csv files (with processed or unprocessed peakTables) 
 in current working directory.
}
\description{
Save a peakTable from an xcms::XCMSnExp, xcms::xcmsSet or xcms::xsAnnotate
 object, and also keep track of analysis status via internal calls to 
 \code{\link{writeStatus}()}.
}
