% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SaveTableModule.R
\name{SaveTableModule}
\alias{SaveTableModule}
\title{SaveTableModule}
\usage{
SaveTableModule(input, output, session,
  values = reactiveValues(projectData = projectData, featureTables =
  NULL), reactives = reactive({     list(df = NULL, filename =
  "table.csv") }), static = list(tooltip = "Save", label = "Save", format
  = c("tsv", "csv"), allowformats = NULL))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{reactives}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
server module for saving Tables
}
