% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SelectActiveTableModule.R
\name{SelectActiveTableModule}
\alias{SelectActiveTableModule}
\title{SelectActiveTableModule}
\usage{
SelectActiveTableModule(input, output, session,
  values = reactiveValues(featureTables = featureTables, MainTable =
  MainTable))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{reactives}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
server module for loading a Project Folder
}
