% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SelectActiveTableModule.R
\name{SelectActiveTableModuleUI}
\alias{SelectActiveTableModuleUI}
\title{SelectActiveTableModuleUI}
\usage{
SelectActiveTableModuleUI(id)
}
\arguments{
\item{id}{id of the Module}
}
\description{
SelectActiveTableModuleUI
}
