% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SelectMS2Module.R
\name{SelectMS2Module}
\alias{SelectMS2Module}
\alias{SelectMS2ModuleUI}
\title{SelectMS2Module}
\usage{
SelectMS2Module(input, output, session, values)

SelectMS2ModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with \code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with \code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with \code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} (must match \code{id} in \code{\link[shiny]{callModule}} for the server Module)}
}
\value{
If the server module for this Module returns something, it is described in \code{Details}.
}
\description{
Module for selection of MS2 spectra. Will search for MS2 spectra matching the
 current selection in the Feature table within specified ppm and RT boundaries,
 and will make the selection available app-wide.
 
 @details Does not return anything, but will modify values:
 \itemize{
 \item{values$MSData$MS2.selectTable:} data.frame listing metadata for all MS2 spectra matching criteria
 \item{values$MSData$MS2.selected:} list with three items: 
 \itemize{
 \item{spectra:} list of selected MS2 spectra as matrices
 \item{mergedSpectrum:} result of merging all selected spectra as a matrix
 \item{specInfo:} data.frame listing metadata for all selected MS2 spectra
}}
}
\section{Functions}{
\itemize{
\item \code{SelectMS2Module}: server logic for SelectMS2Module

\item \code{SelectMS2ModuleUI}: UI elements for SelectMS2Module
}}

