% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SelectMSGroupingModule.R
\name{SelectMSGroupingModule}
\alias{SelectMSGroupingModule}
\title{SelectMSGroupingModule}
\usage{
SelectMSGroupingModule(input, output, session,
  values = reactiveValues(MSData = MSData), static = list(editOnly =
  F))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{static}{Import data from the shiny session}

\item{reactives}{Import data from the shiny session}
}
\description{
server module for loading a Project Folder
}
