% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SelectMSGroupingModule.R
\name{SelectMSGroupingModuleUI}
\alias{SelectMSGroupingModuleUI}
\title{SelectMSGroupingModuleUI}
\usage{
SelectMSGroupingModuleUI(id)
}
\arguments{
\item{id}{id of the Module}
}
\description{
SelectMSGroupingModuleUI
}
