% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SelectProjectFolderModule.R
\name{SelectProjectFolderModule}
\alias{SelectProjectFolderModule}
\alias{SelectProjectFolderModuleUI}
\title{SelectProjectFolderModule}
\usage{
SelectProjectFolderModule(input, output, session,
  values = reactiveValues(projectData = NULL, featureTables = NULL,
  MSData = NULL, GlobalOpts = NULL))

SelectProjectFolderModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
Returns nothing
}
\description{
Module for loading a Project Folder
}
\section{Functions}{
\itemize{
\item \code{SelectProjectFolderModule}: Server logic

\item \code{SelectProjectFolderModuleUI}: UI elements
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
