% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SelectProjectFolderModule.R
\name{SelectProjectFolderModuleUI}
\alias{SelectProjectFolderModuleUI}
\title{SelectProjectFolderModuleUI}
\usage{
SelectProjectFolderModuleUI(id)
}
\arguments{
\item{id}{}
}
\description{
server module for loading Tables
}
