% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_ShowSiriusModule.R
\name{ShowSiriusModule}
\alias{ShowSiriusModule}
\title{ShowSiriusModule}
\usage{
ShowSiriusModule(input, output, session,
  values = reactiveValues(SiriusModule = SiriusModule, GlobalOpts =
  GlobalOpts), reactives = reactive({     list(splash = NULL, mz = NULL)
  }))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{reactives}{Import data from the shiny session}
}
\description{
server module for interactive mass spectrum view
}
