% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_ShowSiriusModule.R
\name{ShowSiriusModuleUI}
\alias{ShowSiriusModuleUI}
\title{ShowSiriusModuleUI}
\usage{
ShowSiriusModuleUI(id)
}
\arguments{
\item{id}{id to be used in ns()}
}
\description{
UI module for interactive SIRIUS interface
}
