% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_simpleTableModule.R
\name{simpleTableModule}
\alias{simpleTableModule}
\alias{simpleTableModuleUI}
\title{simpleTableModule}
\usage{
simpleTableModule(input, output, session, df = reactive({     NULL }),
  static = list(readOnly = T, contextMenu = F, height = "auto"))

simpleTableModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{df}{reactive containing the data.frame to display}

\item{static}{a list of arguments that are not in a reactive context 
and hence do not change while the app is running.}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()}
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
Returns its internalValues, most importantly elements \code{df} 
and \code{liveView}
}
\description{
A simple Table module without sorting or pagination
}
\section{Functions}{
\itemize{
\item \code{simpleTableModule}: Server logic

\item \code{simpleTableModuleUI}: UI elements
}}

