% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_simpleTableModule.R
\name{simpleTableModuleUI}
\alias{simpleTableModuleUI}
\title{simpleTableModuleUI}
\usage{
simpleTableModuleUI(id)
}
\arguments{
\item{id}{id of the shiny module}
}
\description{
simpleTableModuleUI
}
