% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_graph_utilities.R
\name{simplifyGraph}
\alias{simplifyGraph}
\title{simplifyGraph}
\usage{
simplifyGraph(nodes, edges, mergebyedges)
}
\arguments{
\item{nodes}{a node table}

\item{edges}{an edge table with "from" and "to" as the first two columns. "From" and "to" have to refer to rownumbers of nodes.}

\item{mergebyedges}{indices by of those edges in edges that link two nodes that should be joined}
}
\description{
Assign a color from a range of colors to all values in a numeric vector (datarange). NOTE: Edges are expected to not have a direction.
}
