% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SimplifyNetworkModule.R
\name{SimplifyNetworkModule}
\alias{SimplifyNetworkModule}
\title{SimplifyNetworkModule}
\usage{
SimplifyNetworkModule(input, output, session,
  values = reactiveValues(Networks = Networks), reactives = reactive({ 
     list(activeNetwork = NULL) }))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{tag}{id to be used in ns()}

\item{set}{Import data from the shiny session}
}
\description{
server module for loading Tables
}
