% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SiriusModule.R
\name{SiriusModuleUI}
\alias{SiriusModuleUI}
\title{SiriusModuleUI}
\usage{
SiriusModuleUI(id)
}
\arguments{
\item{id}{id to be used in ns()}
}
\description{
UI module for interactive SIRIUS interface
}
