% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SiriusOptionsModule.R
\name{SiriusOptionsModule}
\alias{SiriusOptionsModule}
\title{SiriusOptionsModule}
\usage{
SiriusOptionsModule(input, output, session,
  values = reactiveValues(GlobalOpts = GlobalOpts))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{reactives}{Import data from the shiny session}
}
\description{
server module for interactive mass spectrum view
}
