% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SiriusOptionsModule.R
\name{SiriusOptionsModuleUI}
\alias{SiriusOptionsModuleUI}
\title{SiriusOptionsModuleUI}
\usage{
SiriusOptionsModuleUI(id)
}
\arguments{
\item{id}{id to be used in ns()}
}
\description{
UI module for interactive SIRIUS interface
}
