% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_plotting_modules.R
\name{SpecmoduleUI}
\alias{SpecmoduleUI}
\title{SpecmoduleUI}
\usage{
SpecmoduleUI(id)
}
\arguments{
\item{id}{id to be used in ns()}
}
\description{
UI module for interactive spectrum view
}
