% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_deprecated.R
\name{specplot}
\alias{specplot}
\title{specplot}
\usage{
specplot(x, y, norm = max(y)/100, cx = 1.5, k = 10,
  fileName = "title", yrange = c(0, 100), xrange = range(x),
  maxi = max(y), labels = NULL, mar = c(4, 6, 6, 2),
  ylab = "Relative Intensity (\%)", ylabshift = 0)
}
\arguments{
\item{x}{mz coordinates}

\item{y}{intensity coordinates}

\item{norm}{normalize by}

\item{cx}{font size}

\item{k}{top k intensity peaks will be labeled}

\item{fileName}{plot title}

\item{yrange}{y axis range}

\item{xrange}{x axis range}

\item{maxi}{max intensity to be plotted on side}

\item{labels}{data.frame containing at least x and y coordinates,
plus optional columns: label and color}

\item{mar}{margins passed to par()}

\item{ylab}{y axis label}

\item{ylabshift}{shift horizontal position of y axis label}
}
\value{
plots a spectrum view in the current plotting device
}
\description{
Plot an MS spectrum
}
\details{
NOTE: This function is deprecated. Use \code{\link{specplot2}} instead
}
