% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_specplot.R
\name{specplot2}
\alias{specplot2}
\title{specplot2}
\usage{
specplot2(x = sc[, 1], y = sc[, 2], spectrum, norm = NULL, cx = 1,
  k = 10, fileName = "", yrange = NULL, xrange = NULL,
  maxi = NULL, labels = NULL, highlights = NULL, mar = c(3, 4, 6,
  1), ylab = "Relative Intensity (\%)", ylabshift = 2)
}
\arguments{
\item{x}{mz coordinates}

\item{y}{intensity coordinates}

\item{norm}{normalize by}

\item{cx}{font size}

\item{k}{top k intensity peaks will be labeled}

\item{fileName}{plot title}

\item{yrange}{y axis range}

\item{xrange}{x axis range}

\item{maxi}{max intensity to be plotted on side.}

\item{labels}{data.frame containing at least x and y coordinates,
plus optional columns: label and color}

\item{highlights}{matrix or data.frame with at the first two columns containing x and y coordinates of peaks to highlight.
If \code{highlights} has an additional column called \code{color},
 that column is expected to contain color values for the highlights.
If \code{highlights} has an additional column called \code{type},
 that column is expected to contain point \code{type} values for the highlights,
 as detailed in \link[graphics]{plot.default}.}

\item{mar}{margins passed to par()}

\item{ylab}{y axis label}

\item{ylabshift}{shift horizontal position of y axis label}
}
\description{
Plot an MS spectrum
}
