% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_FeatureTable_analysis.R
\name{sttest}
\alias{sttest}
\title{sttest}
\usage{
sttest(out = NA, ...)
}
\arguments{
\item{out}{what to return if there is an error in the call to t.test, defaults to NA}

\item{...}{arguments passed on to stats::t.test}
}
\description{
helper function for mttest, calculating the p.value between two numeric vectors
}
