% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_EICs.R
\name{subsetEICs}
\alias{subsetEICs}
\title{subsetEICs}
\usage{
subsetEICs(EIClist, group)
}
\arguments{
\item{EIClist}{matrix or list of EIC objects}

\item{group}{file grouping information (named list)}
}
\value{
a list of extracted ion chromatograms that can be read 
by \code{\link{EICplot}}, \code{\link{groupPlot}}, \code{\link{EICgeneral}}
}
\description{
helper function to subset output from multiEICplus
}
