% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_SwitchButton.R
\name{SwitchButton}
\alias{SwitchButton}
\alias{SwitchButtonUI}
\title{SwitchButton}
\usage{
SwitchButton(
  input,
  output,
  session,
  states = list(on = list(label = "", title = "it's on", icon = icon("react"), style =
    "background-color: #1BBF1E;"), off = list(label = "", icon = icon("react"), title =
    "it's off")),
  type = "actionButton"
)

SwitchButtonUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{states}{named list of arguments to be passed to type function. 
NOTE: Do not provide inputId.}

\item{type}{type of UI element, typically "actionButton" or "actionLink"}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()}
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\description{
server module for UI element that changes appearance when interacted with 
(e.g. clicked on) and returns which appearance (state) it is currently using
}
\section{Functions}{
\itemize{
\item \code{SwitchButton}: server logic

\item \code{SwitchButtonUI}: UI elements
}}

