% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_TableAnalysisModule.R
\name{TableAnalysisModule}
\alias{TableAnalysisModule}
\title{TableAnalysisModule}
\usage{
TableAnalysisModule(input, output, session,
  values = reactiveValues(featureTables = featureTables, GlobalOpts =
  values$GlobalOpts, MSData = MSData), reactives = reactive({    
  list(fileGrouping = NULL) }), static = list())
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{reactives}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
server module for interactive mass spectrum view
}
