% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableAnalysisModule.R
\name{TableAnalysisModuleUI}
\alias{TableAnalysisModuleUI}
\title{TableAnalysisModuleUI}
\usage{
TableAnalysisModuleUI(id)
}
\arguments{
\item{id}{id of the shiny module}
}
\description{
TableAnalysisModuleUI
}
