% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_TableModule.R
\name{TableModule}
\alias{TableModule}
\title{TableModule}
\usage{
TableModule(input, output, session, tag, set = list(df = NULL, update = 1,
  layout = list(perpage = 100, height = 300, readOnly = T, contextMenu = T,
  fixedColumnsLeft = 1, invertReadOnly = NULL, format = list(col = NULL, format
  = NULL))))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{tag}{id to be used in ns()}

\item{set}{Import data from the shiny session}
}
\description{
server module for interactive mass spectrum view
}
