% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableModule2.R
\name{TableModule2}
\alias{TableModule2}
\title{TableModule}
\usage{
TableModule2(input, output, session, reactives = reactive({     list() }),
  values = reactiveValues(df = NULL, rowFilters = NULL, colFilters = NULL),
  static = list(perpage = 100, height = 300, readOnly = T, contextMenu = T,
  fixedColumnsLeft = 1, invertReadOnly = NULL, format = list(col = NULL, format
  = NULL)))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{set}{Import data from the shiny session}
}
\description{
server module for interactive mass spectrum view
}
