% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableModule2-bk.R, R/TableModule2.R
\name{TableModule2UI}
\alias{TableModule2UI}
\alias{TableModule2UI}
\title{TableModuleUI}
\usage{
TableModule2UI(id)

TableModule2UI(id)
}
\arguments{
\item{id}{id of the shiny module}

\item{id}{id of the shiny module}
}
\description{
TableModuleUI

TableModuleUI
}
