% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_deprecated.R
\name{TableUpdateChunk}
\alias{TableUpdateChunk}
\title{TableUpdateChunk}
\usage{
TableUpdateChunk()
}
\description{
TableUpdateChunk
}
\details{
\subsection{requires}{

values$featureTables$Maintable - pointing to the MainTableObject
values$featureTables
}
}
\section{NOTE}{

This function is deprecated. Use \code{\link{updateFT}} instead!

Code chunk to update the active Feature Table before triggering a
 rerendering of the rhandsontable object.
 
Chunks are evaluated in their parent environment, and therefore require
 all objects they work on to be present under their correct names.
}

