% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_class_featureTable.R
\name{updateDF}
\alias{updateDF}
\title{updateDF}
\usage{
updateDF(a, b)
}
\arguments{
\item{a}{source data.frame}

\item{b}{target data.frame}
}
\description{
Update dataframe b with data from dataframe a; must have equal number of rows
Existing data in b will be overridden if columns by the same name exist in a, and retained otherwise.
Additional columns in a will be transferred to b. Returns an updated data.frame b.
}
