% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_feature_table.R
\name{updateFeatureTable}
\alias{updateFeatureTable}
\title{updateDF}
\usage{
updateFeatureTable(FT, df)
}
\arguments{
\item{FT}{MosaicFT object}

\item{df}{source data.frame}
}
\description{
update an existing MosaicFT with new data frame columns (leaving intensitiy tables and constants in effect)
}
