% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset_featureTables.R
\name{updateFT}
\alias{updateFT}
\title{updateFT}
\usage{
updateFT(values)
}
\arguments{
\item{values}{an MseekTree (reactivevalues) object}
}
\value{
nothing, but modifies \code{FeatureTable(values)$df}
}
\description{
Update currently active Feature Table with any changes that may have
 been applied via the GUI
}
