% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_class_MSdata.R
\name{updateRawLayout}
\alias{updateRawLayout}
\title{updateRawLayout}
\usage{
updateRawLayout(MSD, new.stem = NULL)
}
\arguments{
\item{MSD}{a rawLayout object with a defined stem}

\item{stem}{if the file paths in rawgrouptable are not full (e.g. subdirectories of the working directory), this should be the path of the working directory.}
}
\description{
Updates the file path information in a rawLayout object
}
