% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_updaterModule.R
\name{updaterModuleUI}
\alias{updaterModuleUI}
\title{updaterModuleUI}
\usage{
updaterModuleUI(id)
}
\arguments{
\item{id}{id to be used in ns()}
}
\description{
module to update Mseek by the click of a button
}
