% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_UploadTableModule.R
\name{UploadTableModule}
\alias{UploadTableModule}
\title{UploadTableModule}
\usage{
UploadTableModule(input, output, session,
  values = reactiveValues(featureTables = NULL), static = list(title =
  NULL, filetypes = NULL, format = list(header = T, sep = "\\t", quote =
  "\\"", stringsAsFactors = F)))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{static}{Import data from the shiny session}
}
\description{
server module for loading Tables
}
