% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_UploadTableModule.R
\name{UploadTableModuleUI}
\alias{UploadTableModuleUI}
\title{UploadTableModuleUI}
\usage{
UploadTableModuleUI(id)
}
\arguments{
\item{id}{id of the shiny module}
}
\description{
UploadTableModuleUI
}
