% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation_templates.R
\name{values}
\alias{values}
\title{values}
\description{
\code{values} is a reactivevalues object with a defined structure
}
\section{Object structure}{

The \code{values} object is a \code{\link[shiny]{reactiveValues}} object,
containing three other \code{reactiveValues} objects, each 
\subsection{values$featureTables}{
contains objects related to feature tables
}
\subsection{values$MSData}{
contains object related to MS data
\describe{
 \item{values$MSData$MS2.selectTable}{data.frame listing metadata for all MS2 spectra matching criteria}
 \item{values$MSData$MS2.selected}{list with three items:
 \itemize{
 \item{spectra:} list of selected MS2 spectra as matrices
 \item{mergedSpectrum:} result of merging all selected spectra as a matrix
 \item{specInfo:} data.frame listing metadata for all selected MS2 spectra
}}
}

}
\subsection{values$GlobalOpts}{
GlobalOpts loads the current MseekOptions into the current session as
\code{reactiveValues}, and is also used by many modules to make values
 accessible app-wide.
}
}

