% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_VennDiagramModule.R
\name{VennDiagramModule}
\alias{VennDiagramModule}
\title{VennDiagramModule}
\usage{
VennDiagramModule(input, output, session,
  values = reactiveValues(featureTables = featureTables, GlobalOpts =
  GlobalOpts), reactives = reactive({     list(active = T, highlights =
  integer(0)) }), static = list(noSelection = T), keys)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{reactives}{Import data from the shiny session}
}
\description{
server module for Venn diagrams in METABOseek
}
