% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_VennDiagramModule.R
\name{VennDiagramModuleUI}
\alias{VennDiagramModuleUI}
\title{VennDiagramModuleUI}
\usage{
VennDiagramModuleUI(id)
}
\arguments{
\item{id}{id to be used in ns()}
}
\description{
UI module for Venn diagrams in METABOseek
}
