% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_VennWidgetModule.R
\name{VennWidgetModule}
\alias{VennWidgetModule}
\title{VennWidgetModule}
\usage{
VennWidgetModule(input, output, session, reactives = reactive({    
  list(x = list(A = 1:105, B = 101:115), ext.percent = c(0.01, 0.01, 0.01),
  filename = NULL) }), static = list())
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{reactives}{Import data from the shiny session}
}
\description{
server module for Venn diagrams in shiny
}
