% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_VennWidgetModule.R
\name{VennWidgetModuleUI}
\alias{VennWidgetModuleUI}
\title{VennWidgetModuleUI}
\usage{
VennWidgetModuleUI(id)
}
\arguments{
\item{id}{id to be used in ns()}
}
\description{
UI module for Venn diagrams in shiny
}
