% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_WelcomePageModule.R
\name{WelcomePageModuleUI}
\alias{WelcomePageModuleUI}
\title{WelcomePageModuleUI}
\usage{
WelcomePageModuleUI(id)
}
\arguments{
\item{id}{id of the shiny module}
}
\description{
WelcomePageModuleUI
}
