% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_SIRIUS.R
\name{writeMS}
\alias{writeMS}
\title{writeMS}
\usage{
writeMS(filename, ms2, parentmz, comments = "", rt = "", ion = NULL,
  charge = 1, splashtag = NULL, scanindices = NULL)
}
\arguments{
\item{filename}{path of file to write to}

\item{ms2}{a matrix with columns mz and intensity}

\item{parentmz}{numeric: parent ion m/z value}

\item{comments}{character: comments}

\item{rt}{numeric: retention time in seconds}

\item{ion}{ion type (e.g. \eqn{[M+H]+})}

\item{charge}{charge (positive or negative integer)}

\item{splashtag}{splash tag of the ms2 spectrum}

\item{scanindices}{which scans were averaged into this ms2 spectrum}
}
\description{
write an .ms file containing a spectrum to be exported to SIRIUS.
will append additional spectrum to file if filename is that of an existing file.
}
