% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_xcmsModule.R
\name{xcmsModule}
\alias{xcmsModule}
\alias{xcmsModuleUI}
\title{xcmsModule}
\usage{
xcmsModule(input, output, session, values = reactiveValues(),
  static = list(servermode = F, rootpath = .MseekOptions$filePaths,
  activateXCMS = T, filePattern = .MseekOptions$filePattern))

xcmsModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{static}{a list of arguments that are not in a reactive context 
and hence do not change while the app is running.}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
Returns nothing
}
\description{
server module for accessing the xcms data analysis workflow
}
\section{Functions}{
\itemize{
\item \code{xcmsModule}: Server logic

\item \code{xcmsModuleUI}: UI elements
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
