% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset_featureTables.R
\name{activeFT}
\alias{activeFT}
\title{activeFT}
\usage{
activeFT(x)
}
\arguments{
\item{x}{an MseekTree (reactivevalues) object}
}
\value{
returns the ID (character(1)) of the currently active Feature Table
}
\description{
activeFT
}
