% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R, R/methods-MseekFT.R
\docType{methods}
\name{analyzeFT,data.frame,listOrNULL,FTAnalysisParam-method}
\alias{analyzeFT,data.frame,listOrNULL,FTAnalysisParam-method}
\alias{analyzeFT}
\alias{removeNAs,MseekFT-method}
\alias{removeNAs}
\title{analyzeFT}
\usage{
\S4method{analyzeFT}{data.frame,listOrNULL,FTAnalysisParam}(object, MSData,
  param)

\S4method{removeNAs}{MseekFT}(object, replacement = 0)
}
\arguments{
\item{object}{an MseekFT or data.frame object.}

\item{MSData}{list of xcmsRaw objects}

\item{param}{a \code{\link{FTAnalysisParam}} object}
}
\value{
an object of the same class as \code{object}, with analyses performed as
defined by \code{param}
}
\description{
analyze Tables using \code{\link{analyzeTable}()},
 recording processing history if \code{object} is of class \code{MseekFT}.
 Will use intensity columns and grouping information from the MseekFT object
 if available
}
