% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_FeatureTable_analysis.R
\name{analyzeTable}
\alias{analyzeTable}
\title{analyzeTable}
\usage{
analyzeTable(df = tab1$df, intensities = tab1$intensities,
  groups = tab1$anagroupnames, analyze = c("Basic analysis",
  "clara_cluster", "t-test", "Peak shapes", "PCA features", "PCA samples"),
  normalize = T, useNormalized = T, MSData = NULL, ppm = 5,
  controlGroup = NULL, numClusters = 2)
}
\arguments{
\item{df}{a data.frame with numeric (intensity) values, rt and mz values}

\item{intensities}{the intensity column names, before normalization (without __norm suffix), will be automatically renamed if useNormalized.}

\item{groups}{named list of non-normalized intensity columns listed by group (as supplied by $anagroupnames of MosaicFT objects), will be automatically renamed if useNormalized.}

\item{analyze}{character vector to select the analyses to be run: "Basic analysis", "clara_cluster", "t-test", "Peak shapes"}

\item{normalize}{normalze intensity columns}

\item{MSData}{a (named) list of xcmsRaw objects, e.g. generated by Mosaic::loadRawM, OR an MSnbase OnDiskMSnExp object required for peak shape analysis}

\item{ppm}{ppm range for peak shape analysis}

\item{controlGroup}{control group for foldChange (part of Basic analysis) analysis (optional)}

\item{numClusters}{number of clusters for clara_clusters analysis}

\item{useNormalize}{use normalized values for analyses; will trigger normalize if there is no normalized data available for all selected intensity columns}
}
\description{
Run a series of analyses on a feature table data.frame
}
