% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_msdata.R
\name{bestgauss}
\alias{bestgauss}
\title{bestgauss}
\usage{
bestgauss(rawdata = rawcoll, ...)
}
\arguments{
\item{rawdata}{a (named) list of xcmsRaw objects, e.g. generated by Mosaic::loadRaw}

\item{...}{additional arguments passed on to multiEIC().}
}
\description{
return best gauss curve fit score for any of the files in rawdata for each feature returned by multiEIC.
}
