% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_EICs.R
\name{bestgauss}
\alias{bestgauss}
\title{bestgauss}
\usage{
bestgauss(...)
}
\arguments{
\item{...}{additional arguments passed on to multiEIC().}

\item{rawdata}{a (named) list of xcmsRaw objects, e.g. generated by Mosaic::loadRawM}
}
\description{
return best gauss curve fit score for any of the files in rawdata for each feature returned by multiEIC.
}
