% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{buildMseekFT,xsAnnotate-method}
\alias{buildMseekFT,xsAnnotate-method}
\alias{buildMseekFT}
\alias{buildMseekFT,XCMSnExp-method}
\alias{buildMseekFT,data.frame-method}
\alias{loadMseekFT,character-method}
\alias{saveMseekFT,MseekFT-method}
\alias{saveMseekFT}
\title{buildMseekFT}
\usage{
\S4method{buildMseekFT}{xsAnnotate}(object, processHistory = list(), ...)

\S4method{buildMseekFT}{XCMSnExp}(object, processHistory = list(), ...)

\S4method{buildMseekFT}{data.frame}(object, processHistory = list(), ...)

\S4method{loadMseekFT}{character}(object)

\S4method{saveMseekFT}{MseekFT}(object, file, writeCSV = FALSE,
  writeRDS = TRUE)
}
\arguments{
\item{object}{an xsAnnotate, xcmsSet or data.frame object.}

\item{processHistory}{a list of \code{\link[xcms]{processHistory}} objects,
will override the processHistory list that may be supplied by \code{object}}

\item{...}{additional arguments passed to \code{\link{constructFeatureTable}}}

\item{file}{file path to write to}

\item{object}{an \code{MseekFT} object.}
}
\value{
an object of class \code{MseekFT}, with processHistory extended from
 \code{object} if applicable.

the \code{MseekFT} object, with saving event added to processHistory
}
\description{
Methods to build an \code{MseekFT} object from an xsAnnotate,
 xcmsSet or data.frame object, and to load or save \code{MseekFT} objects

save a \code{MseekFT} object to a file, registering the save event in the processHistory
}
