% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_buildMseekFT.R
\docType{methods}
\name{buildMseekFT}
\alias{buildMseekFT}
\alias{buildMseekFT,XCMSnExp-method}
\alias{buildMseekFT,data.frame-method}
\alias{loadMseekFT,character-method}
\alias{loadMseekFT}
\alias{saveMseekFT,MseekFT-method}
\alias{saveMseekFT}
\alias{loadMseekGraph,character-method}
\alias{loadMseekGraph}
\title{buildMseekFT}
\usage{
\S4method{buildMseekFT}{xsAnnotate}(object, processHistory = list(), ...)

\S4method{buildMseekFT}{XCMSnExp}(object, processHistory = list(), ...)

\S4method{buildMseekFT}{data.frame}(object, processHistory = list(), ...)

\S4method{loadMseekFT}{character}(object)

\S4method{saveMseekFT}{MseekFT}(object, file, writeCSV = FALSE,
  writeRDS = TRUE)

\S4method{loadMseekGraph}{character}(object)
}
\arguments{
\item{object}{an xsAnnotate, xcmsSet or data.frame object. For \code{saveMseekFT},
an \code{MseekFT} object, for \code{loadMseekFT} a charcater vector of file paths}

\item{processHistory}{a list of \code{\link[xcms]{processHistory}} objects,
will override the processHistory list that may be supplied by \code{object}}

\item{...}{additional arguments passed to \code{\link{constructFeatureTable}}}

\item{file}{file path to write to}

\item{writeCSV}{write object to a .csv file (data frame only)}

\item{writeRDS}{write object to an .mskFT (RDS) file; saves the entire object}
}
\value{
an object of class \code{MseekFT}, with processHistory extended from
 \code{object} if applicable.

the \code{MseekFT} object, with saving event added to processHistory
}
\description{
Methods to build an \code{MseekFT} object from an xsAnnotate,
 xcmsSet or data.frame object, and to load or save \code{MseekFT} objects

\code{loadMseekFT}: load a \code{MseekFT} object from an .mskFT file

\code{saveMseekFT}: save a \code{MseekFT} object to a file, registering the save event in the processHistory

\code{loadMseekGraph}: load a \code{MseekGraph} object from an 
.mskg or .graphml file (file extensions inform how file is loaded in)
}
