% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_calcMZ.R
\name{calcMZs}
\alias{calcMZs}
\title{calcMZs}
\usage{
calcMZs(df, charges = c(1), carrier = "H", monoisotopic = T,
  mf_column = "formula", adduct = NULL)
}
\arguments{
\item{df}{data.frame that contains a column with molecular formulas (MFs)}

\item{charges}{vector of positive or negative integers indicating the charge states to calculate}

\item{carrier}{charge carrier - Molecular formula will be added to or removed from MFs for each charge}

\item{monoisotopic}{if true, will calculate the monoisotopic mass m/z values, will calculate m/z of the most abundant isotope peak otherwise (will be slower but preferrable e.g. for large organic molecules)}

\item{mf_column}{name of column in df that contains MFs}
}
\description{
Calculate charge-to-mass ratios for molecula formulas in a data.frame, retaining the dataframes columns, but overriding columns mz, charge and ion
}
