% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure_functions.R
\name{checkFolders}
\alias{checkFolders}
\title{checkFolders}
\usage{
checkFolders(query = paste0(LETTERS, ":/"))
}
\arguments{
\item{query}{character vector with folders to search for, by default looks for drives in Windows file system}
}
\description{
Looks for folders
}
