% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_columnSelModule.R
\name{columnSelModule}
\alias{columnSelModule}
\title{columnSelModule}
\usage{
columnSelModule(input, output, session, reactives = reactive({    
  (list()) }), values = reactiveValues(), static = list(),
  load = reactive({     list() }))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{reactives}{import reactive data from the shiny session}

\item{values}{import reactiveValues from the shiny session}

\item{static}{import static values}

\item{load}{data to load from previous session (not implemented)}
}
\description{
server module for selection of columns to show in main Table
}
