% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_columnSelModule.R
\name{ColumnSelModule}
\alias{ColumnSelModule}
\alias{ColumnSelModuleUI}
\title{ColumnSelModule}
\usage{
ColumnSelModule(input, output, session,
  values = reactiveValues(featureTables = featureTables, MainTable =
  MainTable))

ColumnSelModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
If the server module for this Module returns something,
 it is described in \code{Details}.
}
\description{
server module for selection of columns to show in main Table
}
\section{Functions}{
\itemize{
\item \code{ColumnSelModule}: server logic for ColumnSelModule

\item \code{ColumnSelModuleUI}: UI elements for ColumnSelModule
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
