% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_densplotModule.R
\name{densplotModule}
\alias{densplotModule}
\title{densplotModule}
\usage{
densplotModule(input, output, session, mx, heading = "Default")
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{mx}{}

\item{heading}{}
}
\description{
Plot density of a vector or matrix
}
