% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_densPlotWidget.R
\name{densplotWidget}
\alias{densplotWidget}
\alias{densplotModuleUI}
\title{densplotWidget}
\usage{
densplotWidget(input, output, session, reactives = reactive({     list()
  }))

densplotModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{reactives}{list of arguments passed to \code{\link{densplot}}}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()}
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\description{
Plot density of a vector or matrix
}
\section{Functions}{
\itemize{
\item \code{densplotWidget}: server logic for densplotWidget

\item \code{densplotModuleUI}: UI elements for densplotWidget
}}

