% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_plotting_modules.R
\name{EICmodule}
\alias{EICmodule}
\title{EICmodule}
\usage{
EICmodule(input, output, session, tag, set = list(layouts = MSData$layouts,
  RTcorr = NULL, activeGroup = "Group1", filelist = MSData$filelist, data =
  MSData$data, mz = 542.32405482, rtr = NULL, active = T), keys)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{tag}{id to be used in ns()}

\item{set}{Import data from the shiny session}
}
\description{
server module for interactive EIC view
}
