% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_EICOptionsModule.R
\name{EICOptionsModule}
\alias{EICOptionsModule}
\alias{EICOptionsModuleUI}
\title{EICOptionsModule}
\usage{
EICOptionsModule(input, output, session, values)

EICOptionsModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
If the server module for this Module returns something,
 it is described in \code{Details}.
}
\description{
Module to change EIC options
}
\details{
Returns nothing, but writes to \code{values$GlobalOpts}
}
\section{Functions}{
\itemize{
\item \code{EICOptionsModule}: server logic for the EICOptionsModule

\item \code{EICOptionsModuleUI}: UI elements for the EICOptionsModule
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
