% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_plotting_EICs.R
\name{EICplot}
\alias{EICplot}
\title{EICplot}
\usage{
EICplot(EICs, cx = 1, ylim = c(0, max(unlist(EICs[[1]][, "tic"]))),
  xlim = c(min(unlist(EICs[[1]][, "rt"])), max(unlist(EICs[[1]][,
  "rt"])))/60, legendtext = paste(sub("^([^.]*).*", "\\\\1",
  basename(row.names(EICs[[1]])))), colr = topo.colors(nrow(EICs[[1]]),
  alpha = 1), heading = "test", relto = NULL, TIC = F, single = F,
  midline = 100, lw = 1, yzoom = 1, ylabshift = 0)
}
\arguments{
\item{ylim}{numeric(2) min and max visible rt value (in seconds)}

\item{xlim}{numeric(2) min and max visible intensity value (in seconds)}

\item{legendtext}{character() with item for each shown EIC for the plot legend}

\item{colr}{color range (actual vector of color values)}

\item{heading}{plot title}

\item{relto}{normalize intensities to this number}

\item{TIC}{if TRUE plot TIC}

\item{single}{TRUE if this is not part of a composite plot}

\item{midline}{numeric() of y-axis positions where a dotted vertical line should be plotted}

\item{lw}{line width for plot lines}

\item{yzoom}{zoom factor into y-axis}

\item{ylabshift}{shift horizontal position of y axis label}

\item{EIClistItem}{item from a list of EICs from METABOseek:multiEIC}
}
\value{
sends a single EIC plot to the current plotting device
}
\description{
generate one EIC plot for multiple files in one plot window.
}
