% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_EICs.R
\name{exIntensities}
\alias{exIntensities}
\title{exIntensities}
\usage{
exIntensities(rawfile = rawdata[[1]], mz = tb$mz, ppm = 5,
  rtw = data.frame(rta[[1]]$rtmin - 5, rta[[1]]$rtmax + 5),
  baselineSubstract = T, SN = NULL)
}
\arguments{
\item{rawfile}{an xcmsRaw object}

\item{mz}{numeric (): m/z values (same length as nrow(rtw)).}

\item{ppm}{m/z width (+/- ppm from values defined in mz)}

\item{rtw}{a data.frame with minimum (column 1) and maximum (column 2) retention time values (in seconds) in each row.}

\item{baselineSubstract}{substract baseline when calculating intensities}

\item{SN}{signal to noise ratio. If not NULL, all peaks with max/min peak intensity below this will be reported as intensity 0. Requires Baselinesubstraction to be off.}
}
\description{
Lightweight variant of multiEIC to extract average in a file, more suitable for large featuretables.
}
