% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_EICs.R
\name{exIntensities}
\alias{exIntensities}
\title{exIntensities}
\usage{
exIntensities(
  rawfile,
  mz,
  ppm = 5,
  rtw = data.frame(0, 10),
  baselineSubtract = T,
  SN = NULL,
  areaMode = F
)
}
\arguments{
\item{rawfile}{an xcmsRaw object}

\item{mz}{numeric (): m/z values (same length as nrow(rtw)).}

\item{ppm}{m/z width (+/- ppm from values defined in mz)}

\item{rtw}{a data.frame with minimum (column 1) and maximum (column 2) retention time values (in seconds) in each row.}

\item{baselineSubtract}{subtract baseline when calculating intensities}

\item{SN}{signal to noise ratio. If not NULL, all peaks with max/min peak intensity below this will be reported as intensity 0. Requires Baselinesubstraction to be off.}

\item{areaMode}{if TRUE, will calculate peak areas rather than mean intensities}
}
\value{
numeric vector with intensity values for all features defined by 
\code{mz} and \code{rt} in \code{rawfile}
}
\description{
Lightweight variant of multiEIC to extract average in a file, more suitable for large featuretables.
}
\section{Functions}{
\itemize{
\item \code{exIntensities}: extract intensities base function
}}

