% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_FeatureTable_analysis.R
\name{featureCalcs}
\alias{featureCalcs}
\title{featureCalcs}
\usage{
featureCalcs(df, massdef = T)
}
\arguments{
\item{df}{a feature table data.frame with a column "mz"}

\item{massdef}{if true, calculate the mass defect for each entry in df from their mz values}
}
\value{
\code{df} with an additional column, \code{massdefppm}
}
\description{
Calculate simple parameters from feature table data.frame. 
Currently ONLY calculates the mass defect in ppm.
}
