% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_featurePlotModule.R
\name{featurePlotModule}
\alias{featurePlotModule}
\title{featurePlotModule}
\usage{
featurePlotModule(input, output, session, FT, rname)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{FT}{Mosaic Feature Table object}

\item{rname}{row names of rows in FT selected for plotting}
}
\description{
Quickplots (ggplot-based plots of feature properties)
}
