% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_getters.R
\name{FeatureTable}
\alias{FeatureTable}
\title{FeatureTable}
\usage{
FeatureTable(x, ...)
}
\arguments{
\item{values}{reactiveValues or list object with at least a values$featureTables$tables}

\item{tableID}{table name}
}
\description{
FeatureTable
}
