% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset_featureTables.R
\name{FeatureTable}
\alias{FeatureTable}
\alias{FeatureTable<-}
\title{FeatureTable}
\usage{
FeatureTable(x, ...)

FeatureTable(x, ...) <- value
}
\arguments{
\item{x}{the \code{values} object}

\item{value}{the value that gets set}

\item{values}{reactiveValues or list object with at least a values$featureTables$tables}

\item{tableID}{table ID to retrieve. will use currently active Feature Table if not set.
For setter, will create a new Feature Table under this ID if ID does not exist, 
otherwise will override or update the selected Feature Table}

\item{replace}{replace the current Feature table with the given tableID? will update it if FALSE.}
}
\description{
Getter and setter methods to retireve a Feature Table from \code{values}
}
