% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container_FeatureTableContainer.R
\name{FeatureTableContainer}
\alias{FeatureTableContainer}
\title{FeatureTableContainer}
\usage{
FeatureTableContainer(input, output, session,
  values = reactiveValues(featureTables = featureTables, MSData = MSData,
  GlobalOpts = GlobalOpts, projectData = projectData))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}
}
\description{
Module to apply filters to a featureTable (UI)
}
