% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_FilterModule.R
\name{FilterModule}
\alias{FilterModule}
\title{FilterModule}
\usage{
FilterModule(input, output, session, lab = "Filter", tag, df, presets)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{lab}{heading of this module in UI}

\item{tag}{same as id in UI module}

\item{df}{data frame to be filtered}

\item{presets}{Mosaic featureTable object filter$filter}
}
\description{
Module to apply filters to a featureTable (UI)
}
