% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_peakPicking.R
\name{find_peaks2}
\alias{find_peaks2}
\title{findPeaks2}
\usage{
find_peaks2(EIC, ...)
}
\arguments{
\item{EIC}{an EIC output item}

\item{...}{arguments passed to peakDetect()}
}
\description{
Find peaks in a multiEIC output list item
}
