% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_FindMS2scansModule.R
\name{FindMS2ScansModule}
\alias{FindMS2ScansModule}
\alias{FindMS2ScansModuleUI}
\title{FindMS2ScansModule}
\usage{
FindMS2ScansModule(input, output, session, values = reactiveValues(MSData
  = MSData, featureTables = featureTables), static = list(tooltip =
  "Find MS2 scans for all parent m/zs in feature table", label =
  "Find MS2 scans"))

FindMS2ScansModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with \code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with \code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with \code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} (must match \code{id} in \code{\link[shiny]{callModule}} for the server Module)}
}
\value{
If the server module for this Module returns something, it is described in \code{Details}.
}
\description{
A one-button modue to enrich a Feature Table with information about available 
MS2 scans for each molecular feature
}
\section{Functions}{
\itemize{
\item \code{FindMS2ScansModule}: server logic

\item \code{FindMS2ScansModuleUI}: UI elements
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
