% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_table_data_analysis.R
\name{foldChange}
\alias{foldChange}
\title{foldChange}
\usage{
foldChange(mx, groups, ctrl = NULL, calc = "mean", topgroup = T,
  maxFold = T, foldMaxK = 2, foldmode = "simple")
}
\arguments{
\item{mx}{a matrix of numeric (intensity) values}

\item{groups}{named list of intensity columns listed by group (as supplied by $anagroupnames or $anagroupnames_norm of MosaicFT objects)}

\item{ctrl}{character() naming the control group(s)}

\item{calc}{currently has to be "mean", compare rowMeans of one group vs. rowMeans of other groups (and optionally rowMeans of controls only)}

\item{topgroup}{if TRUE, return group with highest intensity for each feature}

\item{maxFold}{if TRUE, make a column with maximum fold change between any two groups for each feature}

\item{foldMaxK}{if not NULL, make column with fold change of highest group value over foldMaxK largest group value.}

\item{foldmode}{if "complex", gives ratios between all groups}
}
\description{
calculate fold changes between grouped columns of a matrix
}
