% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset_featureTables.R
\name{FTselection}
\alias{FTselection}
\title{FTselection}
\usage{
FTselection(x, ...)
}
\arguments{
\item{x}{a MseekValues (reactivevalues) object}

\item{liveView}{if TRUE, will get data from the liveView of the MainTable directly (default: FALSE)}
}
\description{
get the rows that are currently selected in MainTable window.
Note: by default, will get its data from the underlying Feature Table,
 not using changes (such as comments) made immediately prior to selection, 
 but including all columns, including those not visible in the current table view.
}
