% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_deprecated.R
\name{getAllScans}
\alias{getAllScans}
\title{getAllScans}
\usage{
getAllScans(scanlist, MSData, removeNoise = NULL, type = c("ms2",
  "ms1", "acquisition"))
}
\arguments{
\item{scanlist}{as returned by makeScanlist2()}

\item{MSData}{list of xcmsRaw objects}

\item{removeNoise}{if not NULL, a numeric specifying a relative intensity
cutoff (<1, relative to maximum peak intensity)}

\item{type}{specify what the scan number represents, either 
"acquisition", "ms1" or "ms2".}
}
\value{
a list of MS spectra (each in matrix format)
}
\description{
get a list of all scans in a scanlist
}
