% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_EICs.R
\name{getgauss}
\alias{getgauss}
\title{getgauss}
\usage{
getgauss(y, pval = 1)
}
\arguments{
\item{y}{numeric() to fit the curve}
}
\description{
fit a gauss curve into a curve (numeric vector). Note that results will be skewed 
if scanrate is low or heterogeneous (e.g. ddMS2 experiments).
}
