% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_GetIntensitiesModule.R
\name{GetIntensitiesModule}
\alias{GetIntensitiesModule}
\alias{GetIntensitiesModuleUI}
\title{GetIntensitiesModule}
\usage{
GetIntensitiesModule(input, output, session,
  values = reactiveValues(MSData = MSData, featureTables = featureTables,
  GlobalOpts = GlobalOpts))

GetIntensitiesModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
If the server module for this Module returns something,
 it is described in \code{Details}.
}
\description{
One-button module to get intensities for all features in a feature 
table across the loaded MS data files
}
\section{Functions}{
\itemize{
\item \code{GetIntensitiesModule}: server logic

\item \code{GetIntensitiesModuleUI}: server logic
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
