% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_GetIntensitiesModule.R
\name{GetIntensitiesModule}
\alias{GetIntensitiesModule}
\title{GetIntensitiesModule}
\usage{
GetIntensitiesModule(input, output, session,
  values = reactiveValues(MSData = MSData, featureTables = featureTables,
  MainTable = MainTable, GlobalOpts = GlobalOpts))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
server module for saving Tables
}
