% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_SIRIUS.R
\name{getSiriusTree}
\alias{getSiriusTree}
\title{getSiriusTree}
\usage{
getSiriusTree(paths, formula)
}
\arguments{
\item{paths}{list generated by getSirius()}

\item{formula}{molecular formula prediction to get details for}
}
\description{
get additional SIRIUS results for a spectrum
}
