% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_GetSiriusWidget.R
\name{GetSiriusWidget}
\alias{GetSiriusWidget}
\alias{GetSiriusWidgetUI}
\title{GetSiriusWidget}
\usage{
GetSiriusWidget(
  input,
  output,
  session,
  reactives = reactive({     list(sirpath = file.path(.MseekOptions$siriusFolder,
    "sirius"), ms2, parentmz, outfolder = getwd(), ms1 = NULL, ion = "[M+?]+", comments =
    "", rt = "", charge = 1, fingerid = T, scanindices = "", moreOpts = "", config =
    list(IsotopeSettings.filter = TRUE, FormulaSearchDB = c("ALL_BUT_INSILICO", "ALL",
    "BIO", "METACYC", "CHEBI", "COCONUT", "ECOCYCMINE", "GNPS", "HMDB", "HSDB", "KEGG",
    "KEGGMINE", "KNAPSACK", "MACONDA", "MESH", "NORMAN", "UNDP", "PLANTCYC", "PUBCHEM",
    "PUBMED", "YMDB", "YMDBMINE",          "ZINCBIO"), Timeout.secondsPerTree = 0,
    FormulaSettings.enforced = "HCNOP[5]S", Timeout.secondsPerInstance = 0,
    AdductSettings.detectable =
    "[[M+K]+,[M+H3N+H]+,[M+Na]+,[M-H4O2+H]+,[M-H2O+H]+,[M+H]+]",
    UseHeuristic.mzToUseHeuristicOnly = 650, AlgorithmProfile = "orbitrap",
    IsotopeMs2Settings = "IGNORE", MS2MassDeviation.allowedMassDeviation = "5.0ppm",
    NumberOfCandidatesPerIon = 1, UseHeuristic.mzToUseHeuristic = 300,
    FormulaSettings.detectable = ",", NumberOfCandidates = 10, StructureSearchDB =
    c("ALL_BUT_INSILICO",          "ALL", "BIO", "METACYC", "CHEBI", "COCONUT",
    "ECOCYCMINE", "GNPS", "HMDB", "HSDB", "KEGG", "KEGGMINE", "KNAPSACK", "MACONDA",
    "MESH", "NORMAN", "UNDP", "PLANTCYC", "PUBCHEM", "PUBMED", "YMDB", "YMDBMINE",
    "ZINCBIO"), AdductSettings.fallback = "[[M+K]+,[M+Na]+,[M+H]+]", RecomputeResults =
    TRUE), force = T) })
)

GetSiriusWidgetUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{reactives}{a reactive() returning a list of \code{\link{runSirius}()}
parameters}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()}
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\description{
Widget that starts a SIRIUS analysis as specified in reactives({}, which will 
add an entry to the index file in the SIRIUS folder to be observed by the SiriusModule
}
\section{Functions}{
\itemize{
\item \code{GetSiriusWidget}: server logic

\item \code{GetSiriusWidgetUI}: UI elements
}}

