% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset_MSData.R
\name{getSpectra}
\alias{getSpectra}
\alias{getSpectra.reactivevalues}
\alias{getSpectra.OnDiskMSnExp}
\title{getSpectra}
\usage{
getSpectra(data, ...)

\method{getSpectra}{reactivevalues}(data, ...)

\method{getSpectra}{OnDiskMSnExp}(data, index, xcmsRawobject = NULL, SpectrumClass = F)
}
\arguments{
\item{data}{an OnDiskMSnExp or MseekValues (reactivevalues) object}

\item{...}{arguments passed to \code{getSpectra.OnDiskMSnExp}}

\item{index}{a numeric vector of scan indices}

\item{xcmsRawobject}{list of xcmsRaw objects}

\item{SpectrumClass}{if TRUE, will return a list of Spectum class objects, 
otherwise will return a list of matrices with mz and intensity values}
}
\value{
a list of Spectrum1 or Spectrum2 objects
}
\description{
get a list of spectra using indices from the OnDiskMSnExp object
}
