% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_getters.R
\name{getSpectra}
\alias{getSpectra}
\title{rawSpectrum2}
\usage{
getSpectra(index, MSnData, xcmsRawobject = NULL)
}
\description{
create a list of Spectrum (Spectrum1 and/or Spectrum2) objects from a list of xcmsRaw onjects if available
}
