% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_GlobalOptionsModule.R
\name{GlobalOptionsModule}
\alias{GlobalOptionsModule}
\title{GlobalOptionsModule}
\usage{
GlobalOptionsModule(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\description{
Module to change global options
}
