% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_plotting_EICs.R
\name{legendplot}
\alias{legendplot}
\title{legendplot}
\usage{
legendplot(...)
}
\arguments{
\item{...}{all arguments passed on to \code{\link[graphics]{legend}()}}
}
\value{
calls a new \code{plot()} with a legend
}
\description{
Plot a legend in an otherwise empty plot
}
