% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_Scans.R
\name{listMS2scans}
\alias{listMS2scans}
\title{listMS2scans}
\usage{
listMS2scans(mz, rt, ppm, rtw, MSData, rtMatch = F)
}
\arguments{
\item{mz}{expected target (parent) m/z}

\item{rt}{expected target (parent) retention time in seconds}

\item{ppm}{target (parent) m/z tolerance in ppm}

\item{rtw}{target (parent) rt tolerance in seconds}

\item{MSData}{list of xcmsRaw objects}

\item{rtMatch}{assign MS2 scans only to the matching feature with the closest rt}
}
\value{
a character vector
}
\description{
Searches for MS2 scans with a given parent m/z and directly converts the
 resulting table into a single string
}
\details{
Also, see \code{\link{Parentsearch}}
}
