% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_formatting_utilities.R
\name{ListToReactiveValues}
\alias{ListToReactiveValues}
\title{ListToReactiveValues}
\usage{
ListToReactiveValues(ls)
}
\arguments{
\item{ls}{a list object}
}
\description{
recursively converts lists to reactiveValues
}
