% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_LoadMSDataModule.R
\name{LoadMSDataModule}
\alias{LoadMSDataModule}
\alias{LoadMSDataModuleUI}
\title{LoadMSDataModule}
\usage{
LoadMSDataModule(input, output, session,
  values = reactiveValues(projectData = NULL, featureTables = NULL,
  MSData = NULL, GlobalOpts = NULL))

LoadMSDataModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
Nothing, but modifies \code{\link{values}}
}
\description{
Module for loading MS data into Metaboseek
}
\section{Functions}{
\itemize{
\item \code{LoadMSDataModule}: server logic

\item \code{LoadMSDataModuleUI}: UI elements
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
