% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_LoadNetworkModule.R
\name{LoadNetworkModule}
\alias{LoadNetworkModule}
\title{LoadNetworkModule}
\usage{
LoadNetworkModule(input, output, session,
  values = reactiveValues(featureTables = featureTables, MSData =
  MSData), reactives = reactive({     list(active = NetMod$active) }))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{tag}{id to be used in ns()}

\item{set}{Import data from the shiny session}
}
\description{
server module for loading Tables
}
