% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_msdata.R
\name{loadRawM}
\alias{loadRawM}
\title{loadRawM}
\usage{
loadRawM(filelist = mzxml_pos, MSn = T, workers = 10, rnames = filelist)
}
\arguments{
\item{filelist}{a list of mzXML or mzML files (character vector)}

\item{MSn}{should MSn data be read in? defaults to TRUE}

\item{workers}{How many cores to use (cf. BiocParallel and SnowParam, argument only used if more than 10 files are loaded).}

\item{rnames}{names of the xcmsRaw objects in the list returned, defaults to the filepaths of the source files.}
}
\description{
TIME CONSUMING. This step does not need to be repeated when adjusting other parameters
(e.g. feature list, EIC ppm or RT) Generates an R-readable data structure (a list of xcmsRaw objects) 
in memory from MS data files defined in the file list.
}
\details{
Note that there is an xcms function of the same name.
}
