% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_LoadTableModule.R
\name{LoadTableModule}
\alias{LoadTableModule}
\title{LoadTableModule}
\usage{
LoadTableModule(input, output, session,
  values = reactiveValues(projectData = projectData, featureTables =
  NULL), static = list(tooltip = "Load", label = "Load", format =
  list(header = T, sep = "\\t", quote = "\\"", stringsAsFactors = F),
  pattern = "\\\\.tGrouping$"))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{static}{Import data from the shiny session}

\item{reactives}{Import data from the shiny session}
}
\description{
server module for loading Tables
}
