% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MainTableModule.R
\name{MainTableModule}
\alias{MainTableModule}
\title{MainTableModule}
\usage{
MainTableModule(input, output, session,
  values = reactiveValues(featureTables = featureTables, GlobalOpts =
  GlobalOpts, projectData = projectData), static = list(height = 300,
  readOnly = T, contextMenu = T, fixedColumnsLeft = 1, invertReadOnly =
  NULL, controls = T, format = list(col = NULL, format = NULL)))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{set}{Import data from the shiny session}
}
\description{
server module for interactive mass spectrum view
}
