% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_formatting_utilities.R
\name{matchedToCharacter}
\alias{matchedToCharacter}
\title{matchedToCharacter}
\usage{
matchedToCharacter(matchedPatterns, sep = "|")
}
\arguments{
\item{matchedPatterns}{a list of named logical vectors}
}
\value{
a character vector
}
\description{
summarize named logical vectors into a single character vector
}
