% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_peakPicking.R
\name{mergepeaks2}
\alias{mergepeaks2}
\title{mergepeaks2}
\usage{
mergepeaks2(pl, rttol = 3, minint = 0, minrelint = 0, topN = 100)
}
\arguments{
\item{pl}{a peaklist, such as list items returned by find_peaks() (or its rbind product)}

\item{rttol}{retention time tolerance in seconds to merge peaks}

\item{minint}{minimum intensity for a peak to be retained (before merging)}

\item{minrelint}{minimum intensity of a peak to be retained (before merging), relative to largest peak}

\item{topN}{number of highest intensity peaks retained after merging}
}
\description{
merge peaks in a peaklist
}
