% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure_functions.R
\name{MosaicExamplePreload}
\alias{MosaicExamplePreload}
\title{MosaicExamplePreload}
\usage{
MosaicExamplePreload(tables = T, data = T)
}
\arguments{
\item{tables}{load example tables?}

\item{data}{load example MS data?}
}
\description{
Loads example data into the parent environment for testing and development purposes
}
