% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure_functions.R
\name{MosaicMinimalServer}
\alias{MosaicMinimalServer}
\title{MosaicMinimalServer}
\usage{
MosaicMinimalServer(exampleData = T, diagnostics = T)
}
\description{
A minimal server logic for Mosaic. Function has to be called in a shiny server function context. Will load example data if it is not present in the parent environment.
Save time by providing the parent environment with objects \code{tab1}, \code{tab2} and \code{MSD} by calling MosaicExamplePreload()
}
