% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_components.R
\name{MosaicMinimalUI}
\alias{MosaicMinimalUI}
\title{MosaicMinimalUI}
\usage{
MosaicMinimalUI(..., diagnostics = T, dashboard = F, id = NULL)
}
\description{
A minimal UI for Mosaic that can be extended with additional objects for testing and development purposes
}
