% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure_functions.R
\name{MosaicMinimalUI}
\alias{MosaicMinimalUI}
\title{MosaicMinimalUi}
\usage{
MosaicMinimalUI(..., diagnostics = T, dashboard = F)
}
\description{
A minimal UI for Mosaic that can be extended with additional objects for testing and development purposes
}
