% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_MosaicOptions.R
\name{MosaicOptions}
\alias{MosaicOptions}
\title{MosaicOptions}
\usage{
MosaicOptions(..., defaults = F)
}
\arguments{
\item{...}{parameters to be modified}

\item{defaults}{if TRUE, default MosaicOptions are loaded}
}
\description{
Load and/or change .MosaicOptions
}
