% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_FeatureTable_analysis.R
\name{MosCluster}
\alias{MosCluster}
\title{MosCluster}
\usage{
MosCluster(method = "clara", ...)
}
\arguments{
\item{method}{name of a function, defaults to "clara" (from the cluster package)}

\item{...}{additional arguments to the clustering method}

\item{mx}{matrix (of intensity values)}
}
\value{
a data.frame with one column, called \code{cluster__METHODNAME}, 
where METHODNAME is the applied \code{method}
}
\description{
Cluster a matrix
}
