% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MS2BrowserModule.R
\name{MS2BrowserModule}
\alias{MS2BrowserModule}
\title{MS2BrowserModule}
\usage{
MS2BrowserModule(input, output, session, tag, set = list(MSData = xcmsRaws,
  query = list(mz = NULL, rt = NULL)), keys = reactive({     keys() }))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{tag}{id to be used in ns()}

\item{set}{Import data from the shiny session}
}
\description{
server module for loading Tables
}
