% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container_MseekContainer.R
\name{MseekContainer}
\alias{MseekContainer}
\alias{MseekContainerUI}
\title{MseekContainer}
\usage{
MseekContainer(input, output, session)

MseekContainerUI(id)
}
\arguments{
\item{input}{arguments necessary for use with \link[shiny]{callModule}()}

\item{output}{arguments necessary for use with \link[shiny]{callModule}()}

\item{session}{arguments necessary for use with \link[shiny]{callModule}()}

\item{id}{id to be used to define a namespace via  \link[shiny]{NS}() (must match \code{id} in \link[shiny]{callModule} for the server Module)}
}
\description{
Module that contains the entire METABOseek program
}
\section{Functions}{
\itemize{
\item \code{MseekContainer}: server logic module, to be called with \link[shiny]{callModule}()

\item \code{MseekContainerUI}: contains the UI for the entire METABOseek program
}}

