% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation_templates.R
\name{MseekContainers}
\alias{MseekContainers}
\alias{MseekContainersUI}
\title{MseekContainers}
\usage{
MseekContainers(input, output, session, values)

MseekContainersUI(id)
}
\arguments{
\item{input, output, session}{arguments necessary for use with \code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} (must match \code{id} in \code{\link[shiny]{callModule}()} for the server Module)}
}
\value{
The server module for this container returns nothing
}
\description{
This is the MseekContainer template
}
\section{Functions}{
\itemize{
\item \code{MseekContainers}: generic server logic example for MseekContainers

\item \code{MseekContainersUI}: generic UI example for MseekContainers
}}

\section{MseekContainers}{

Containers are used to keep the code organised and can be characterized
 as Modules which contain other Modules, but do not contain add any functionality themselves.
  They should not contain any observers and are only here to pass \code{values} into modules,
   or provide ways for individual Modules to interact with each other.
}

\examples{
\dontrun{
ui <- function(request){
MseekContainerUI("Mseek")
}
server <- function(input, output, session) {
   callModule(MseekContainer, "Mseek")
}
Create Shiny app ----
shinyApp(ui, server)
}

}
\seealso{
\code{\link{values}} for a description of the \code{values} object
\itemize{
\item{General Information:} \code{\link{MseekContainers}}
\item{Specific containers:} \code{\link{MseekContainer}}, \code{\link{MainPageContainer}} 
\item{More on Shiny modules:} \url{http://shiny.rstudio.com/articles/modules.html}
}
}
