% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R, R/methods-MseekFT.R
\docType{methods}
\name{analyzeFT,MseekFT,listOrNULL,FTAnalysisParam-method}
\alias{analyzeFT,MseekFT,listOrNULL,FTAnalysisParam-method}
\alias{analyzeFT}
\alias{addProcessHistory,MseekFT,ProcessHistory-method}
\alias{addProcessHistory}
\alias{processHistory,MseekFT-method}
\alias{processHistory}
\alias{hasAdjustedRtime,MseekFT-method}
\alias{hasAdjustedRtime}
\alias{rename,MseekFT-method}
\alias{previousStep,MseekFT-method}
\alias{previousStep}
\alias{\code{previousStep()}}
\alias{extract}
\alias{the}
\alias{most}
\alias{recent}
\alias{\code{ProcessHistory}}
\alias{object}
\alias{from}
\alias{an}
\alias{object,}
\alias{representing}
\alias{last}
\alias{recorded}
\alias{changes}
\alias{made}
\alias{to}
\alias{object.}
\alias{getMseekIntensities,MseekFT,listOrNULL,missing-method}
\alias{getMseekIntensities}
\alias{searchFunParam,MseekFT-method}
\alias{intensityCols,MseekFT-method}
\alias{intensityCols<-,MseekFT-method}
\alias{groupingTable,MseekFT-method}
\alias{groupingTable<-,MseekFT,data.frame-method}
\alias{FTNormalize,MseekFT-method}
\alias{FTNormalize}
\alias{FTBasicAnalysis,MseekFT-method}
\alias{FTOldPeakShapes,MseekFT,listOrNULL-method}
\alias{FTPeakShapes,MseekFT,listOrNULL-method}
\alias{FTMzMatch,MseekFT-method}
\alias{FTT.test,MseekFT-method}
\alias{FTAnova,MseekFT-method}
\alias{FTCluster,MseekFT-method}
\alias{FTPCA,MseekFT-method}
\alias{FTFilter,data.frame-method}
\alias{FTFilter,MseekFT-method}
\title{MseekFT-class}
\usage{
\S4method{analyzeFT}{MseekFT,listOrNULL,FTAnalysisParam}(object, MSData,
  param)

\S4method{addProcessHistory}{MseekFT,ProcessHistory}(object, ph)

\S4method{processHistory}{MseekFT}(object)

\S4method{hasAdjustedRtime}{MseekFT}(object)

\S4method{rename}{MseekFT}(object, name)

\S4method{previousStep}{MseekFT}(object)

\S4method{getMseekIntensities}{MseekFT,listOrNULL,missing}(object, rawdata,
  adjustedRT = TRUE, ppm = 5, rtrange = TRUE, rtw = 5)

\S4method{searchFunParam}{MseekFT}(object, fun = "", index = FALSE)

\S4method{intensityCols}{MseekFT}(object)

\S4method{intensityCols}{MseekFT}(object) <- value

\S4method{groupingTable}{MseekFT}(object)

\S4method{groupingTable}{MseekFT,data.frame}(object) <- value

\S4method{FTNormalize}{MseekFT}(object, logNormalized = FALSE)

\S4method{FTBasicAnalysis}{MseekFT}(object, intensityCols = NULL,
  grouping = NULL, controlGroup = NULL)

\S4method{FTOldPeakShapes}{MseekFT,listOrNULL}(object, rawdata, ppm = 5,
  workers = 1)

\S4method{FTPeakShapes}{MseekFT,listOrNULL}(object, rawdata, ppm = 5,
  workers = 1)

\S4method{FTMzMatch}{MseekFT}(object, db, ppm = 5, mzdiff = 0.001)

\S4method{FTT.test}{MseekFT}(object, intensityCols = NULL,
  grouping = NULL, adjmethod = "bonferroni")

\S4method{FTAnova}{MseekFT}(object, intensityCols = NULL,
  grouping = NULL)

\S4method{FTCluster}{MseekFT}(object, intensityCols = NULL,
  numClusters = 100L)

\S4method{FTPCA}{MseekFT}(object, intensityCols = NULL,
  featureMode = FALSE)

\S4method{FTFilter}{data.frame}(object, filters = list(),
  sortBy = charachter(), decreasing = TRUE)

\S4method{FTFilter}{MseekFT}(object, filters = list(),
  sortBy = charachter(), decreasing = TRUE)
}
\arguments{
\item{object}{an \code{MseekFT} object.}

\item{name}{file path to write to}

\item{adjustedRT}{use adjusted RTs for all samples for which it is available}

\item{index}{get index in list instead of actual value}

\item{featureMode}{if TRUE, will cluster molecular features by intensities 
across samples. If FALSE, will cluster samples by intensities across features}

\item{filters}{a list of filters}

\item{...}{additional arguments passed to \code{exIntensities}}

\item{filters}{a list of filters}
}
\value{
The \code{addProcessHistory} method returns the input object with the
provided \code{\link{ProcessHistory}} appended to the process history.

For \code{hasAdjustedRtime}: A logical, indicating if the object 
contains retention time correction information.
}
\description{
Copied description and Method template from \code{xcms}.
\code{addProcessHistory}: adds (appends) a single
\code{\link{ProcessHistory}} object to the \code{.processHistory} slot.

\code{processHistory}: extract a list of \code{ProcessHistory} objects from an object,
 representing changes made to the object.

\code{hasAdjustedRtime}: check if the object contains retention time
correction information.

save a \code{MseekFT} object to a file, registering the save event in the processHistory

\code{FTNormalize}: Normalize a feature table such that the mean
 values of all intensity columns will be equal.

\code{FTPCA}: Calculate Principal Component Analysis to cluster either features or samples

\code{FTFilter}: apply a list of filters to a MseekFT object

\code{FTFilter}: apply a list of filters to a MseekFT object
}
\author{
Johannes Rainer, Maximilian Helf
}
