% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{processHistory,MseekFT-method}
\alias{processHistory,MseekFT-method}
\alias{rename,MseekFT-method}
\title{MseekFT-class}
\usage{
\S4method{processHistory}{MseekFT}(object)

\S4method{rename}{MseekFT}(object, name)
}
\arguments{
\item{object}{an \code{MseekFT} object.}

\item{name}{file path to write to}
}
\description{
extract a list of \code{ProcessHistory} objects from an object,
 representing changes made to the object.

save a \code{MseekFT} object to a file, registering the save event in the processHistory
}
