% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_MseekGraph.R
\name{setLayout,MseekGraph-method}
\alias{setLayout,MseekGraph-method}
\alias{setLayout}
\alias{MseekGraphs}
\alias{buildMseekGraph}
\alias{saveMseekGraph,MseekGraph-method}
\alias{saveMseekGraph}
\alias{loadMseekGraph,character-method}
\alias{loadMseekGraph}
\alias{simplify,MseekGraph-method}
\title{MseekGraphs}
\usage{
\S4method{setLayout}{MseekGraph}(object, layoutFunction = "qgraph::qgraph.layout.fruchtermanreingold")

\S4method{buildMseekGraph}{MseekFT}(
  object,
  cosineThreshold = 0.6,
  layoutFunction = "qgraph::qgraph.layout.fruchtermanreingold"
)

\S4method{saveMseekGraph}{MseekGraph}(object, file, writeGraphML = FALSE, writeRDS = TRUE)

\S4method{loadMseekGraph}{character}(
  object,
  layoutFunction = "qgraph::qgraph.layout.fruchtermanreingold"
)

\S4method{simplify}{MseekGraph}(
  object,
  rankBy = NULL,
  maxK = 10,
  cosineThreshold = NULL,
  layoutFunction = "qgraph::qgraph.layout.fruchtermanreingold"
)
}
\arguments{
\item{layoutFunction}{function to use for layout of the resulting graph}

\item{cosineThreshold}{if length > 0, will remove all edges with a cosine value below this}

\item{file}{file path to write to}

\item{writeGraphML}{write object to a .graphML file (will drop metadata and some columns)}

\item{writeRDS}{write object to an .mskg (RDS) file; saves the entire object}

\item{rankBy}{if length > 0, will use this edge attribute to rank keep only the top \code{maxK} edges}

\item{maxK}{if length > 0, will only alllow at most this many edges from each node}
}
\value{
the \code{MseekFT} object, with saving event added to processHistory
}
\description{
\code{setLayout():} change the graph layout of an existing MseekGraph object.

\code{MseekGraph} objects are \code{\link[igraph]{igraph}} objects
which have layout information and slots that are the same as MseekFT objects,
except that the bulk of data is kept in an \code{igraph} object instead of 
a \code{data.frame}. These objects are read and displayed by the 
\code{\link{NetworkModule}}.

\code{saveMseekGraph}: save a \code{MseekGraph} object to a
 file, registering the save event in the processHistory if the format allows

\code{loadMseekGraph}: load a \code{MseekGraph} object from an 
.mskg or .graphml file (file extensions inform how file is loaded in)

\code{simplify}: simplify an MseekGraph object, reducting the number of edges 
using additional filters.
}
