% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_components.R
\name{MseekMinimalUI}
\alias{MseekMinimalUI}
\title{MseekMinimalUI}
\usage{
MseekMinimalUI(..., diagnostics = T, dashboard = F, id = NULL)
}
\description{
A minimal UI for Mseek that can be extended with additional objects for testing and development purposes
}
