% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_MseekOptions.R
\name{MseekOptions}
\alias{MseekOptions}
\title{MseekOptions}
\usage{
MseekOptions(..., defaults = F)
}
\arguments{
\item{...}{parameters to be modified}

\item{defaults}{if TRUE, default MseekOptions are loaded}
}
\description{
Load and/or change .MseekOptions
}
