% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_EICs.R
\name{multiEIC}
\alias{multiEIC}
\title{multiEIC}
\usage{
multiEIC(rawdata = rawcoll, mz = data.frame(mzmin = pcame_mini$mzmin, mzmax
  = pcame_mini$mzmax), rt = data.frame(rtmin = pcame_mini$rt - 5, rtmax =
  pcame_mini$rt + 5), rnames = row.names(pcame_mini), byFile = F, XIC = F,
  getgauss = F, RTcorr = NULL)
}
\arguments{
\item{rawdata}{a (named) list of xcmsRaw objects, e.g. generated by Mosaic::loadRawM, OR an MSnbase OnDiskMSnExp object}

\item{mz}{a data.frame with minimum (column 1) and maximum (column 2) m/z values in each row.}

\item{rt}{a data.frame with minimum (column 1) and maximum (column 2) retention time values (in seconds) in each row.}

\item{rnames}{names of the rows in the result matrix, typically the row.names of the feature table, or the filenames if byFile =T.}

\item{byFile}{a data.frame with columns File and Group , holding file paths and group names, respectively.}

\item{XIC}{deprecated}

\item{getgauss}{additionally, fit a gauss curve to each EIC (time consuming), defaults to FALSE}

\item{RTcorr}{if not NULL, this RTcorr object will be used to adjust retention times.}
}
\description{
Extract EICs from multiple files for multiple features.
}
