% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_msdata.R
\name{multiEICplus}
\alias{multiEICplus}
\title{multiEICplus}
\usage{
multiEICplus(adducts = c(0, 1, 2, 3), mz, rt, ...)
}
\arguments{
\item{adducts}{numeric() of mass shifts}

\item{...}{all other arguments passed on to multiEIC}
}
\description{
get EICs for adducts, isotopes and neutral loss masses
}
