% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MultiFilterModule.R
\name{MultiFilterModule}
\alias{MultiFilterModule}
\alias{MultiFilterModuleUI}
\title{MultiFilterModule}
\usage{
MultiFilterModule(
  input,
  output,
  session,
  values,
  static = list(lab = "Filter", modFeatureTable = T, name = NULL, activate = F)
)

MultiFilterModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{static}{a list of arguments that are not in a reactive context 
and hence do not change while the app is running.}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
Returns its internalValues
}
\description{
Module to apply filters to the active \code{\link{FeatureTable}}.
}
\section{Functions}{
\itemize{
\item \code{MultiFilterModule}: Server logic

\item \code{MultiFilterModuleUI}: UI elements
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
