% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MultiSpecmodule.R
\name{MultiSpecmodule}
\alias{MultiSpecmodule}
\title{MultiSpecmodule}
\usage{
MultiSpecmodule(input, output, session, tag, set = list(spec = list(xrange =
  NULL, yrange = NULL, maxxrange = NULL, maxyrange = NULL, sel = NULL, mz =
  NULL, data = NULL, MS2 = T), layout = list(lw = 1, cex = 1, controls = F, ppm
  = 5, active = T, height = 350), msdata = NULL), keys, static = list(title =
  "MS spectra"))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{tag}{id to be used in ns()}

\item{set}{Import data from the shiny session}
}
\description{
server module for multiple interactive mass spectrum views
}
