% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MultiSpecmodule.R
\name{MultiSpecmoduleUI}
\alias{MultiSpecmoduleUI}
\title{MultiSpecmoduleUI}
\usage{
MultiSpecmoduleUI(id)
}
\arguments{
\item{id}{id to be used in ns()}
}
\description{
UI module for interactive view of multiple spectra
}
