% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_FeatureTable_analysis.R
\name{multittest}
\alias{multittest}
\title{multittest}
\usage{
multittest(df = as.data.frame(mx), groups, ttest = T,
  adjmethod = "bonferroni")
}
\arguments{
\item{df}{a data.frame with numeric (intensity) values}

\item{groups}{named list of intensity columns listed by group (as supplied by $anagroupnames or $anagroupnames_norm of MosaicFT objects)}

\item{ttest}{if TRUE, ttest will be calculated}

\item{adjmethod}{method to adjust p values (passed on to stats::p.adjust)}
}
\description{
Calculate per-row t-test between grouped columns of a data.frame.
Note that for column groups with less than 2 members, a pseudo-ttest will be calculated via Mosaic::ttestx to avoid throwing errors.
}
