% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MZcalcModule.R
\name{MZcalcModule}
\alias{MZcalcModule}
\title{MZcalcModule}
\usage{
MZcalcModule(input, output, session,
  values = reactiveValues(featureTables = featureTables))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
server module for saving Tables
}
