% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MzqueryModule.R
\name{MzqueryModule}
\alias{MzqueryModule}
\title{MzqueryModule}
\usage{
MzqueryModule(input, output, session, tag, set = list(search =
  list(elements = NULL, mz = NULL, data = NULL)))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{tag}{id to be used in ns()}

\item{set}{Import data from the shiny session}
}
\description{
Chemcalc mz query module
server module to get molecular formula predictions from chemcalc
}
