% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_MzqueryModule.R
\name{MzqueryModule}
\alias{MzqueryModule}
\alias{MzqueryModuleUI}
\title{MzqueryModule}
\usage{
MzqueryModule(input, output, session, values, reactives = reactive({    
  list(mz = NULL) }))

MzqueryModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{reactives}{a list, wrapped in a \code{\link[shiny]{reactive}()} 
expression with settings for this module. See \code{details}.}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
Returns its internalValues
}
\description{
Use the \code{MassTools} and \code{Rdisop} packages to predict molecular formulas
}
\section{Functions}{
\itemize{
\item \code{MzqueryModule}: Server logic

\item \code{MzqueryModuleUI}: UI elements
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
