% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SF_prediction_module.R
\name{MzqueryModule}
\alias{MzqueryModule}
\title{massquery}
\usage{
MzqueryModule(input, output, session, tag, set = list(search = list(elements =
  NULL, mz = NULL, data = NULL)))
}
\arguments{
\item{input}{m/z value}

\item{output}{maximum absolute error}

\item{session}{maximum relative error in ppm}

\item{tag}{id to be used in ns()}

\item{set}{Import data from the shiny session}
}
\description{
Chemcalc mz query module
server module to get molecular formula predictions from chemcalc
}
