% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_graph_utilities.R
\name{network1}
\alias{network1}
\title{network1}
\usage{
network1(spec1, spec2, mztol = 0.005, parentshift = 0,
  method = "cosine", minpeaks = 6, nonmatched = T)
}
\arguments{
\item{spec1}{spectrum 1}

\item{spec2}{spectrum 2}

\item{mztol}{allowed m/z difference between peaks to match matching}

\item{parentshift}{m/z difference between parentmasses (importantly, 
has to be Parent(spec2) - Parent(spec1) ) to find alternative matches /
 neutral loss matches; will only be calculated if \code{abs(parentshift) >  abs(mztol)}}

\item{method}{method passed on to \code{\link{pairCompare}()}; 
"cosine" will use lsa::cosine(), "pearson" will use stats::cor()}

\item{minpeaks}{minimum number 
of peaks that have to be matched, otherwise returns 0}

\item{nonmatched}{if TRUE, will add non-matching peaks to calculation, 
with 0 intensity in the spectrum missing the peak}
}
\value{
the result of the method call, a numeric similarity score
}
\description{
Compare two spectra and calculate a similarity score 
(e.g. intensity values of matched MS peaks)
NOTE: Current limitations: (1) peaks from one spectrum can match multiple 
peaks from the other scan (workaround: call quickMergeMS with ppm = 0 and 
mzdiff same as mzdiff here on both spectra prior to calling network1)
}
