% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_modules.R
\name{NetworkModule}
\alias{NetworkModule}
\title{NetworkModule}
\usage{
NetworkModule(input, output, session, tag, set = list(net = list(xrange =
  NULL, yrange = NULL, maxxrange = NULL, maxyrange = NULL, sel = NULL, mz =
  NULL, data = NULL, tables = list(nodes = NULL, edges = NULL)), layout =
  list(lw = 1, cex = 1, controls = F, ppm = 5, active = T)), keys)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{tag}{id to be used in ns()}

\item{set}{Import data from the shiny session}
}
\description{
server module for interactive mass spectrum view
}
