% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_extract_Scans.R
\name{Parentsearch}
\alias{Parentsearch}
\title{Parentsearch}
\usage{
Parentsearch(xcmsRaws, mz, rt, ppm = 5, rtw = 200)
}
\arguments{
\item{xcmsRaws}{list of xcmsRaw objects}

\item{mz}{parent mz values (mumeric vector)}

\item{rt}{parent retention time (in seconds, numeric vector), needs to be same length as mz}

\item{ppm}{parent mz tolerance in ppm}

\item{rtw}{parent rt tolerance in seconds}
}
\value{
a data.frame with metadata for MS2 scans that match the request
}
\description{
Make a list of all MS2 scans with the defined parent masses at a given retention time
}
