% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_msdata.R
\name{Parentsearch}
\alias{Parentsearch}
\title{Parentsearch}
\usage{
Parentsearch(xcmsRaws, mz = c(499.11085), rt = c(366.65), ppm = 5,
  rtw = 200)
}
\description{
Make a list of all MS2 scans with the defined parent masses at a given retention time
}
\details{
@param xcmsRaws list of xcmsRaw objects
 @param mz parent mz values (mumeric vector)
 @param rt parent retention time (in seconds, numeric vector), needs to be same length as mz
 @param ppm parent mz tolerance in ppm
 @param rtw parent rt tolerance in seconds
}
