% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_peakPicking.R
\name{peakDetect}
\alias{peakDetect}
\title{peakDetect}
\usage{
peakDetect(trace, SN = 1, minwidth = 4, localNoise = 7,
  localNoiseFactor = 0.5, globalNoiseFactor = 0.5, extend = T)
}
\arguments{
\item{trace}{numeric vector of intensity values. it is assumed that the data points were collected in regular intervals (problems may occur e.g. with variable scan rate due to ddMS2 experiments)}

\item{SN}{factor by which the maximum of a peak has to be larger than its edges (start  and  end scan) in final filter step}

\item{minwidth}{minimum width of a peak to not be thrown out in initial filter step}

\item{localNoise}{window size for local noise level calculation - default 7 means data from each point plus/minus 3 scans is used}

\item{localNoiseFactor}{factor to adjust local noise level. can be negative to allow very small peaks}

\item{globalNoiseFactor}{factor to adjust global noise level (based on entire trace). can be negative to allow very small peaks}

\item{extend}{if true, will try to expand peak boundaries to include downstream tails that may have been cut of before. Aslo depends on local noise settings.}
}
\description{
more recent try at peak detection
}
