% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_PeptideCalcModule.R
\name{PeptideCalcModule}
\alias{PeptideCalcModule}
\alias{PeptideCalcModuleUI}
\title{PeptideCalcModule}
\usage{
PeptideCalcModule(input, output, session, values)

PeptideCalcModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
If the server module for this Module returns something,
 it is described in \code{Details}.
}
\description{
This module generates a new feature table with mz values based on peptide sequences
and charges specified in the current FeatureTable, and also applies variable and fixed peptide modifications.
}
\details{
Does not return anything, but
\subsection{requires}{
\code{PeptideModificationsModule}
}
\subsection{modifies}{
\code{FeatureTable()}, \code{activeFT(values)}
}
\subsection{reads}{
\code{values$GlobalOpts$Peptides.variableMods},
 \code{values$GlobalOpts$Peptides.fixedMods},
 \code{FeatureTable(values)$df}
}
}
\section{Functions}{
\itemize{
\item \code{PeptideCalcModule}: server logic for PeptideCalcModule

\item \code{PeptideCalcModuleUI}: UI elements for PeptideCalcModule
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
