% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_PlotBrowserModule.R
\name{PlotBrowserModule}
\alias{PlotBrowserModule}
\alias{PlotBrowserModuleUI}
\title{PlotBrowserModule}
\usage{
PlotBrowserModule(
  input,
  output,
  session,
  reactives = reactive({     reactiveValues(PCAtable = df, active = T) }),
  static = list(patterns = list(axis = "PCA__", color = "", hover = ""))
)

PlotBrowserModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{reactives}{a list, wrapped in a \code{\link[shiny]{reactive}()} 
expression with settings for this module. See \code{details}.}

\item{static}{a list of arguments that are not in a reactive context 
and hence do not change while the app is running.}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()}
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
Returns its internalValues
}
\description{
MseekWidget to customize ggplot views
}
\section{Functions}{
\itemize{
\item \code{PlotBrowserModule}: Server logic

\item \code{PlotBrowserModuleUI}: UI elements
}}

