% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBrowserModule.R
\name{PlotModule}
\alias{PlotModule}
\title{PlotModule}
\usage{
PlotModule(input, output, session, reactives = reactive({    
  list(reactiveValues(plot = p, interactive = F)) }))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{reactives}{Import data from the shiny session}
}
\description{
server module for ggplot2 handler
}
