% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_peakPicking.R
\name{plotpeaks}
\alias{plotpeaks}
\title{plotpeaks}
\usage{
plotpeaks(EIC, peaklist)
}
\arguments{
\item{EIC}{a single EIC object (list item of multiEIC output)}

\item{peaklist}{a peaklist, such as list items returned by find_peaks() (or its rbind product)}
}
\description{
plot detected peaks for each file in an EIC object
}
