% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_rcdk.R
\name{plotSMILE}
\alias{plotSMILE}
\title{plotSMILE}
\usage{
plotSMILE(SMILE, ...)
}
\arguments{
\item{SMILE}{a character(1) SMILE term}

\item{...}{additional arguments passed to \link[adsf]{view.image.2d}()}
}
\description{
draw single molecule from a SMILE
}
