% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_msdata.R
\name{PlotWindow}
\alias{PlotWindow}
\title{PlotWindow}
\usage{
PlotWindow(cx = 1, ylim = c(0, max(unlist(EIClistItem[, "tic"]))),
  xlim = c(min(unlist(EIClistItem[, "rt"])), max(unlist(EIClistItem[,
  "rt"])))/60, heading = "test", single = F, relto = NULL)
}
\arguments{
\item{cx}{character expansion factor (font size)}

\item{ylim}{numeric(2) of y-axis range}

\item{xlim}{numeric(2) of x-axis range}

\item{heading}{heading of the plot}

\item{single}{if TRUE, this plot is expected to be the only plot in a com,posite (different margin settings)}
}
\description{
Initiate a plot and draw axes
}
