% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_msdata.R
\name{PlotWindow}
\alias{PlotWindow}
\title{PlotWindow}
\usage{
PlotWindow(cx = 1, ylim = c(0, max(unlist(EIClistItem[, "tic"]))),
  xlim = c(min(unlist(EIClistItem[, "rt"])), max(unlist(EIClistItem[,
  "rt"])))/60, heading = "test", single = F, par = T, relto = NULL,
  ylab = "Intensity", xlab = "RT (min)", ysci = T, liwi = 1)
}
\arguments{
\item{cx}{character expansion factor (font size)}

\item{ylim}{numeric(2) of y-axis range}

\item{xlim}{numeric(2) of x-axis range}

\item{heading}{heading of the plot}

\item{single}{if TRUE, this plot is expected to be the only plot in a composite (different margin settings)}

\item{par}{if FALSE, par margin settings are not set inside the function and should be set outside}

\item{relto}{show y axis values relative to relto if not NULL.}

\item{ylab}{y axis label}

\item{xlab}{x axis label}

\item{ysci}{if TRUE, y axis label numbers are shown in scientific format}
}
\description{
Initiate a plot and draw axes
}
