% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_msdata.R
\name{rawGrouping}
\alias{rawGrouping}
\title{rawGrouping}
\usage{
rawGrouping(rawgrouptable)
}
\arguments{
\item{rawgrouptable}{a data.frame with columns File and Group , holding file paths and group names, respectively.}
}
\description{
Groups a rawgrouptable into a named list.
}
