% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_RegroupMSDataModule.R
\name{RegroupMSDataModule}
\alias{RegroupMSDataModule}
\alias{RegroupMSDataModuleUI}
\title{RegroupMSDataModule}
\usage{
RegroupMSDataModule(
  input,
  output,
  session,
  values = reactiveValues(MSData = NULL, projectData = NULL)
)

RegroupMSDataModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
If the server module for this Module returns something,
 it is described in \code{Details}.
}
\description{
Select and group MS files in plotting layouts
}
\section{Functions}{
\itemize{
\item \code{RegroupMSDataModule}: server logic

\item \code{RegroupMSDataModuleUI}: UI elements
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
