% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms_retcorAdjust.R
\name{rtadjust}
\alias{rtadjust}
\title{rtadjust}
\usage{
rtadjust(rtexportObj, rts)
}
\arguments{
\item{rtexportObj}{rtexport output object}

\item{rts}{data.frame with rt values}
}
\description{
recalculate original rt values from rt values as reported in featuretables after rt correction for each file
}
