% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms_retcorAdjust.R
\name{rtexport}
\alias{rtexport}
\title{rtexport}
\usage{
rtexport(xset)
}
\arguments{
\item{xset}{an xcms::XCMSnExp, xcms::xcmsSet or xcms::xsAnnotate object}
}
\description{
Extract corrected and uncorrected retention times for each scan from an xcms::XCMSnExp, xcms::xcmsSet or xcms::xsAnnotate object
}
