% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_retcorAdjust.R
\name{RTplot}
\alias{RTplot}
\title{RTplot}
\usage{
RTplot(rtt, colscheme = "topo.colors", liwi = 2, cx = 1)
}
\arguments{
\item{rtt}{RTcorr object}

\item{colscheme}{function for color palette}

\item{liwi}{line width}

\item{cx}{font size}
}
\description{
Plot retention time deviation for each file generated by an RTexport() generated RTcorr object
}
