% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMosaic.R
\name{runMosaic}
\alias{runMosaic}
\title{Run Mosaic}
\usage{
runMosaic(devel = F, server = T, ...)
}
\arguments{
\item{devel}{Enable troubleshooting tools (default: FALSE). May constitute security risk if run on a publicly available server.}

\item{server}{Enable server mode (users can not access file paths on the server directly), default is TRUE.}

\item{...}{Additional arguments passed to shiny::runApp, e.g. to set up the port used or to make Mosaic reachable from other computers (with host = getOption("shiny.host", "0.0.0.0"))}
}
\description{
Run Mosaic in your webbrowser.
}
