% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_runMosaic.R, R/Function_runMseek.R
\name{runMseek}
\alias{runMseek}
\title{RunMseek}
\usage{
runMseek(...)

runMseek(...)
}
\arguments{
\item{...}{Additional arguments passed to shiny::runApp, e.g. to set up the port used or to make Mseek reachable from other computers (with host = getOption("shiny.host", "0.0.0.0"))}

\item{...}{Additional arguments passed to shiny::runApp, e.g. to set up the port used or to make Mseek reachable from other computers (with host = getOption("shiny.host", "0.0.0.0"))}
}
\description{
Run METABOseek in your webbrowser.

Run Mseek in your webbrowser.
}
