% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_SIRIUS.R
\name{runSirius}
\alias{runSirius}
\title{runSirius}
\usage{
runSirius(outfolder, ms1 = NULL, ms2, instrument, parentmz,
  comments = "", rt = "", ion, charge = 1, fingerid = T,
  scanindices = NULL, sirpath, moreOpts = "")
}
\arguments{
\item{outfolder}{output folder for the SIRIUS results}

\item{ms2}{a list of ms2 spectra (matrices with columns mz and intensity)}

\item{parentmz}{numeric: parent ion m/z value}

\item{comments}{character: comments}

\item{rt}{numeric: retention time in seconds}

\item{ion}{ion type (e.g. \eqn{[M+H]+})}

\item{charge}{charge (positive or negative integer)}

\item{scanindices}{which scans were averaged into this ms2 spectrum}

\item{sirpath}{path to SIRIUS executable}

\item{moreOpts}{character with additional options to be passed to SIRIUS}
}
\description{
run SIRIUS externally for a list of ms2 spectra
}
