% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_SIRIUS.R
\name{runSirius}
\alias{runSirius}
\title{runSirius}
\usage{
runSirius(outfolder, ms1 = NULL, ms2, instrument, parentmz,
  comments = "", rt = "", ion, charge = 1, fingerid = T,
  scanindices = NULL, sirpath, moreOpts = "", force = T)
}
\arguments{
\item{outfolder}{output folder for the SIRIUS results}

\item{ms2}{a list (or list of lists) of ms2 spectra 
(matrices with columns mz and intensity), see \code{Details}}

\item{parentmz}{numeric: parent ion m/z value}

\item{comments}{character: comments}

\item{rt}{numeric: retention time in seconds}

\item{ion}{ion type (e.g. \eqn{[M+H]+})}

\item{charge}{charge (positive or negative integer)}

\item{scanindices}{which scans were averaged into this ms2 spectrum}

\item{sirpath}{path to SIRIUS executable}

\item{moreOpts}{character with additional options to be passed to SIRIUS}

\item{force}{force calculation, even if same results should exist according to indexfile}
}
\value{
Will cause SIRIUS to run a job and save its results in
 a timestamped subfolder of  \code{outfolder}, and will register the job in
 the \code{index.csv} file in \code{outfolder}.
}
\description{
run SIRIUS externally for a list of ms2 spectra
}
\details{
\describe{
   \item{ms2}{can be a list of matrices, or a list of list of matrices. 
   In the latter case, the (optional) names of the nested list items are 
   expected to be in the format "collisionXX", where XX denotes an integer 
   collision energy.}
}
}
