% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_plotting_genericData.R
\name{safelog}
\alias{safelog}
\title{safelog}
\usage{
safelog(x, base = 10, replaceZeros = NULL, replaceNonFinites = NA_real_)
}
\arguments{
\item{x}{numeric vector of positive values}

\item{base}{of log}

\item{replaceZeros}{replace zeros with this value, if NULL will use 
smallest non-zero absolute value in x}

\item{replaceNonFinites}{replace non-finite values (\code{Inf, -Inf, NA}) 
with this value before applying log}
}
\value{
a numeric vector of same length as \code{x}, with logarithmized values
}
\description{
apply a log function to a numeric vector, but replaces zeros and uses abs values
}
