% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SaveNetworkModule.R
\name{SaveNetworkModule}
\alias{SaveNetworkModule}
\alias{SaveNetworkModuleUI}
\title{SaveNetworkModule}
\usage{
SaveNetworkModule(input, output, session,
  values = reactiveValues(Networks = Networks), reactives = reactive({ 
     list(graphname = NULL, filename = "table.graphml") }),
  static = list(tooltip = "Save Network as a graphml file", label =
  "Save Network", format = c("graphml")))

SaveNetworkModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{reactives}{a list, wrapped in a \code{\link[shiny]{reactive}()} 
expression with settings for this module. See \code{details}.}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
Returns nothing
}
\description{
Module for saving molecular network data
}
\section{Functions}{
\itemize{
\item \code{SaveNetworkModule}: Server logic

\item \code{SaveNetworkModuleUI}: UI elements
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
