% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_script_functions.R
\name{savetable}
\alias{savetable}
\title{savetable}
\usage{
savetable(xset, status = NULL, fill = FillChromPeaksParam(expandMz = 0.005,
  expandRt = 5, ppm = 3), nonfill = F, filename = "tableoutxx.csv",
  bparams = SnowParam(workers = 1), intensities = list(ppm = 5, rtw = 5,
  rtrange = T), rawdata = NULL, saveR = T)
}
\arguments{
\item{xset}{an xcms::XCMSnExp, xcms::xcmsSet or xcms::xsAnnotate object from which to extract the peaklist}

\item{status}{a writeStatus object to be modified and returned by the function.}

\item{fill}{xcms::FillChromPeaksParam object to specify how to fill peaks (only works if xset is an XCMSnExp object). No peak filling if NULL.}

\item{nonfill}{also write the non-filled peaktable data? TRUE or FALSE}

\item{filename}{filename to use when writing the csv file (including file extension). "_filled" will be added for filled data.}

\item{bparams}{BBPARAMs (BiocParallel::BiocParallelParam object with parameters for parallel computing used in fill peaks)}

\item{intensities}{list of parameters for Mosaic EIC intensity extraction (for each feature in the peaklist). ppm: mz window, rtw: rt window (+/- seconds), rtrange: if TRUE, rtw extends from rtmin and rtmax, if FALSE rtw centers around value in column rt. Not performed if intensities = NULL.}

\item{rawdata}{named list of xcmsRaw objects (as returned by Mosaic::loadRawM), required for Mosaic EIC intensity extraction which is not performed if NULL.}

\item{saveR}{save the xset object as .Rdata file}
}
\description{
Write or update a status dataframe with custom messages, time stamp and time elapsed since last status. Useful to document time consumed analysis by steps.
}
