% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SaveTableModule.R
\name{SaveTableModule}
\alias{SaveTableModule}
\alias{SaveTableModuleUI}
\title{SaveTableModule}
\usage{
SaveTableModule(input, output, session,
  values = reactiveValues(projectData = projectData, featureTables =
  NULL), reactives = reactive({     list(df = NULL, filename =
  "table.csv") }), static = list(tooltip = "Save", label = "Save", format
  = c("tsv", "csv"), allowformats = NULL))

SaveTableModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{reactives}{a list, wrapped in a \code{\link[shiny]{reactive}()} 
expression with settings for this module. See \code{details}.}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
Returns nothing
}
\description{
Module for saving Tables
}
\section{Functions}{
\itemize{
\item \code{SaveTableModule}: Server logic

\item \code{SaveTableModuleUI}: UI elements
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
