% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SelectActiveTableModule.R
\name{SelectActiveTableModule}
\alias{SelectActiveTableModule}
\alias{SelectActiveTableModuleUI}
\title{SelectActiveTableModule}
\usage{
SelectActiveTableModule(
  input,
  output,
  session,
  values = reactiveValues(featureTables = NULL, MainTable = NULL)
)

SelectActiveTableModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
Returns nothing
}
\description{
Module to select the active Feature Table
}
\section{Functions}{
\itemize{
\item \code{SelectActiveTableModule}: Server logic

\item \code{SelectActiveTableModuleUI}: UI elements
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
