% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_simpleTableModule.R
\name{simpleTableModule}
\alias{simpleTableModule}
\title{simpleTableModule}
\usage{
simpleTableModule(input, output, session, df = reactive({     NULL }),
  static = list(readOnly = T, contextMenu = F, height = "auto"))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{set}{Import data from the shiny session}
}
\description{
server module for interactive mass spectrum view
}
