% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SiriusModule.R
\name{SiriusModule}
\alias{SiriusModule}
\title{SiriusModule}
\usage{
SiriusModule(input, output, session, values = reactiveValues(GlobalOpts =
  GlobalOpts))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{reactives}{Import data from the shiny session}
}
\description{
server module for interactive mass spectrum view
}
