% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SiriusOptionsModule.R
\name{SiriusOptionsModule}
\alias{SiriusOptionsModule}
\alias{SiriusOptionsModuleUI}
\title{SiriusOptionsModule}
\usage{
SiriusOptionsModule(input, output, session, values)

SiriusOptionsModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
If the server module for this Module returns something,
 it is described in \code{Details}.
}
\description{
Module to change Sirius analysis options
}
\section{Functions}{
\itemize{
\item \code{SiriusOptionsModule}: Server logic

\item \code{SiriusOptionsModuleUI}: UI elements
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
