% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Specmodule.R
\name{Specmodule}
\alias{Specmodule}
\title{Specmodule}
\usage{
Specmodule(input, output, session, tag, set = list(spec = list(xrange = NULL,
  yrange = NULL, maxxrange = NULL, maxyrange = NULL, sel = NULL, mz = NULL, data
  = NULL, MS2 = T), layout = list(lw = 1, cex = 1, controls = F, ppm = 5, active
  = T, highlights = NULL, height = 550), msdata = NULL), keys)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{tag}{id to be used in ns()}

\item{set}{Import data from the shiny session}

\item{keys}{registered keystrokes}
}
\description{
server module for interactive mass spectrum view
}
