% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SpecModule2.R
\name{SpecModule2}
\alias{SpecModule2}
\alias{SpecModule2UI}
\title{SpecModule2}
\usage{
SpecModule2(input, output, session, values = values,
  reactives = reactive({     list(scantable = NULL, type = "ms2",
  spectrum = NULL, specinfo = NULL, xrange = NULL, markermz = NULL) }))

SpecModule2UI(id)
}
\arguments{
\item{input, output, session}{arguments necessary for use with \link[shiny]{callModule}()}

\item{values}{a \link[shiny]{reactiveValues} object that in effect gives read and write access to external objects}

\item{id}{id to be used to define a namespace via  \link[shiny]{NS}() (must match \code{id} in \link[shiny]{callModule} for the server Module)}

\item{reactive}{a \link[shiny]{reactive} object with settings for this module. See \code{details}.}
}
\description{
server module for interactive mass spectrum view
}
\details{
\subsection{reactives}{
Can either be a list with entries \code{scanTable} and \code{type}, 
specifying inputs for the \link[METABOseek]{getScanInfo}, OR
a list with entries \code{spectrum} and \code{specinfo}. 

\describe{
\item{scanTable}{}
\item{type}{}
\item{spectrum}{}
\item{specinfo}{}
}
}
}
\section{Functions}{
\itemize{
\item \code{SpecModule2}: Server module, to be called with \link[shiny]{callModule}()

\item \code{SpecModule2UI}: UI function for SpecModule2
}}

