% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_specplot.R
\name{specplot}
\alias{specplot}
\title{specplot}
\usage{
specplot(x = sc[, 1], y = sc[, 2], norm = max(y)/100, cx = 1.5,
  k = 10, fileName = "title", yrange = c(0, 100),
  xrange = range(x), maxi = max(y), mar = c(4, 6, 6, 2))
}
\arguments{
\item{x}{mz coordinates}

\item{y}{intensity coordinates}

\item{norm}{normalize by}

\item{cx}{font size}

\item{k}{top k intensity peaks will be labeled}

\item{fileName}{plot title}

\item{yrange}{y axis range}

\item{xrange}{x axis range}

\item{maxi}{max intensity to be plotted on side}
}
\description{
Plot an MS spectrum
}
