% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_SpecModuleWidget.R
\name{SpecplotWidget}
\alias{SpecplotWidget}
\alias{SpecplotWidgetUI}
\title{SpecplotWidget}
\usage{
SpecplotWidget(input, output, session, reactives = reactive({     list(x
  = sc[, 1], y = sc[, 2], norm = max(y)/100, cx = 1.5, k = 10, fileName =
  "title", yrange = c(0, 100), xrange = range(x), maxi = max(y), labels =
  NULL, mar = c(4, 6, 6, 2), ylab = "Relative Intensity (\%)", ylabshift =
  0) }), layout = reactive({     list(active = T, height = 550,
  selectCallback = T) }), keys = reactive({     "NO" }))

SpecplotWidgetUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{reactives}{a list, wrapped in a \code{\link[shiny]{reactive}()} 
expression with settings for this module. See \code{details}.}

\item{keys}{a \link[shiny]{reactive} object registering keystrokes}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\description{
Module for interactive mass spectrum view
}
\section{Functions}{
\itemize{
\item \code{SpecplotWidget}: Server module, to be called with \link[shiny]{callModule}()

\item \code{SpecplotWidgetUI}: UI function for SpecplotWidget
}}

