% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_SpecModuleWidget.R
\name{SpecplotWidget}
\alias{SpecplotWidget}
\alias{SpecplotWidgetUI}
\title{SpecplotWidget}
\usage{
SpecplotWidget(input, output, session, reactives = reactive({     list(x
  = sc[, 1], y = sc[, 2], norm = max(y)/100, cx = 1.5, k = 10, fileName =
  "title", yrange = c(0, 100), xrange = range(x), maxi = max(y), labels =
  NULL, mar = c(4, 6, 6, 2), ylab = "Relative Intensity (\%)", ylabshift =
  0) }), layout = reactive({     list(active = T, height = 550,
  selectCallback = T) }), keys)

SpecplotWidgetUI(id)
}
\arguments{
\item{keys}{a \link[shiny]{reactive} object registering keystrokes}
}
\description{
server and UI module for interactive mass spectrum view
}
\section{Functions}{
\itemize{
\item \code{SpecplotWidgetUI}: UI function for SpecplotWidget
}}

