% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableAnalysisModule.R
\name{TableAnalysisModule}
\alias{TableAnalysisModule}
\title{TableAnalysisModule}
\usage{
TableAnalysisModule(input, output, session, reactives = reactive({    
  list(fileGrouping = NULL) }), values = reactiveValues(featureTables =
  featureTables, MSData = MSData), static = list())
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{reactives}{Import data from the shiny session}

\item{values}{Import data from the shiny session}

\item{static}{Import data from the shiny session}
}
\description{
server module for interactive mass spectrum view
}
