% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_TableAnalysisModule.R
\name{TableAnalysisModule}
\alias{TableAnalysisModule}
\alias{TableAnalysisModuleUI}
\title{TableAnalysisModule}
\usage{
TableAnalysisModule(input, output, session, values,
  reactives = reactive({     list(fileGrouping = NULL) }),
  static = list())

TableAnalysisModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{values}{a \code{\link[shiny]{reactiveValues}} object that in effect 
gives read and write access to external objects}

\item{reactives}{a list, wrapped in a \code{\link[shiny]{reactive}()} 
expression with settings for this module. See \code{details}.}

\item{static}{a list of arguments that are not in a reactive context 
and hence do not change while the app is running.}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()} 
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
Returns its internalValues
}
\description{
Module for Feature Table analysis
}
\section{Functions}{
\itemize{
\item \code{TableAnalysisModule}: Server logic

\item \code{TableAnalysisModuleUI}: UI elements
}}

\seealso{
\code{\link{values}} for a description of the \code{values} object
}
