% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_class_featureTable.R
\name{tableGrouping}
\alias{tableGrouping}
\title{tableGrouping}
\usage{
tableGrouping(df = NULL, anagrouptable)
}
\arguments{
\item{df}{feature table as data.frame, with intensity values.}

\item{anagrouptable}{Analysis grouping table: a data.frame with columns "Column" (containing column names from df with intensity values) and "Group" (defining a group for each entry in "Column")}
}
\description{
Groups a anagrouptable into a named list. Helper function to reformat Table analysis grouping tables
}
