% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chunk_TableUpdateChunk.R
\name{TableUpdateChunk}
\alias{TableUpdateChunk}
\title{TableUpdateChunk}
\usage{
TableUpdateChunk()
}
\description{
Code chunk to update the active Feature Table before triggering a rerendering of the rhandsontable object.
 
Chunks are evaluated in their parent environment, and therefore require all objects they work on to be present under their correct names.
}
\details{
This chunk requires:

values$featureTables$Maintable - pointing to the MainTableObject
values$featureTables
}
