% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_class_featureTable.R
\name{tableWriter}
\alias{tableWriter}
\title{tableWriter}
\usage{
tableWriter(df, fname, format = c("csv", "tsv", "instrumentList"),
  moreArgs = list())
}
\arguments{
\item{df}{feature table as data.frame, with mz and rt columns.}

\item{fname}{file name}

\item{format}{output format}

\item{moreArgs}{additional arguments}

\item{instrument}{which instrument is this table for?}

\item{listType}{"inclusion" or "exclusion"}
}
\description{
Writes a feature table to a file in a given format.
}
