% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_class_featureTable.R
\name{updateFeatureTable}
\alias{updateFeatureTable}
\title{updateDF}
\usage{
updateFeatureTable(FT, df)
}
\arguments{
\item{FT}{MseekFT object}

\item{df}{source data.frame}
}
\description{
update an existing MseekFT with new data frame columns (leaving intensitiy tables and constants in effect)
}
