% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_class_featureTable.R
\name{updateFeatureTable}
\alias{updateFeatureTable}
\title{updateFeatureTable}
\usage{
updateFeatureTable(FT, df)
}
\arguments{
\item{FT}{MseekFT object}

\item{df}{source data.frame}
}
\value{
An updated \code{MseekFT} object
}
\description{
update an existing MseekFT with new data.frame columns (leaving intensitiy
 column names and constants in effect, but updating information on which 
 "other" columns are present)
}
