% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_feature_table.R
\name{updateFTgrouping}
\alias{updateFTgrouping}
\title{updateFTgrouping}
\usage{
updateFTgrouping(FT, anagrouptable)
}
\arguments{
\item{FT}{MosaicFT object}

\item{anagrouptable}{Analysis grouping table: a data.frame with columns "Column" (containing column names from df with intensity values) and "Group" (defining a group for each entry in "Column")}
}
\description{
update or construct grouping information of a MosaicFT object
}
