% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_class_featureTable.R
\name{updateFTIndex}
\alias{updateFTIndex}
\title{updateFTgrouping}
\usage{
updateFTIndex(tables)
}
\arguments{
\item{table}{named list of MseekFT objects}
}
\value{
named list of MseekFT tablenames
}
\description{
update the FT index of the Mseek App (needed because this index is a named 
list - extracts tablename entry from each MseekFT object in list)
}
