% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_feature_table.R
\name{updateFTIndex}
\alias{updateFTIndex}
\title{updateFTgrouping}
\usage{
updateFTIndex(tables)
}
\arguments{
\item{table}{list of MosaicFT objects}
}
\description{
update the FT index of the Mosaic App (needed because this index is a named list - extracts tablename entry from each MosaicFT object in list)
}
