% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_updaterModule.R
\name{updaterModule}
\alias{updaterModule}
\title{updaterModule}
\usage{
updaterModule(input, output, session, tag, set = list(package =
  "METABOseek", refs = c("master", "devel", "devel_raw"), active = T))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{tag}{id to be used in ns()}

\item{set}{Import data from the shiny session}
}
\description{
module to update Mseek by the click of a button
}
