% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_VennWidgetModule.R
\name{VennWidgetModule}
\alias{VennWidgetModule}
\alias{VennWidgetModuleUI}
\title{VennWidgetModule}
\usage{
VennWidgetModule(input, output, session, reactives = reactive({    
  list(x = list(A = 1:105, B = 101:115), ext.percent = c(0.01, 0.01, 0.01),
  filename = NULL) }), static = list())

VennWidgetModuleUI(id)
}
\arguments{
\item{input}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{output}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{session}{arguments necessary for use with 
\code{\link[shiny]{callModule}()}}

\item{reactives}{a list, wrapped in a \code{\link[shiny]{reactive}()} 
expression with settings for this module. See \code{details}.}

\item{static}{a list of arguments that are not in a reactive context 
and hence do not change while the app is running.}

\item{id}{id to be used to define a namespace via  \code{\link[shiny]{NS}()}
(must match \code{id} in \code{\link[shiny]{callModule}} 
for the server Module)}
}
\value{
Returns nothing
}
\description{
MseekWidget for plotting Venn diagrams
}
\section{Functions}{
\itemize{
\item \code{VennWidgetModule}: Server logic

\item \code{VennWidgetModuleUI}: UI elements
}}

