% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_WelcomePageModule.R
\name{WelcomePageModule}
\alias{WelcomePageModule}
\title{WelcomePageModule}
\usage{
WelcomePageModule(input, output, session, values = reactiveValues(MSData
  = MSData))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\description{
module to welcome Mosaic users.
}
