% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_WelcomePageModule.R
\name{WelcomePageModule}
\alias{WelcomePageModule}
\title{WelcomePageModule}
\usage{
WelcomePageModule(input, output, session,
  values = reactiveValues(projectData = values$projectData, featureTables
  = values$featureTables, MSData = values$MSData, GlobalOpts =
  values$GlobalOpts), show = reactive({     T }))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\description{
module to welcome Mseek users.
}
