% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_script_functions.R
\name{writeStatus}
\alias{writeStatus}
\title{writeStatus}
\usage{
writeStatus(previous = NULL, message = list(Status = "Starting analysis",
  Details = "loading files"), wd = getwd())
}
\arguments{
\item{previous}{a writeStatus object to be updated (must have same column names); new object generated if NULL.}

\item{message}{named list of columns with custom messages}

\item{wd}{directory where the status.csv file should be written and updated. If NULL, no status.csv is written.}
}
\description{
Write or update a status dataframe with custom messages, time stamp and time elapsed since last status. Useful to document time consumed analysis by steps.
}
\examples{

status <- writeStatus(previous = NULL,
            message = list(Status = "Starting analysis",
                           Details = "loading files"
                          ),
           wd = getwd())
           
           
status <- writeStatus(previous = status,
            message = list(Status = "Running analysis",
                           Details = "detecting features"
                          ),
           wd = getwd())

}
