% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_xcmsModule.R
\name{xcmsModule}
\alias{xcmsModule}
\title{xcmsModule}
\usage{
xcmsModule(input, output, session, values = reactiveValues(),
  static = list(servermode = F, rootpath = .MosaicOptions$filePaths,
  activateXCMS = T, filePattern = .MosaicOptions$filePattern))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{values}{import reactiveValues from the shiny session}

\item{static}{import static values}
}
\description{
server module for accessing the xcms data analysis workflow
}
