% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms_module.R
\name{xcmsModule}
\alias{xcmsModule}
\title{Specmodule}
\usage{
xcmsModule(input, output, session, tag, reactives = reactive({     (list())
  }), values = reactiveValues(), static = list(servermode = F),
  load = reactive({     list() }))
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{tag}{id to be used in ns()}

\item{reactives}{import reactive data from the shiny session}

\item{values}{import reactiveValues from the shiny session}

\item{static}{import static values}

\item{load}{data to load from previous session (not implemented)}
}
\description{
server module for interactive mass spectrum view
}
